Put datatables enums in dedicated headers
