+++ source/D2Common/CMakeLists.txt
@@ -58,6 +58,13 @@ target_sources(D2Common
     include/DataTbls/SkillsTbls.h
     include/DataTbls/TokenTbls.h
     include/DataTbls/TransformTbls.h
+	
+	# Enums for asset indices, should be generated from .txt files
+	include/DataTbls/LevelsIds.h
+	include/DataTbls/ObjectsIds.h
+	include/DataTbls/MissilesIds.h
+	include/DataTbls/MonsterIds.h
+	include/DataTbls/SkillsIds.h
 )
 
 # Dynamic Random Level Generation
