+++ source/D2Common/include/D2Items.h
@@ -5,6 +5,190 @@
 //TODO: Revise
 
 
+enum D2C_ItemQualities
+{
+	ITEMQUAL_INFERIOR = 0x01, 	//0x01 Inferior
+	ITEMQUAL_NORMAL = 0x02, 	//0x02 Normal
+	ITEMQUAL_SUPERIOR = 0x03, 	//0x03 Superior
+	ITEMQUAL_MAGIC = 0x04, 		//0x04 Magic
+	ITEMQUAL_SET = 0x05, 		//0x05 Set
+	ITEMQUAL_RARE = 0x06, 		//0x06 Rare
+	ITEMQUAL_UNIQUE = 0x07, 	//0x07 Unique
+	ITEMQUAL_CRAFT = 0x08, 		//0x08 Crafted
+	ITEMQUAL_TEMPERED = 0x09	//0x09 Tempered
+};
+
+enum D2C_ItemTypes
+{
+	ITEMTYPE_NONE_1 = 0,
+	ITEMTYPE_NONE_2,
+	ITEMTYPE_SHIELD,
+	ITEMTYPE_ARMOR,
+	ITEMTYPE_GOLD,
+	ITEMTYPE_BOW_QUIVER,
+	ITEMTYPE_CROSSBOW_QUIVER,
+	ITEMTYPE_PLAYER_BODY_PART,
+	ITEMTYPE_HERB,
+	ITEMTYPE_POTION,
+	ITEMTYPE_RING,
+	ITEMTYPE_ELIXIR,
+	ITEMTYPE_AMULET,
+	ITEMTYPE_CHARM,
+	ITEMTYPE_NONE_3,
+	ITEMTYPE_BOOTS,
+	ITEMTYPE_GLOVES,
+	ITEMTYPE_NONE_4,
+	ITEMTYPE_BOOK,
+	ITEMTYPE_BELT,
+	ITEMTYPE_GEM,
+	ITEMTYPE_TORCH,
+	ITEMTYPE_SCROLL,
+	ITEMTYPE_NONE_5,
+	ITEMTYPE_SCEPTER,
+	ITEMTYPE_WAND,
+	ITEMTYPE_STAFF,
+	ITEMTYPE_BOW,
+	ITEMTYPE_AXE,
+	ITEMTYPE_CLUB,
+	ITEMTYPE_SWORD,
+	ITEMTYPE_HAMMER,
+	ITEMTYPE_KNIFE,
+	ITEMTYPE_SPEAR,
+	ITEMTYPE_POLEARM,
+	ITEMTYPE_CROSSBOW,
+	ITEMTYPE_MACE,
+	ITEMTYPE_HELM,
+	ITEMTYPE_MISSILE_POTION,
+	ITEMTYPE_QUEST,
+	ITEMTYPE_BODY_PART,
+	ITEMTYPE_KEY,
+	ITEMTYPE_THROWING_KNIFE,
+	ITEMTYPE_THROWING_AXE,
+	ITEMTYPE_JAVELIN,
+	ITEMTYPE_WEAPON,
+	ITEMTYPE_MELEE_WEAPON,
+	ITEMTYPE_MISSILE_WEAPON,
+	ITEMTYPE_THROWN_WEAPON,
+	ITEMTYPE_COMBO_WEAPON,
+	ITEMTYPE_ANY_ARMOR,
+	ITEMTYPE_ANY_SHIELD,
+	ITEMTYPE_MISCELLANEOUS,
+	ITEMTYPE_SOCKET_FILLER,
+	ITEMTYPE_SECOND_HAND,
+	ITEMTYPE_STAVES_AND_RODS,
+	ITEMTYPE_MISSILE,
+	ITEMTYPE_BLUNT,
+	ITEMTYPE_JEWEL,
+	ITEMTYPE_CLASS_SPECIFIC,
+	ITEMTYPE_AMAZON_ITEM,
+	ITEMTYPE_BARBARIAN_ITEM,
+	ITEMTYPE_NECROMANCER_ITEM,
+	ITEMTYPE_PALADIN_ITEM,
+	ITEMTYPE_SORCERESS_ITEM,
+	ITEMTYPE_ASSASSIN_ITEM,
+	ITEMTYPE_DRUID_ITEM,
+	ITEMTYPE_HAND_TO_HAND,
+	ITEMTYPE_ORB,
+	ITEMTYPE_VOODOO_HEADS,
+	ITEMTYPE_AURIC_SHIELDS,
+	ITEMTYPE_PRIMAL_HELM,
+	ITEMTYPE_PELT,
+	ITEMTYPE_CLOAK,
+	ITEMTYPE_RUNE,
+	ITEMTYPE_CIRCLET,
+	ITEMTYPE_HEALING_POTION,
+	ITEMTYPE_MANA_POTION,
+	ITEMTYPE_REJUV_POTION,
+	ITEMTYPE_STAMINA_POTION,
+	ITEMTYPE_ANTIDOTE_POTION,
+	ITEMTYPE_THAWING_POTION,
+	ITEMTYPE_SMALL_CHARM,
+	ITEMTYPE_MEDIUM_CHARM,
+	ITEMTYPE_LARGE_CHARM,
+	ITEMTYPE_AMAZON_BOW,
+	ITEMTYPE_AMAZON_SPEAR,
+	ITEMTYPE_AMAZON_JAVELIN,
+	ITEMTYPE_HAND_TO_HAND_2,
+	ITEMTYPE_MAGIC_BOW_QUIV,
+	ITEMTYPE_UNK,
+	ITEMTYPE_CHIPPED_GEM,
+	ITEMTYPE_FLAWED_GEM,
+	ITEMTYPE_STANDARD_GEM,
+	ITEMTYPE_FLAWLESS_GEM,
+	ITEMTYPE_PERFECT_GEM,
+	ITEMTYPE_AMETHYST,
+	ITEMTYPE_DIAMOND,
+	ITEMTYPE_EMERALD,
+	ITEMTYPE_RUBY,
+	ITEMTYPE_SAPPHIRE,
+	ITEMTYPE_TOPAZ,
+	ITEMTYPE_SKULL,
+	// Own ItemTypes
+	ITEMTYPE_TOME,
+	ITEMTYPE_CLUESCROLL
+};
+
+enum D2C_ItemModes
+{
+	IMODE_STORED, 		//Item is in Storage (inventory, cube, Stash?)
+	IMODE_EQUIP, 		//Item is Equippped
+	IMODE_INBELT, 		//Item is in Belt Rows
+	IMODE_ONGROUND, 	//Item is on Ground
+	IMODE_ONCURSOR, 	//Item is on Cursor
+	IMODE_DROPPING, 	//Item is Being Dropped
+	IMODE_SOCKETED		//Item is Socketed in another Item
+};
+
+enum D2C_ItemFlags
+{
+	IFLAG_NEWITEM = 0x00000001,
+	IFLAG_TARGET = 0x00000002,
+	IFLAG_TARGETING = 0x00000004,
+	IFLAG_DELETED = 0x00000008,
+	IFLAG_IDENTIFIED = 0x00000010,
+	IFLAG_QUANTITY = 0x00000020,
+	IFLAG_SWITCHIN = 0x00000040,
+	IFLAG_SWITCHOUT = 0x00000080,
+	IFLAG_BROKEN = 0x00000100,
+	IFLAG_REPAIRED = 0x00000200,
+	IFLAG_UNK1 = 0x00000400,
+	IFLAG_SOCKETED = 0x00000800,
+	IFLAG_NOSELL = 0x00001000,
+	IFLAG_INSTORE = 0x00002000,
+	IFLAG_NOEQUIP = 0x00004000,
+	IFLAG_NAMED = 0x00008000,
+	IFLAG_ISEAR = 0x00010000,
+	IFLAG_STARTITEM = 0x00020000,
+	IFLAG_UNK2 = 0x00040000,
+	IFLAG_INIT = 0x00080000,
+	IFLAG_UNK3 = 0x00100000,
+	IFLAG_COMPACTSAVE = 0x00200000,
+	IFLAG_ETHEREAL = 0x00400000,
+	IFLAG_JUSTSAVED = 0x00800000,
+	IFLAG_PERSONALIZED = 0x01000000,
+	IFLAG_LOWQUALITY = 0x02000000,
+	IFLAG_RUNEWORD = 0x04000000,
+	IFLAG_ITEM = 0x08000000
+};
+
+enum D2C_UniqueItemFlags
+{
+	UNIQUEITEM_ENABLED = 0x00000001,
+	UNIQUEITEM_NOLIMIT = 0x00000002,
+	UNIQUEITEM_CARRY1 = 0x00000004,
+	UNIQUEITEM_LADDER = 0x00000008,
+};
+
+enum D2C_ItemDropFlags
+{
+	ITEMDROPFLAG_ALWAYSETH = 0x00000004,
+};
+
+enum D2C_ItemCmdFlags
+{
+	ITEMCMDFLAG_DELETE = 0x00000020,
+};
+
 
 //D2Common.0x6FD98380 (#10687)
 void __stdcall ITEMS_AllocItemData(void* pMemPool, D2UnitStrc* pItem);
