+++ source/D2Common/include/DataTbls/HoradricCube.h
@@ -2,6 +2,56 @@
 
 #include <D2BasicTypes.h>
 
+enum D2C_CubeOps
+{
+	CUBEOP_COWPORTAL = 1,
+	CUBEOP_UBERDUNGEON = 2,
+	CUBEOP_UBERTRISTRAM = 3,
+	CUBEOP_SKILLPTS = 4,
+	CUBEOP_STATPTS = 5,
+	CUBEOP_REDPORTAL = 128,
+	CUBEOP_ADDSTAT = 129,
+	CUBEOP_ITEMCODE = 252,
+	CUBEOP_ITEMTYPE = 253,
+	CUBEOP_USEITEM = 254,
+	CUBEOP_USETYPE = 255,
+};
+
+enum D2C_CubeOutFlags
+{
+	CUBEFLAG_OUT_COPYMODS = 0x01,
+	CUBEFLAG_OUT_SOCKET = 0x02,
+	CUBEFLAG_OUT_ETHEREAL = 0x04,
+	CUBEFLAG_OUT_SPECIAL = 0x08,
+	CUBEFLAG_OUT_UNSOCKET = 0x10,
+	CUBEFLAG_OUT_REMOVE = 0x20,
+	CUBEFLAG_OUT_NORMAL = 0x40,
+	CUBEFLAG_OUT_EXCEPTIONAL = 0x80,
+	CUBEFLAG_OUT_ELITE = 0x100,
+	CUBEFLAG_OUT_REPAIR = 0x200,
+	CUBEFLAG_OUT_RECHARGE = 0x400,
+
+	//Custom Flags
+	CUBEFLAG_OUT_KEEPONLYSOCKETABLES = 0x800,
+	CUBEFLAG_OUT_DUPE = 0x1000,
+};
+
+enum D2C_CubeInFlags
+{
+	CUBEFLAG_IN_USEANY = 0x01,
+	CUBEFLAG_IN_ITEMCODE = 0x02,
+	CUBEFLAG_IN_NOSOCKET = 0x04,
+	CUBEFLAG_IN_SOCKETED = 0x08,
+	CUBEFLAG_IN_ETHEREAL = 0x10,
+	CUBEFLAG_IN_NOETHEREAL = 0x20,
+	CUBEFLAG_IN_SPECIAL = 0x40,
+	CUBEFLAG_IN_UPGRADED = 0x80,
+	CUBEFLAG_IN_NORMAL = 0x100,
+	CUBEFLAG_IN_EXCEPTIONAL = 0x200,
+	CUBEFLAG_IN_ELITE = 0x400,
+	CUBEFLAG_IN_NORUNES = 0x800,
+};
+
 struct D2CubeInputItem
 {
 	uint16_t wInputFlags;				//0x00
