+++ source/D2Common/include/DataTbls/MissilesIds.h
@@ -1,0 +1,690 @@
+#pragma once
+
+enum D2C_MissilesHcIdx
+{
+	MISSILE_ARROW,
+	MISSILE_JAVELIN,
+	MISSILE_BIGHEAD1,
+	MISSILE_BIGHEAD2,
+	MISSILE_BIGHEAD3,
+	MISSILE_BIGHEAD4,
+	MISSILE_BIGHEAD5,
+	MISSILE_SPIKE1,
+	MISSILE_SPIKE2,
+	MISSILE_SPIKE3,
+	MISSILE_SPIKE4,
+	MISSILE_SPIKE5,
+	MISSILE_FIREARROW,
+	MISSILE_CR_ARROW1,
+	MISSILE_CR_ARROW2,
+	MISSILE_CR_ARROW3,
+	MISSILE_CR_ARROW4,
+	MISSILE_CR_ARROW5,
+	MISSILE_BLOOD1,
+	MISSILE_BLOOD2,
+	MISSILE_BIGBLOOD1,
+	MISSILE_BIGBLOOD2,
+	MISSILE_SHAFIRE1,
+	MISSILE_SHAFIRE2,
+	MISSILE_SHAFIRE3,
+	MISSILE_SHAFIRE4,
+	MISSILE_SHAFIRE5,
+	MISSILE_MAGICARROW,
+	MISSILE_ICEARROW,
+	MISSILE_FIREEXPLODE,
+	MISSILE_ICEEXPLODE,
+	MISSILE_BOLT,
+	MISSILE_ANDARIELSPRAY,
+	MISSILE_BIGHEADEXP,
+	MISSILE_SHAMANEXP,
+	MISSILE_THROWAXE,
+	MISSILE_THROWKNIFE,
+	MISSILE_GLAIVE,
+	MISSILE_POISONJAV,
+	MISSILE_POISONJAVCLOUD,
+	MISSILE_COLDARROW,
+	MISSILE_EXPLODINGARROW,
+	MISSILE_EXPLODINGARROWEXP,
+	MISSILE_PLAGUEJAVELIN,
+	MISSILE_OILPOTION,
+	MISSILE_EXPLOSIVEPOTION,
+	MISSILE_FULMINATINGPOTION,
+	MISSILE_RANCIDGASEPOTION,
+	MISSILE_CHOKINGGASPOITION,
+	MISSILE_STRANGLINGGASPOTION,
+	MISSILE_NOTUSED50,
+	MISSILE_EXPLOSIVEPOTIONEXP,
+	MISSILE_EXPLOSIVEPOTIONDEBRIS1,
+	MISSILE_EXPLOSIVEPOTIONDEBRIS2,
+	MISSILE_EXPLOSIVEPOTIONDEBRIS3,
+	MISSILE_HOLYBOLT,
+	MISSILE_CHARGEDBOLT,
+	MISSILE_SANCTUARYBOLT,
+	MISSILE_FIREBOLT,
+	MISSILE_ICEBOLT,
+	MISSILE_INFERNOFLAME1,
+	MISSILE_INFERNOFLAME2,
+	MISSILE_FIREBALL,
+	MISSILE_MUMMY1,
+	MISSILE_MUMMY2,
+	MISSILE_MUMMY3,
+	MISSILE_MUMMY4,
+	MISSILE_BLAZE,
+	MISSILE_FIREWALLMAKER,
+	MISSILE_FIREWALL,
+	MISSILE_GOOSPIT1,
+	MISSILE_GOOSPIT2,
+	MISSILE_GOOSPIT3,
+	MISSILE_GOOSPIT4,
+	MISSILE_GOOSPIT5,
+	MISSILE_GOOSPLAT,
+	MISSILE_SAND_PILE,
+	MISSILE_UNHOLYBOLT1,
+	MISSILE_UNHOLYBOLT2,
+	MISSILE_UNHOLYBOLT3,
+	MISSILE_UNHOLYBOLT4,
+	MISSILE_SANCTUARYCENTER,
+	MISSILE_FIREEXPLOSION,
+	MISSILE_STUCKARROW,
+	MISSILE_FOOTPRINT,
+	MISSILE_IMMOLATIONARROW,
+	MISSILE_GUIDEDARROW,
+	MISSILE_FREEZINGARROW,
+	MISSILE_FREEZINGARROWEXP1,
+	MISSILE_FREEZINGARROWEXP2,
+	MISSILE_NOVA,
+	MISSILE_ICEBLAST,
+	MISSILE_BLESSEDHAMMER,
+	MISSILE_CHAINLIGHTNING,
+	MISSILE_FISTOFARES,
+	MISSILE_CHILLBLOOD,
+	MISSILE_GLACIALSPIKE,
+	MISSILE_TELEPORT,
+	MISSILE_LIGHTNINGBOLT,
+	MISSILE_LIGHTNINGHIT,
+	MISSILE_METEOR,
+	MISSILE_METEORCENTER,
+	MISSILE_METEORTAIL,
+	MISSILE_METEOREXPLODE,
+	MISSILE_FIRESMALL,
+	MISSILE_FIREMEDIUM,
+	MISSILE_MONBLIZCENTER,
+	MISSILE_MONBLIZ1,
+	MISSILE_MONBLIZ2,
+	MISSILE_MONBLIZ3,
+	MISSILE_MONBLIZ4,
+	MISSILE_MONBLIZEXPLODE1,
+	MISSILE_MONBLIZEXPLODE2,
+	MISSILE_MONBLIZEXPLODE3,
+	MISSILE_TEETH,
+	MISSILE_CORPSEEXPLOSION,
+	MISSILE_POISONCORPSEEXPLOSION,
+	MISSILE_MONSTERCORPSEEXPLODE,
+	MISSILE_POISONNOVA,
+	MISSILE_FROSTNOVA,
+	MISSILE_ROGUE1,
+	MISSILE_ROGUE2,
+	MISSILE_ROGUE3,
+	MISSILE_BAT_LIGHTNING_BOLT,
+	MISSILE_BAT_LIGHTNING_TRAIL,
+	MISSILE_SKMAGE1,
+	MISSILE_SKMAGE2,
+	MISSILE_SKMAGE3,
+	MISSILE_SKMAGE4,
+	MISSILE_VAMPIREFIREBALL,
+	MISSILE_VAMPIREFIREWALLMAKER,
+	MISSILE_VAMPIREFIREWALL,
+	MISSILE_VAMPIREMETEOR,
+	MISSILE_VAMPIREMETEORCENTER,
+	MISSILE_VAMPIREMETEOREXP,
+	MISSILE_RAVEN1,
+	MISSILE_RAVEN2,
+	MISSILE_AMPHIBIANGOO1,
+	MISSILE_AMPHIBIANGOO2,
+	MISSILE_TENTACLEGOO,
+	MISSILE_AMPHIBIANEXPLODE,
+	MISSILE_POISONPUFF,
+	MISSILE_CURSEEFFECTRED,
+	MISSILE_SPIDERGOOLAY,
+	MISSILE_FETISHINFERNO1,
+	MISSILE_FETISHINFERNO2,
+	MISSILE_SPIDERGOO,
+	MISSILE_CURSECAST,
+	MISSILE_HOWL,
+	MISSILE_SHOUT,
+	MISSILE_DUST,
+	MISSILE_REDLIGHTMISSILE,
+	MISSILE_GREENLIGHTMISSILE,
+	MISSILE_BLUELIGHTMISSILE,
+	MISSILE_WHITELIGHTMISSILE,
+	MISSILE_CORPSEPOISONCLOUD,
+	MISSILE_CHILLBLOODCLOUD,
+	MISSILE_CHILLBLOODPUFF,
+	MISSILE_BLIZZARDCENTER,
+	MISSILE_BLIZZARD1,
+	MISSILE_BLIZZARD2,
+	MISSILE_BLIZZARD3,
+	MISSILE_BLIZZARD4,
+	MISSILE_BLIZZARDEXPLODE1,
+	MISSILE_BLIZZARDEXPLODE2,
+	MISSILE_BLIZZARDEXPLODE3,
+	MISSILE_THUNDERSTORM1,
+	MISSILE_THUNDERSTORM2,
+	MISSILE_THUNDERSTORM3,
+	MISSILE_THUNDERSTORM4,
+	MISSILE_MONSTERLIGHT,
+	MISSILE_POISONBALL,
+	MISSILE_DIABLIGHT,
+	MISSILE_REDEMPTION,
+	MISSILE_REDEMPTIONFAIL,
+	MISSILE_HANDOFGOD,
+	MISSILE_DIABFIRE,
+	MISSILE_FINGERMAGESPIDER,
+	MISSILE_ELECTRIC_THROWAXE,
+	MISSILE_DIABWALLMAKER,
+	MISSILE_DIABWALL,
+	MISSILE_CURSEAMPLIFYDAMAGE,
+	MISSILE_CURSEDIMVISION,
+	MISSILE_CURSEWEAKEN,
+	MISSILE_CURSEIRONMAIDEN,
+	MISSILE_CURSETERROR,
+	MISSILE_CURSEATTRACT,
+	MISSILE_CURSEREVERSEVAMPIRE,
+	MISSILE_CURSECONFUSE,
+	MISSILE_CURSEDECREPIFY,
+	MISSILE_CURSELOWERRESIST,
+	MISSILE_CURSECENTER,
+	MISSILE_BONESPEAR,
+	MISSILE_BONESPIRIT,
+	MISSILE_COLDUNIQUE,
+	MISSILE_LIGHTUNIQUE,
+	MISSILE_SKBOWARROW1,
+	MISSILE_SKBOWARROW2,
+	MISSILE_SKBOWARROW3,
+	MISSILE_SKBOWARROW4,
+	MISSILE_SKBOWARROW5,
+	MISSILE_NOVA1,
+	MISSILE_NOVA2,
+	MISSILE_ANDYPOISONBOLT,
+	MISSILE_TEETHEXPLODE,
+	MISSILE_LIGHTNINGJAVELIN,
+	MISSILE_LIGHTNINGFURY,
+	MISSILE_BONEWALLMAKER,
+	MISSILE_NECROMAGE1,
+	MISSILE_NECROMAGE2,
+	MISSILE_NECROMAGE3,
+	MISSILE_NECROMAGE4,
+	MISSILE_SPARKLE,
+	MISSILE_MULTIPLESHOTARROW,
+	MISSILE_MULTIPLESHOTBOLT,
+	MISSILE_CHARGEDSTRIKEBOLT,
+	MISSILE_BONESPEAREXPLODE,
+	MISSILE_POISONEXPLOSIONCLOUD,
+	MISSILE_BONECAST,
+	MISSILE_BATTLECRY,
+	MISSILE_PRIMEPOISONCLOUD,
+	MISSILE_PLAGUEJAVCLOUD,
+	MISSILE_RANCIDGASCLOUD,
+	MISSILE_CHOKINGGASCLOUD,
+	MISSILE_STRANGLINGGASCLOUD,
+	MISSILE_BUGLIGHTNING,
+	MISSILE_PANTHERJAV1,
+	MISSILE_PANTHERJAV2,
+	MISSILE_PANTHERJAV3,
+	MISSILE_PANTHERJAV4,
+	MISSILE_IMMOLATIONFIRE,
+	MISSILE_FURYLIGHTNING,
+	MISSILE_LIGHTNINGSTRIKE,
+	MISSILE_FISTOFTHEHEAVENSDELAY,
+	MISSILE_FISTOFTHEHEAVENSBOLT,
+	MISSILE_WARCRY,
+	MISSILE_BATTLECOMMAND,
+	MISSILE_BATTLEORDERS,
+	MISSILE_PANTHERPOTORANGE,
+	MISSILE_PANTHERPOTGREEN,
+	MISSILE_METEORFIRE,
+	MISSILE_TRAP_SPIKE_RIGHT,
+	MISSILE_TRAP_SPIKE_LEFT,
+	MISSILE_TRAP_CURSED_SKULL_RIGHT,
+	MISSILE_TRAP_CURSED_SKULL_LEFT,
+	MISSILE_TRAP_POISON_BALL_RIGHT,
+	MISSILE_TRAP_POISON_BALL_LEFT,
+	MISSILE_HYDRA,
+	MISSILE_BONESPEARTRAIL,
+	MISSILE_GRIMWARDSMALLSTART,
+	MISSILE_GRIMWARDSMALL,
+	MISSILE_GRIMWARDSMALLSTOP,
+	MISSILE_GRIMWARDMEDIUMSTART,
+	MISSILE_GRIMWARDMEDIUM,
+	MISSILE_GRIMWARDMEDIUMSTOP,
+	MISSILE_GRIMWARDLARGESTART,
+	MISSILE_GRIMWARDLARGE,
+	MISSILE_GRIMWARDLARGESTOP,
+	MISSILE_ZAKARUMLIGHT,
+	MISSILE_GRIMWARDSCARE,
+	MISSILE_FROZENORB,
+	MISSILE_FROZENORBBOLT,
+	MISSILE_FROZENORBNOVA,
+	MISSILE_FROZENORBEXPLODE,
+	MISSILE_CHILLINGARMORBOLT,
+	MISSILE_FIREEXPLOSION2,
+	MISSILE_BLOWGUN,
+	MISSILE_CHAINLIGHTNING2,
+	MISSILE_REVIVESMALL,
+	MISSILE_REVIVEMEDIUM,
+	MISSILE_REVIVELARGE,
+	MISSILE_MONGLACIALSPIKE,
+	MISSILE_ICEBREAKSMALL,
+	MISSILE_ICEBREAKMEDIUM,
+	MISSILE_ICEBREAKLARGE,
+	MISSILE_ICEBREAKSMOKE,
+	MISSILE_MEPHISTO,
+	MISSILE_FIREHEAD,
+	MISSILE_WHILRWIND,
+	MISSILE_ARCANELIGHTNINGBOLT,
+	MISSILE_FROGFIRE,
+	MISSILE_FROGCOLD,
+	MISSILE_FROGPOIS,
+	MISSILE_DESERTFIREBALL,
+	MISSILE_BRDEATHCONTROL,
+	MISSILE_BRDEATHLIGHTNINGBOLT,
+	MISSILE_BRDEATHLIGHTNINGHIT,
+	MISSILE_DENOFEVILLIGHT,
+	MISSILE_CAIRNSTONES,
+	MISSILE_CAIRNSTONESSKY,
+	MISSILE_CAIRNSTONESGROUND,
+	MISSILE_TOWERMIST,
+	MISSILE_TOWERMISTTRAIL,
+	MISSILE_BRDEATHSMOKES1,
+	MISSILE_BRDEATHSMOKENU,
+	MISSILE_BRDEATHSMOKEDT,
+	MISSILE_BRDEATHSPIRITS1,
+	MISSILE_BRDEATHSPIRITNU,
+	MISSILE_BRDEATHSPIRITDT,
+	MISSILE_MEPHISTODEATHCONTROL,
+	MISSILE_MEPHISTOFIREWALLMAKER,
+	MISSILE_MEPHISTOFIREWALL,
+	MISSILE_MEPHISTOFLYINGROCKSBIG,
+	MISSILE_MEPHISTOEXPLOSIONBIG,
+	MISSILE_MEPHISTOFLYINGROCKSSMALL,
+	MISSILE_MEPHISTOEXPLOSIONSMALL,
+	MISSILE_MEPHISTODONOTDRAW,
+	MISSILE_ANDYCONTROL0,
+	MISSILE_ANDYFIREWALLMAKER,
+	MISSILE_ANDYFIREWALL,
+	MISSILE_ANDYCOLUMNFIREBASE,
+	MISSILE_ANDYCOLUMNFIRE,
+	MISSILE_ANDYFALLINGDEBRIS1,
+	MISSILE_ANDYFALLINGDEBRIS2,
+	MISSILE_ANDYFALLINGDEBRIS3,
+	MISSILE_ANDYDEBRISEXPLOSION1,
+	MISSILE_ANDYDEBRISEXPLOSION2,
+	MISSILE_ANDYDEBRISEXPLOSION3,
+	MISSILE_ANDYDEBRISEXPLOSION4,
+	MISSILE_ANDYDEBRISEXPLOSION5,
+	MISSILE_WILLOWISPLIGHTNINGBOLT,
+	MISSILE_QUEENPOISONCLOUD,
+	MISSILE_DIRT_PILE,
+	MISSILE_UNDEADMISSILE1,
+	MISSILE_UNDEADMISSILE2,
+	MISSILE_UNDEADMISSILE3,
+	MISSILE_UNDEADMISSILE4,
+	MISSILE_BONESPIRITEXPLODE,
+	MISSILE_DOPPLEZONEXPLODE,
+	MISSILE_MONBONESPIRIT,
+	MISSILE_TOWERMISTFADE,
+	MISSILE_COUNTESSFIREWALL,
+	MISSILE_TOWERCHESTSPAWNER,
+	MISSILE_HELLMETEORLAUNCH1,
+	MISSILE_HELLMETEORLAUNCH2,
+	MISSILE_HELLMETEORUP,
+	MISSILE_HELLMETEORDOWN,
+	MISSILE_HELLMETEORBALL,
+	MISSILE_HORADRICSTAFF,
+	MISSILE_HORADRICLIGHTNING,
+	MISSILE_HORADRICLIGHT,
+	MISSILE_REGURGITATORCORPSE,
+	MISSILE_REGURGITATORCORPSEEXPL,
+	MISSILE_HIGHPRIESTLIGHTNING,
+	MISSILE_ICEBREAKSMALLMELT,
+	MISSILE_ICEBREAKLARGEMELT,
+	MISSILE_LEAPKNOCKBACK,
+	MISSILE_RADAMENTDEATH,
+	MISSILE_RADAMENTHANDOFGOD,
+	MISSILE_RADAMENTHOLYBOLT,
+	MISSILE_TAINTEDSUNCONTROL,
+	MISSILE_TAINTEDSUNFLASH,
+	MISSILE_TAINTEDSUNBALL,
+	MISSILE_QUEENDEATHCENTER,
+	MISSILE_QUEENDEATHGLOB,
+	MISSILE_QUEENDEATHSPLAT1,
+	MISSILE_QUEENDEATHSPLAT2,
+	MISSILE_HEALINGBOLT,
+	MISSILE_MEPHISTOHOLEDELAY,
+	MISSILE_MEPHISTOHOLEBIRTH,
+	MISSILE_MEPHISTOHOLENEUTRAL,
+	MISSILE_MEPHISTOHOLEDEATH,
+	MISSILE_MEPHISTOHOLEDEAD,
+	MISSILE_DURIELDEATHCONTROL,
+	MISSILE_DURIELDEATHROCK,
+	MISSILE_DURIELDEATHDEBRIS,
+	MISSILE_DURIELDEATHSMOKE,
+	MISSILE_MEPHISTOEXPLOSION,
+	MISSILE_ORBMIST,
+	MISSILE_ORBMISTTRAIL,
+	MISSILE_ORBMISTFADE,
+	MISSILE_PILUM,
+	MISSILE_DIABLO_APPEARS,
+	MISSILE_HFCONTROL,
+	MISSILE_HFFRAGMENT1,
+	MISSILE_HFFRAGMENT2,
+	MISSILE_HFFRAGMENT3,
+	MISSILE_HFSPIRIT1,
+	MISSILE_HFRESERVED3,
+	MISSILE_IZUAL_CONTROL,
+	MISSILE_IZUAL_MIST_LOOP,
+	MISSILE_IZUAL_MIST_FADE,
+	MISSILE_IZUAL_LIGHTNING,
+	MISSILE_IZUAL_LIGHTNING_TRAIL,
+	MISSILE_CAIRN_STONES_BOLT,
+	MISSILE_BOMB_IN_AIR,
+	MISSILE_BOMB_ON_GROUND,
+	MISSILE_BOMB_EXPLOSION,
+	MISSILE_SHOCK_FIELD_IN_AIR,
+	MISSILE_SHOCK_FIELD_ON_GROUND,
+	MISSILE_THROWINGSTAR,
+	MISSILE_ACIDSPRAY,
+	MISSILE_BLADE_CREEPER,
+	MISSILE_DISTRACTION,
+	MISSILE_DISTRACTION_FOG,
+	MISSILE_DISTRACTION_PUFF,
+	MISSILE_DISTRACTION_START,
+	MISSILE_DISTRACTION_END,
+	MISSILE_IMPINFERNOFLAME1,
+	MISSILE_IMPINFERNOFLAME2,
+	MISSILE_BAALLIGHTNINGBOLT,
+	MISSILE_BAALLIGHTNINGTRAIL,
+	MISSILE_BAALLIGHTNINGBOLT2,
+	MISSILE_BAALLIGHTNINGTRAIL2,
+	MISSILE_IMPFIREBALL,
+	MISSILE_IMPFIREBALLEXPLODE,
+	MISSILE_CATAPULTCHARGEDBALL_ON,
+	MISSILE_CATAPULTCHARGEDBALL,
+	MISSILE_CATAPULTCHARGEDBALLBOLT,
+	MISSILE_IMP_SPAWN_MONSTERS,
+	MISSILE_CATAPULT_SPIKE_BALL_ON,
+	MISSILE_CATAPULT_SPIKE_BALL,
+	MISSILE_CATAPULT_SPIKE_IN_AIR,
+	MISSILE_CATAPULT_SPIKE_ON_GROUND,
+	MISSILE_CATAPULT_SPIKE_EXPLOSION,
+	MISSILE_CATAPULT_COLD_BALL_ON,
+	MISSILE_CATAPULT_COLD_BALL,
+	MISSILE_CATAPULT_COLD_EXPLOSION,
+	MISSILE_CATAPULT_PLAGUE_BALL_ON,
+	MISSILE_CATAPULT_PLAGUE_BALL,
+	MISSILE_CATAPULT_PLAGUE_CLOUD,
+	MISSILE_CATAPULT_METEOR_BALL_ON,
+	MISSILE_CATAPULT_METEOR_BALL,
+	MISSILE_CATAPULT_METEOR_FIRE,
+	MISSILE_TOWER_DEATH,
+	MISSILE_HEALING_VORTEX,
+	MISSILE_SUICIDECORPSEEXPLODE,
+	MISSILE_SUICIDEFIREEXPLODE,
+	MISSILE_SUICIDEICEEXPLODE,
+	MISSILE_EXPLODINGJAVALIN,
+	MISSILE_EXPLODINGJAVALINEXP,
+	MISSILE_LIGHTINGTRAILINGJAVALIN,
+	MISSILE_LIGHTJAVALINTRAIL,
+	MISSILE_LIGHTJAVALINEXPLOSION,
+	MISSILE_ICEJAVALIN,
+	MISSILE_ICEJAVALINEXPLODE,
+	MISSILE_PLAGUEJAVELIN2,
+	MISSILE_PLAGUEJAVLINEXPLODE,
+	MISSILE_ADVLIGHTTRAILINGJAV,
+	MISSILE_ADVLIGHTTRAILINGJAV2,
+	MISSILE_ADVLIGHTJAVEXPLODE,
+	MISSILE_SUCFIREBALL,
+	MISSILE_SUCFIREBALLEXPLODE,
+	MISSILE_SUCFIREBALLTRAIL,
+	MISSILE_SUCSHOCKFIELDMISSILE,
+	MISSILE_SUCSHOCKFIELDMISSILEEXP,
+	MISSILE_SUCSHOCKFIELD,
+	MISSILE_HELLFIREMISSILE,
+	MISSILE_HELLFIREEXA,
+	MISSILE_HELLFIREEXB,
+	MISSILE_IMP_CHARGED_BOLT,
+	MISSILE_IMP_TELEPORT,
+	MISSILE_MOLTENBOULDER,
+	MISSILE_MOLTENBOULDEREMERGE,
+	MISSILE_MOLTENBOULDEREXPLODE,
+	MISSILE_MOLTENBOULDERFIREPATH,
+	MISSILE_MOLTENBOULDER_FLYINGROCKS,
+	MISSILE_FIRESTORM,
+	MISSILE_FIRESTORMMAKER,
+	MISSILE_ARCTICBLAST1,
+	MISSILE_ARCTICBLAST2,
+	MISSILE_ERRUPTION_CENTER,
+	MISSILE_ERRUPTION_CRACK_1,
+	MISSILE_ERRUPTION_CRACK_2,
+	MISSILE_ERRUPTION_SMOKE_1,
+	MISSILE_ERRUPTION_SMOKE_2,
+	MISSILE_VINE_BEAST_WALK_1,
+	MISSILE_VINE_BEAST_WALK_2,
+	MISSILE_VINE_BEAST_NEUTRAL,
+	MISSILE_VINE_BEAST_ATTACK,
+	MISSILE_VINE_BEAST_DEATH,
+	MISSILE_VINES,
+	MISSILE_VINES_TRAIL,
+	MISSILE_VINES_WITHER,
+	MISSILE_PLAGUE_VINES,
+	MISSILE_PLAGUE_VINES_TRAIL,
+	MISSILE_PLAGUE_VINES_WITHER,
+	MISSILE_TWISTER,
+	MISSILE_TORNADO,
+	MISSILE_VOLCANO,
+	MISSILE_VOLCANO_OVERLAY_FIRE,
+	MISSILE_VOLCANO_DEBRIS_2,
+	MISSILE_VOLCANO_EXPLOSION,
+	MISSILE_VOLCANO_SMALL_FIRE,
+	MISSILE_DRAGONBREATH_MISSILE,
+	MISSILE_LUREPROJECTILE,
+	MISSILE_LURECENTER,
+	MISSILE_LURECLOUD,
+	MISSILE_IMPMISS1,
+	MISSILE_IMPMISS2,
+	MISSILE_IMPMISS3,
+	MISSILE_IMPMISS4,
+	MISSILE_IMPMISS5,
+	MISSILE_FROZENHORROR_ARCTICBLAST1,
+	MISSILE_FROZENHORROR_ARCTICBLAST2,
+	MISSILE_SENTRYCHARGEDBOLT,
+	MISSILE_SENTRYSPIKEINAIR,
+	MISSILE_SENTRYSPIKEONGROUND,
+	MISSILE_RECYCLER_DELAY,
+	MISSILE_RECYCLER_VINE,
+	MISSILE_RECYCLER_FADE,
+	MISSILE_RECYCLER_EXPLOSION,
+	MISSILE_DEATH_MAULER,
+	MISSILE_DEATH_MAULER_TRAIL,
+	MISSILE_DEATH_MAULER_TRAIL_FADE,
+	MISSILE_BLADEFURY1,
+	MISSILE_BLADEFRAGMENT1,
+	MISSILE_BLADEFURY2,
+	MISSILE_BLADEFRAGMENT2,
+	MISSILE_BLADEFURY3,
+	MISSILE_BLADEFRAGMENT3,
+	MISSILE_SHOCKWAVE,
+	MISSILE_LIGHTNINGTALONS,
+	MISSILE_LIGHTNINGTALONSTRAIL,
+	MISSILE_PHOENIXTRAIL,
+	MISSILE_RABIESPLAGUE,
+	MISSILE_RABIESCONTAGION,
+	MISSILE_WAKE_OF_DESTRUCTION_MAKER,
+	MISSILE_WAKE_OF_DESTRUCTION,
+	MISSILE_DEATHSENTRYEXPLODE,
+	MISSILE_TIGERFURY,
+	MISSILE_TIGERFURYTRAIL,
+	MISSILE_TIGERFURYTRAIL2,
+	MISSILE_INFERNO_SENTRY_1,
+	MISSILE_INFERNO_SENTRY_2,
+	MISSILE_ANCIENT_THROWING_AXE,
+	MISSILE_SENTRYLIGHTNINGBOLT,
+	MISSILE_SENTRYLIGHTNINGHIT,
+	MISSILE_ANYA_CENTER,
+	MISSILE_ANYA_ICICLE,
+	MISSILE_ANYA_ICEIMPACT,
+	MISSILE_ANYA_ICESTEAM,
+	MISSILE_ANYA_ICEMAGIC,
+	MISSILE_DRAGONTAIL_MISSILE,
+	MISSILE_DRAGONFLIGHT,
+	MISSILE_DRAGONFLIGHTMAKER,
+	MISSILE_PROGRESSIVE_RADIUS_DAMAGE,
+	MISSILE_VINE_BEAST_WALK_1_FADE,
+	MISSILE_VINE_BEAST_WALK_2_FADE,
+	MISSILE_VINE_BEAST_NEUTRAL_FADE,
+	MISSILE_VINE_RECYCLER_DELAY,
+	MISSILE_ANCIENT_DEATH_CENTER,
+	MISSILE_ANCIENT_DEATH_CLOUD,
+	MISSILE_LIGHTNING_CHARGE_UP_NOVA,
+	MISSILE_CHAINLIGHTNINGCHARGE_UP,
+	MISSILE_PAIN_WORM_APPEAR,
+	MISSILE_BAAL_TAUNT_CONTROL,
+	MISSILE_BAAL_TAUNT_LIGHTNING,
+	MISSILE_BAAL_TAUNT_LIGHTNING_TRAIL,
+	MISSILE_BAAL_TAUNT_POISON,
+	MISSILE_BAAL_SPAWN_MONSTERS,
+	MISSILE_MINDBLAST_HIT,
+	MISSILE_BLADE_SHIELD_MISSILE,
+	MISSILE_BLADE_SHIELD_ATTACHMENT,
+	MISSILE_BAAL_INFERNO,
+	MISSILE_BAAL_NOVA,
+	MISSILE_FISTSOFFIREEXPLODE,
+	MISSILE_FISTSOFFIREFIREWALL,
+	MISSILE_CLAWSOFTHUNDERBOLT,
+	MISSILE_CLAWSOFTHUNDERNOVA,
+	MISSILE_BLADESOFICEEXPLODE,
+	MISSILE_BLADESOFICECUBES,
+	MISSILE_BLADESOFICECUBESMELT,
+	MISSILE_ROYALSTRIKEMETEOR,
+	MISSILE_ROYALSTRIKEMETEORCENTER,
+	MISSILE_ROYALSTRIKEMETEORTAIL,
+	MISSILE_ROYALSTRIKEMETEOREXPLODE,
+	MISSILE_ROYALSTRIKEMETEORFIRE,
+	MISSILE_ROYALSTRIKECHAINLIGHTNING,
+	MISSILE_ROYALSTRIKECHAOSICE,
+	MISSILE_WORLD_STONE_CHIP_1,
+	MISSILE_WORLD_STONE_CHIP_2,
+	MISSILE_WORLD_STONE_CHIP_3,
+	MISSILE_WORLD_STONE_CHIP_4,
+	MISSILE_HIGHPRIESTLIGHTNING2,
+	MISSILE_INFERNOFLAME3,
+	MISSILE_MINDBLAST_CENTER,
+	MISSILE_ARMAGEDDONCONTROL,
+	MISSILE_ARMAGEDDONROCK,
+	MISSILE_ARMAGEDDONTAIL,
+	MISSILE_ARMAGEDDONEXPLOSION,
+	MISSILE_HURRICANESWOOSH,
+	MISSILE_HURRICANECART,
+	MISSILE_HURRICANEROCK,
+	MISSILE_HURRICANESACK,
+	MISSILE_HURRICANETREE,
+	MISSILE_HURRICANEVASE,
+	MISSILE_BAALCORPSEEXPLODEDELAY,
+	MISSILE_BAALCORPSEEXPLODEEXPL,
+	MISSILE_BAAL_COLD_MAKER,
+	MISSILE_BAAL_COLD_TRAIL,
+	MISSILE_BAAL_SPAWN_MONSTERS_EXP,
+	MISSILE_IMPMISS21,
+	MISSILE_IMPMISS22,
+	MISSILE_IMPMISS23,
+	MISSILE_IMPMISS24,
+	MISSILE_IMPMISS25,
+	MISSILE_ANYASTEAM1,
+	MISSILE_ANYASTEAM2,
+	MISSILE_ANCIENTSGUIDE,
+	MISSILE_ANCIENTSMARKER,
+	MISSILE_ANCIENTSCONTROL,
+	MISSILE_OVERSEERCONTROL,
+	MISSILE_NIHLITHAK1,
+	MISSILE_NIHLITHAK2,
+	MISSILE_NIHLITHAK3,
+	MISSILE_NEHLITHAKCONTROL,
+	MISSILE_NEHLITHAKSWOOSH,
+	MISSILE_NEHLITHAKDEBRIS1,
+	MISSILE_NEHLITHAKDEBRIS2,
+	MISSILE_NEHLITHAKDEBRIS3,
+	MISSILE_NEHLITHAKDEBRIS4,
+	MISSILE_NEHLITHAKGLOW,
+	MISSILE_BAALTELEPORT,
+	MISSILE_BAALCLONEDEATH,
+	MISSILE_ANYASTEAMVENT,
+	MISSILE_ANYASTEAM,
+	MISSILE_NEHLITHAKHOLE,
+	MISSILE_NEHLITHAKHOLELIGHT,
+	MISSILE_VOLCANOFIRETRAIL,
+	MISSILE_NEHLITHAKGLOW2,
+	MISSILE_NEHLITHAKBONECHIPS,
+	MISSILE_BAALCORPSEEXPLODEFADE,
+	MISSILE_ARMAGEDDONFIRE,
+	MISSILE_ICESPARKLE,
+	MISSILE_BAALFX_CONTROL,
+	MISSILE_BAALFX_SPIRIT_1,
+	MISSILE_BAALFX_SPIRIT_2,
+	MISSILE_BAALFX_SPIRIT_3,
+	MISSILE_BAALFX_SPIRIT_4,
+	MISSILE_BAALFX_SPIRIT_5,
+	MISSILE_BAALFX_BAAL_HEAD_APPEAR,
+	MISSILE_BAALFX_BAAL_HEAD_1,
+	MISSILE_BAALFX_BAAL_HEAD_2,
+	MISSILE_BAALFX_BAAL_HEAD_3,
+	MISSILE_BAALFX_TYREAL_DEBRIS_1,
+	MISSILE_BAALFX_TYREAL_DEBRIS_2,
+	MISSILE_BAALFX_TYREAL_DEBRIS_3,
+	MISSILE_BAALFX_TYREAL_DEBRIS_BREAK,
+	MISSILE_WORLDSTONE_SHAKE,
+	MISSILE_BLESSEDHAMMEREX,
+	MISSILE_SENTRYLIGHTNINGBOLT2,
+	MISSILE_SENTRYLIGHTNINGHIT2,
+	MISSILE_LIGHTNINGTOWERNOVA,
+	MISSILE_SKBOWARROW6,
+	MISSILE_SKBOWARROW7,
+	MISSILE_SKBOWARROW8,
+	MISSILE_BIGHEAD6,
+	MISSILE_VIPER_POISJAV,
+	MISSILE_VIPER_POISJAVCLOUD,
+	MISSILE_VIPER_FIRE,
+	MISSILE_VIPER_FIRECLOUD,
+	MISSILE_VIPER_BONESPEAR,
+	MISSILE_COUNTESSFIREWALLMAKER,
+	MISSILE_BAAL_TAUNT_LIGHTNING_CONTROL,
+	MISSILE_BAAL_TAUNT_POISON_CONTROL,
+	MISSILE_EXPLODINGARROWEXP2,
+	MISSILE_FREEZINGARROWEXP3,
+	MISSILE_PANTHERJAV5,
+	MISSILE_SPIKE6,
+	MISSILE_CR_ARROW6,
+	MISSILE_SKMAGEPOIS,
+	MISSILE_SKMAGECOLD,
+	MISSILE_SKMAGEFIRE,
+	MISSILE_SKMAGELTNG,
+	MISSILE_SUCCUBUSMISS,
+	MISSILE_WILLOWISPLIGHTNINGBOLT2,
+	MISSILE_MUMMYEX,
+	MISSILE_GOOSPITEX,
+	MISSILE_IMPMISSEX,
+	MISSILE_DIABLOGEDDONCONTROL,
+	MISSILE_DIABLOGEDDONROCK,
+	MISSILE_DIABLOGEDDONTAIL,
+	MISSILE_DIABLOGEDDONEXPLOSION,
+	MISSILE_DIABLOGEDDONFIRE,
+	MISSILE_MEGADEMONINFERNO,
+	MISSILE_TRAPFIREBOLT,
+	MISSILE_TRAPPOISONJAVCLOUD,
+	MISSILE_TRAPNOVA,
+	MISSILE_MEPHFROSTNOVA,
+	MISSILE_MEPHLIGHT,
+	MISSILE_VAMPIREMETEORFIRE,
+	MISSILE_STRAFEARROW,
+	MISSILE_STRAFEBOLT,
+	MISSILE_SIEGE_BEAST_STOMP_MISSILE
+};
