+++ source/D2Common/include/DataTbls/MonsterIds.h
@@ -1,0 +1,811 @@
+#pragma once 
+
+
+enum D2C_MonsterIds
+{
+	MONSTER_SKELETON1 = 0,
+	MONSTER_SKELETON2,
+	MONSTER_SKELETON3,
+	MONSTER_SKELETON4,
+	MONSTER_SKELETON5,
+	MONSTER_ZOMBIE1,
+	MONSTER_ZOMBIE2,
+	MONSTER_ZOMBIE3,
+	MONSTER_ZOMBIE4,
+	MONSTER_ZOMBIE5,
+	MONSTER_BIGHEAD1,
+	MONSTER_BIGHEAD2,
+	MONSTER_BIGHEAD3,
+	MONSTER_BIGHEAD4,
+	MONSTER_BIGHEAD5,
+	MONSTER_FOULCROW1,
+	MONSTER_FOULCROW2,
+	MONSTER_FOULCROW3,
+	MONSTER_FOULCROW4,
+	MONSTER_FALLEN1,
+	MONSTER_FALLEN2,
+	MONSTER_FALLEN3,
+	MONSTER_FALLEN4,
+	MONSTER_FALLEN5,
+	MONSTER_BRUTE2,
+	MONSTER_BRUTE3,
+	MONSTER_BRUTE4,
+	MONSTER_BRUTE5,
+	MONSTER_BRUTE1,
+	MONSTER_SANDRAIDER1,
+	MONSTER_SANDRAIDER2,
+	MONSTER_SANDRAIDER3,
+	MONSTER_SANDRAIDER4,
+	MONSTER_SANDRAIDER5,
+	MONSTER_GORGON1,
+	MONSTER_GORGON2,
+	MONSTER_GORGON3,
+	MONSTER_GORGON4,
+	MONSTER_WRAITH1,
+	MONSTER_WRAITH2,
+	MONSTER_WRAITH3,
+	MONSTER_WRAITH4,
+	MONSTER_WRAITH5,
+	MONSTER_CORRUPTROGUE1,
+	MONSTER_CORRUPTROGUE2,
+	MONSTER_CORRUPTROGUE3,
+	MONSTER_CORRUPTROGUE4,
+	MONSTER_CORRUPTROGUE5,
+	MONSTER_BABOON1,
+	MONSTER_BABOON2,
+	MONSTER_BABOON3,
+	MONSTER_BABOON4,
+	MONSTER_BABOON5,
+	MONSTER_GOATMAN1,
+	MONSTER_GOATMAN2,
+	MONSTER_GOATMAN3,
+	MONSTER_GOATMAN4,
+	MONSTER_GOATMAN5,
+	MONSTER_FALLENSHAMAN1,
+	MONSTER_FALLENSHAMAN2,
+	MONSTER_FALLENSHAMAN3,
+	MONSTER_FALLENSHAMAN4,
+	MONSTER_FALLENSHAMAN5,
+	MONSTER_QUILLRAT1,
+	MONSTER_QUILLRAT2,
+	MONSTER_QUILLRAT3,
+	MONSTER_QUILLRAT4,
+	MONSTER_QUILLRAT5,
+	MONSTER_SANDMAGGOT1,
+	MONSTER_SANDMAGGOT2,
+	MONSTER_SANDMAGGOT3,
+	MONSTER_SANDMAGGOT4,
+	MONSTER_SANDMAGGOT5,
+	MONSTER_CLAWVIPER1,
+	MONSTER_CLAWVIPER2,
+	MONSTER_CLAWVIPER3,
+	MONSTER_CLAWVIPER4,
+	MONSTER_CLAWVIPER5,
+	MONSTER_SANDLEAPER1,
+	MONSTER_SANDLEAPER2,
+	MONSTER_SANDLEAPER3,
+	MONSTER_SANDLEAPER4,
+	MONSTER_SANDLEAPER5,
+	MONSTER_PANTHERWOMAN1,
+	MONSTER_PANTHERWOMAN2,
+	MONSTER_PANTHERWOMAN3,
+	MONSTER_PANTHERWOMAN4,
+	MONSTER_SWARM1,
+	MONSTER_SWARM2,
+	MONSTER_SWARM3,
+	MONSTER_SWARM4,
+	MONSTER_SCARAB1,
+	MONSTER_SCARAB2,
+	MONSTER_SCARAB3,
+	MONSTER_SCARAB4,
+	MONSTER_SCARAB5,
+	MONSTER_MUMMY1,
+	MONSTER_MUMMY2,
+	MONSTER_MUMMY3,
+	MONSTER_MUMMY4,
+	MONSTER_MUMMY5,
+	MONSTER_UNRAVELER1,
+	MONSTER_UNRAVELER2,
+	MONSTER_UNRAVELER3,
+	MONSTER_UNRAVELER4,
+	MONSTER_UNRAVELER5,
+	MONSTER_CHAOSHORDE1,
+	MONSTER_CHAOSHORDE2,
+	MONSTER_CHAOSHORDE3,
+	MONSTER_CHAOSHORDE4,
+	MONSTER_VULTURE1,
+	MONSTER_VULTURE2,
+	MONSTER_VULTURE3,
+	MONSTER_VULTURE4,
+	MONSTER_MOSQUITO1,
+	MONSTER_MOSQUITO2,
+	MONSTER_MOSQUITO3,
+	MONSTER_MOSQUITO4,
+	MONSTER_WILLOWISP1,
+	MONSTER_WILLOWISP2,
+	MONSTER_WILLOWISP3,
+	MONSTER_WILLOWISP4,
+	MONSTER_ARACH1,
+	MONSTER_ARACH2,
+	MONSTER_ARACH3,
+	MONSTER_ARACH4,
+	MONSTER_ARACH5,
+	MONSTER_THORNHULK1,
+	MONSTER_THORNHULK2,
+	MONSTER_THORNHULK3,
+	MONSTER_THORNHULK4,
+	MONSTER_VAMPIRE1,
+	MONSTER_VAMPIRE2,
+	MONSTER_VAMPIRE3,
+	MONSTER_VAMPIRE4,
+	MONSTER_VAMPIRE5,
+	MONSTER_BATDEMON1,
+	MONSTER_BATDEMON2,
+	MONSTER_BATDEMON3,
+	MONSTER_BATDEMON4,
+	MONSTER_BATDEMON5,
+	MONSTER_FETISH1,
+	MONSTER_FETISH2,
+	MONSTER_FETISH3,
+	MONSTER_FETISH4,
+	MONSTER_FETISH5,
+	MONSTER_CAIN1,
+	MONSTER_GHEED,
+	MONSTER_AKARA,
+	MONSTER_CHICKEN,
+	MONSTER_KASHYA,
+	MONSTER_RAT,
+	MONSTER_ROGUE1,
+	MONSTER_HELLMETEOR,
+	MONSTER_CHARSI,
+	MONSTER_WARRIV1,
+	MONSTER_ANDARIEL,
+	MONSTER_BIRD1,
+	MONSTER_BIRD2,
+	MONSTER_BAT,
+	MONSTER_CR_ARCHER1,
+	MONSTER_CR_ARCHER2,
+	MONSTER_CR_ARCHER3,
+	MONSTER_CR_ARCHER4,
+	MONSTER_CR_ARCHER5,
+	MONSTER_CR_LANCER1,
+	MONSTER_CR_LANCER2,
+	MONSTER_CR_LANCER3,
+	MONSTER_CR_LANCER4,
+	MONSTER_CR_LANCER5,
+	MONSTER_SK_ARCHER1,
+	MONSTER_SK_ARCHER2,
+	MONSTER_SK_ARCHER3,
+	MONSTER_SK_ARCHER4,
+	MONSTER_SK_ARCHER5,
+	MONSTER_WARRIV2,
+	MONSTER_ATMA,
+	MONSTER_DROGNAN,
+	MONSTER_FARA,
+	MONSTER_COW,
+	MONSTER_MAGGOTBABY1,
+	MONSTER_MAGGOTBABY2,
+	MONSTER_MAGGOTBABY3,
+	MONSTER_MAGGOTBABY4,
+	MONSTER_MAGGOTBABY5,
+	MONSTER_CAMEL,
+	MONSTER_BLUNDERBORE1,
+	MONSTER_BLUNDERBORE2,
+	MONSTER_BLUNDERBORE3,
+	MONSTER_BLUNDERBORE4,
+	MONSTER_MAGGOTEGG1,
+	MONSTER_MAGGOTEGG2,
+	MONSTER_MAGGOTEGG3,
+	MONSTER_MAGGOTEGG4,
+	MONSTER_MAGGOTEGG5,
+	MONSTER_ACT2MALE,
+	MONSTER_ACT2FEMALE,
+	MONSTER_ACT2CHILD,
+	MONSTER_GREIZ,
+	MONSTER_ELZIX,
+	MONSTER_GEGLASH,
+	MONSTER_JERHYN,
+	MONSTER_LYSANDER,
+	MONSTER_ACT2GUARD1,
+	MONSTER_ACT2VENDOR1,
+	MONSTER_ACT2VENDOR2,
+	MONSTER_CROWNEST1,
+	MONSTER_CROWNEST2,
+	MONSTER_CROWNEST3,
+	MONSTER_CROWNEST4,
+	MONSTER_MESHIF1,
+	MONSTER_DURIEL,
+	MONSTER_BONEFETISH1,
+	MONSTER_BONEFETISH2,
+	MONSTER_BONEFETISH3,
+	MONSTER_BONEFETISH4,
+	MONSTER_BONEFETISH5,
+	MONSTER_DARKGUARD1,
+	MONSTER_DARKGUARD2,
+	MONSTER_DARKGUARD3,
+	MONSTER_DARKGUARD4,
+	MONSTER_DARKGUARD5,
+	MONSTER_BLOODMAGE1,
+	MONSTER_BLOODMAGE2,
+	MONSTER_BLOODMAGE3,
+	MONSTER_BLOODMAGE4,
+	MONSTER_BLOODMAGE5,
+	MONSTER_MAGGOT,
+	MONSTER_SARCOPHAGUS,
+	MONSTER_RADAMENT,
+	MONSTER_FIREBEAST,
+	MONSTER_ICEGLOBE,
+	MONSTER_LIGHTNINGBEAST,
+	MONSTER_POISONORB,
+	MONSTER_FLYINGSCIMITAR,
+	MONSTER_ZEALOT1,
+	MONSTER_ZEALOT2,
+	MONSTER_ZEALOT3,
+	MONSTER_CANTOR1,
+	MONSTER_CANTOR2,
+	MONSTER_CANTOR3,
+	MONSTER_CANTOR4,
+	MONSTER_MEPHISTO,
+	MONSTER_DIABLO,
+	MONSTER_CAIN2,
+	MONSTER_CAIN3,
+	MONSTER_CAIN4,
+	MONSTER_FROGDEMON1,
+	MONSTER_FROGDEMON2,
+	MONSTER_FROGDEMON3,
+	MONSTER_SUMMONER,
+	MONSTER_TYRAEL1,
+	MONSTER_ASHEARA,
+	MONSTER_HRATLI,
+	MONSTER_ALKOR,
+	MONSTER_ORMUS,
+	MONSTER_IZUAL,
+	MONSTER_HALBU,
+	MONSTER_TENTACLE1,
+	MONSTER_TENTACLE2,
+	MONSTER_TENTACLE3,
+	MONSTER_TENTACLEHEAD1,
+	MONSTER_TENTACLEHEAD2,
+	MONSTER_TENTACLEHEAD3,
+	MONSTER_MESHIF2,
+	MONSTER_CAIN5,
+	MONSTER_NAVI,
+	MONSTER_BLOODRAVEN,
+	MONSTER_BUG,
+	MONSTER_SCORPION,
+	MONSTER_ROGUE2,
+	MONSTER_ROGUEHIRE,
+	MONSTER_ROGUE3,
+	MONSTER_GARGOYLETRAP,
+	MONSTER_SKMAGE_POIS1,
+	MONSTER_SKMAGE_POIS2,
+	MONSTER_SKMAGE_POIS3,
+	MONSTER_SKMAGE_POIS4,
+	MONSTER_FETISHSHAMAN1,
+	MONSTER_FETISHSHAMAN2,
+	MONSTER_FETISHSHAMAN3,
+	MONSTER_FETISHSHAMAN4,
+	MONSTER_FETISHSHAMAN5,
+	MONSTER_LARVA,
+	MONSTER_MAGGOTQUEEN1,
+	MONSTER_MAGGOTQUEEN2,
+	MONSTER_MAGGOTQUEEN3,
+	MONSTER_MAGGOTQUEEN4,
+	MONSTER_MAGGOTQUEEN5,
+	MONSTER_CLAYGOLEM,
+	MONSTER_BLOODGOLEM,
+	MONSTER_IRONGOLEM,
+	MONSTER_FIREGOLEM,
+	MONSTER_FAMILIAR,
+	MONSTER_ACT3MALE,
+	MONSTER_BABOON6,
+	MONSTER_ACT3FEMALE,
+	MONSTER_NATALYA,
+	MONSTER_VILEMOTHER1,
+	MONSTER_VILEMOTHER2,
+	MONSTER_VILEMOTHER3,
+	MONSTER_VILECHILD1,
+	MONSTER_VILECHILD2,
+	MONSTER_VILECHILD3,
+	MONSTER_FINGERMAGE1,
+	MONSTER_FINGERMAGE2,
+	MONSTER_FINGERMAGE3,
+	MONSTER_REGURGITATOR1,
+	MONSTER_REGURGITATOR2,
+	MONSTER_REGURGITATOR3,
+	MONSTER_DOOMKNIGHT1,
+	MONSTER_DOOMKNIGHT2,
+	MONSTER_DOOMKNIGHT3,
+	MONSTER_QUILLBEAR1,
+	MONSTER_QUILLBEAR2,
+	MONSTER_QUILLBEAR3,
+	MONSTER_QUILLBEAR4,
+	MONSTER_QUILLBEAR5,
+	MONSTER_SNAKE,
+	MONSTER_PARROT,
+	MONSTER_FISH,
+	MONSTER_EVILHOLE1,
+	MONSTER_EVILHOLE2,
+	MONSTER_EVILHOLE3,
+	MONSTER_EVILHOLE4,
+	MONSTER_EVILHOLE5,
+	MONSTER_TRAP_FIREBOLT,
+	MONSTER_TRAP_HORZMISSILE,
+	MONSTER_TRAP_VERTMISSILE,
+	MONSTER_TRAP_POISONCLOUD,
+	MONSTER_TRAP_LIGHTNING,
+	MONSTER_ACT2GUARD2,
+	MONSTER_INVISOSPAWNER,
+	MONSTER_DIABLOCLONE,
+	MONSTER_SUCKERNEST1,
+	MONSTER_SUCKERNEST2,
+	MONSTER_SUCKERNEST3,
+	MONSTER_SUCKERNEST4,
+	MONSTER_ACT2HIRE,
+	MONSTER_MINISPIDER,
+	MONSTER_BONEPRISON1,
+	MONSTER_BONEPRISON2,
+	MONSTER_BONEPRISON3,
+	MONSTER_BONEPRISON4,
+	MONSTER_BONEWALL,
+	MONSTER_COUNCILMEMBER1,
+	MONSTER_COUNCILMEMBER2,
+	MONSTER_COUNCILMEMBER3,
+	MONSTER_TURRET1,
+	MONSTER_TURRET2,
+	MONSTER_TURRET3,
+	MONSTER_HYDRA1,
+	MONSTER_HYDRA2,
+	MONSTER_HYDRA3,
+	MONSTER_TRAP_MELEE,
+	MONSTER_SEVENTOMBS,
+	MONSTER_DOPPLEZON,
+	MONSTER_VALKYRIE,
+	MONSTER_ACT2GUARD3,
+	MONSTER_ACT3HIRE,
+	MONSTER_MEGADEMON1,
+	MONSTER_MEGADEMON2,
+	MONSTER_MEGADEMON3,
+	MONSTER_NECROSKELETON,
+	MONSTER_NECROMAGE,
+	MONSTER_GRISWOLD,
+	MONSTER_COMPELLINGORB,
+	MONSTER_TYRAEL2,
+	MONSTER_DARKWANDERER,
+	MONSTER_TRAP_NOVA,
+	MONSTER_SPIRITMUMMY,
+	MONSTER_LIGHTNINGSPIRE,
+	MONSTER_FIRETOWER,
+	MONSTER_SLINGER1,
+	MONSTER_SLINGER2,
+	MONSTER_SLINGER3,
+	MONSTER_SLINGER4,
+	MONSTER_ACT2GUARD4,
+	MONSTER_ACT2GUARD5,
+	MONSTER_SKMAGE_COLD1,
+	MONSTER_SKMAGE_COLD2,
+	MONSTER_SKMAGE_COLD3,
+	MONSTER_SKMAGE_COLD4,
+	MONSTER_SKMAGE_FIRE1,
+	MONSTER_SKMAGE_FIRE2,
+	MONSTER_SKMAGE_FIRE3,
+	MONSTER_SKMAGE_FIRE4,
+	MONSTER_SKMAGE_LTNG1,
+	MONSTER_SKMAGE_LTNG2,
+	MONSTER_SKMAGE_LTNG3,
+	MONSTER_SKMAGE_LTNG4,
+	MONSTER_HELLBOVINE,
+	MONSTER_WINDOW1,
+	MONSTER_WINDOW2,
+	MONSTER_SLINGER5,
+	MONSTER_SLINGER6,
+	MONSTER_FETISHBLOW1,
+	MONSTER_FETISHBLOW2,
+	MONSTER_FETISHBLOW3,
+	MONSTER_FETISHBLOW4,
+	MONSTER_FETISHBLOW5,
+	MONSTER_MEPHISTOSPIRIT,
+	MONSTER_SMITH,
+	MONSTER_TRAPPEDSOUL1,
+	MONSTER_TRAPPEDSOUL2,
+	MONSTER_JAMELLA,
+	MONSTER_IZUALGHOST,
+	MONSTER_FETISH11,
+	MONSTER_MALACHAI,
+	MONSTER_HEPHASTO,
+	MONSTER_WAKEOFDESTRUCTION,
+	MONSTER_CHARGEBOLTSENTRY,
+	MONSTER_LIGHTNINGSENTRY,
+	MONSTER_BLADECREEPER,
+	MONSTER_INVISOPET,
+	MONSTER_INFERNOSENTRY,
+	MONSTER_DEATHSENTRY,
+	MONSTER_SHADOWWARRIOR,
+	MONSTER_SHADOWMASTER,
+	MONSTER_DRUIDHAWK,
+	MONSTER_SPIRITWOLF,
+	MONSTER_FENRIS,
+	MONSTER_SPIRITOFBARBS,
+	MONSTER_HEARTOFWOLVERINE,
+	MONSTER_OAKSAGE,
+	MONSTER_PLAGUEPOPPY,
+	MONSTER_CYCLEOFLIFE,
+	MONSTER_VINECREATURE,
+	MONSTER_DRUIDBEAR,
+	MONSTER_EAGLE,
+	MONSTER_WOLF,
+	MONSTER_BEAR,
+	MONSTER_BARRICADEDOOR1,
+	MONSTER_BARRICADEDOOR2,
+	MONSTER_PRISONDOOR,
+	MONSTER_BARRICADETOWER,
+	MONSTER_REANIMATEDHORDE1,
+	MONSTER_REANIMATEDHORDE2,
+	MONSTER_REANIMATEDHORDE3,
+	MONSTER_REANIMATEDHORDE4,
+	MONSTER_REANIMATEDHORDE5,
+	MONSTER_SIEGEBEAST1,
+	MONSTER_SIEGEBEAST2,
+	MONSTER_SIEGEBEAST3,
+	MONSTER_SIEGEBEAST4,
+	MONSTER_SIEGEBEAST5,
+	MONSTER_SNOWYETI1,
+	MONSTER_SNOWYETI2,
+	MONSTER_SNOWYETI3,
+	MONSTER_SNOWYETI4,
+	MONSTER_WOLFRIDER1,
+	MONSTER_WOLFRIDER2,
+	MONSTER_WOLFRIDER3,
+	MONSTER_MINION1,
+	MONSTER_MINION2,
+	MONSTER_MINION3,
+	MONSTER_MINION4,
+	MONSTER_MINION5,
+	MONSTER_MINION6,
+	MONSTER_MINION7,
+	MONSTER_MINION8,
+	MONSTER_SUICIDEMINION1,
+	MONSTER_SUICIDEMINION2,
+	MONSTER_SUICIDEMINION3,
+	MONSTER_SUICIDEMINION4,
+	MONSTER_SUICIDEMINION5,
+	MONSTER_SUICIDEMINION6,
+	MONSTER_SUICIDEMINION7,
+	MONSTER_SUICIDEMINION8,
+	MONSTER_SUCCUBUS1,
+	MONSTER_SUCCUBUS2,
+	MONSTER_SUCCUBUS3,
+	MONSTER_SUCCUBUS4,
+	MONSTER_SUCCUBUS5,
+	MONSTER_SUCCUBUSWITCH1,
+	MONSTER_SUCCUBUSWITCH2,
+	MONSTER_SUCCUBUSWITCH3,
+	MONSTER_SUCCUBUSWITCH4,
+	MONSTER_SUCCUBUSWITCH5,
+	MONSTER_OVERSEER1,
+	MONSTER_OVERSEER2,
+	MONSTER_OVERSEER3,
+	MONSTER_OVERSEER4,
+	MONSTER_OVERSEER5,
+	MONSTER_MINIONSPAWNER1,
+	MONSTER_MINIONSPAWNER2,
+	MONSTER_MINIONSPAWNER3,
+	MONSTER_MINIONSPAWNER4,
+	MONSTER_MINIONSPAWNER5,
+	MONSTER_MINIONSPAWNER6,
+	MONSTER_MINIONSPAWNER7,
+	MONSTER_MINIONSPAWNER8,
+	MONSTER_IMP1,
+	MONSTER_IMP2,
+	MONSTER_IMP3,
+	MONSTER_IMP4,
+	MONSTER_IMP5,
+	MONSTER_CATAPULT1,
+	MONSTER_CATAPULT2,
+	MONSTER_CATAPULT3,
+	MONSTER_CATAPULT4,
+	MONSTER_FROZENHORROR1,
+	MONSTER_FROZENHORROR2,
+	MONSTER_FROZENHORROR3,
+	MONSTER_FROZENHORROR4,
+	MONSTER_FROZENHORROR5,
+	MONSTER_BLOODLORD1,
+	MONSTER_BLOODLORD2,
+	MONSTER_BLOODLORD3,
+	MONSTER_BLOODLORD4,
+	MONSTER_BLOODLORD5,
+	MONSTER_LARZUK,
+	MONSTER_DREHYA,
+	MONSTER_MALAH,
+	MONSTER_NIHLATHAK,
+	MONSTER_QUAL_KEHK,
+	MONSTER_CATAPULTSPOTTER1,
+	MONSTER_CATAPULTSPOTTER2,
+	MONSTER_CATAPULTSPOTTER3,
+	MONSTER_CATAPULTSPOTTER4,
+	MONSTER_CAIN6,
+	MONSTER_TYRAEL3,
+	MONSTER_ACT5BARB1,
+	MONSTER_ACT5BARB2,
+	MONSTER_BARRICADEWALL1,
+	MONSTER_BARRICADEWALL2,
+	MONSTER_NIHLATHAKBOSS,
+	MONSTER_DREHYAICED,
+	MONSTER_EVILHUT,
+	MONSTER_DEATHMAULER1,
+	MONSTER_DEATHMAULER2,
+	MONSTER_DEATHMAULER3,
+	MONSTER_DEATHMAULER4,
+	MONSTER_DEATHMAULER5,
+	MONSTER_ACT5POW,
+	MONSTER_ACT5BARB3,
+	MONSTER_ACT5BARB4,
+	MONSTER_ANCIENTSTATUE1,
+	MONSTER_ANCIENTSTATUE2,
+	MONSTER_ANCIENTSTATUE3,
+	MONSTER_ANCIENTBARB1,
+	MONSTER_ANCIENTBARB2,
+	MONSTER_ANCIENTBARB3,
+	MONSTER_BAALTHRONE,
+	MONSTER_BAALCRAB,
+	MONSTER_BAALTAUNT,
+	MONSTER_PUTRIDDEFILER1,
+	MONSTER_PUTRIDDEFILER2,
+	MONSTER_PUTRIDDEFILER3,
+	MONSTER_PUTRIDDEFILER4,
+	MONSTER_PUTRIDDEFILER5,
+	MONSTER_PAINWORM1,
+	MONSTER_PAINWORM2,
+	MONSTER_PAINWORM3,
+	MONSTER_PAINWORM4,
+	MONSTER_PAINWORM5,
+	MONSTER_BUNNY,
+	MONSTER_BAALHIGHPRIEST,
+	MONSTER_VENOMLORD,
+	MONSTER_BAALCRABSTAIRS,
+	MONSTER_ACT5HIRE1,
+	MONSTER_ACT5HIRE2,
+	MONSTER_BAALTENTACLE1,
+	MONSTER_BAALTENTACLE2,
+	MONSTER_BAALTENTACLE3,
+	MONSTER_BAALTENTACLE4,
+	MONSTER_BAALTENTACLE5,
+	MONSTER_INJUREDBARB1,
+	MONSTER_INJUREDBARB2,
+	MONSTER_INJUREDBARB3,
+	MONSTER_BAALCLONE,
+	MONSTER_BAALMINION1,
+	MONSTER_BAALMINION2,
+	MONSTER_BAALMINION3,
+	MONSTER_WORLDSTONEEFFECT,
+	MONSTER_SK_ARCHER6,
+	MONSTER_SK_ARCHER7,
+	MONSTER_SK_ARCHER8,
+	MONSTER_SK_ARCHER9,
+	MONSTER_SK_ARCHER10,
+	MONSTER_BIGHEAD6,
+	MONSTER_BIGHEAD7,
+	MONSTER_BIGHEAD8,
+	MONSTER_BIGHEAD9,
+	MONSTER_BIGHEAD10,
+	MONSTER_GOATMAN6,
+	MONSTER_GOATMAN7,
+	MONSTER_GOATMAN8,
+	MONSTER_GOATMAN9,
+	MONSTER_GOATMAN10,
+	MONSTER_FOULCROW5,
+	MONSTER_FOULCROW6,
+	MONSTER_FOULCROW7,
+	MONSTER_FOULCROW8,
+	MONSTER_CLAWVIPER6,
+	MONSTER_CLAWVIPER7,
+	MONSTER_CLAWVIPER8,
+	MONSTER_CLAWVIPER9,
+	MONSTER_CLAWVIPER10,
+	MONSTER_SANDRAIDER6,
+	MONSTER_SANDRAIDER7,
+	MONSTER_SANDRAIDER8,
+	MONSTER_SANDRAIDER9,
+	MONSTER_SANDRAIDER10,
+	MONSTER_DEATHMAULER6,
+	MONSTER_QUILLRAT6,
+	MONSTER_QUILLRAT7,
+	MONSTER_QUILLRAT8,
+	MONSTER_VULTURE5,
+	MONSTER_THORNHULK5,
+	MONSTER_SLINGER7,
+	MONSTER_SLINGER8,
+	MONSTER_SLINGER9,
+	MONSTER_CR_ARCHER6,
+	MONSTER_CR_ARCHER7,
+	MONSTER_CR_LANCER6,
+	MONSTER_CR_LANCER7,
+	MONSTER_CR_LANCER8,
+	MONSTER_BLUNDERBORE5,
+	MONSTER_BLUNDERBORE6,
+	MONSTER_SKMAGE_FIRE5,
+	MONSTER_SKMAGE_FIRE6,
+	MONSTER_SKMAGE_LTNG5,
+	MONSTER_SKMAGE_LTNG6,
+	MONSTER_SKMAGE_COLD5,
+	MONSTER_SKMAGE_POIS5,
+	MONSTER_SKMAGE_POIS6,
+	MONSTER_PANTHERWOMAN5,
+	MONSTER_PANTHERWOMAN6,
+	MONSTER_SANDLEAPER6,
+	MONSTER_SANDLEAPER7,
+	MONSTER_WRAITH6,
+	MONSTER_WRAITH7,
+	MONSTER_WRAITH8,
+	MONSTER_SUCCUBUS6,
+	MONSTER_SUCCUBUS7,
+	MONSTER_SUCCUBUSWITCH6,
+	MONSTER_SUCCUBUSWITCH7,
+	MONSTER_SUCCUBUSWITCH8,
+	MONSTER_WILLOWISP5,
+	MONSTER_WILLOWISP6,
+	MONSTER_WILLOWISP7,
+	MONSTER_FALLEN6,
+	MONSTER_FALLEN7,
+	MONSTER_FALLEN8,
+	MONSTER_FALLENSHAMAN6,
+	MONSTER_FALLENSHAMAN7,
+	MONSTER_FALLENSHAMAN8,
+	MONSTER_SKELETON6,
+	MONSTER_SKELETON7,
+	MONSTER_BATDEMON6,
+	MONSTER_BATDEMON7,
+	MONSTER_BLOODLORD6,
+	MONSTER_BLOODLORD7,
+	MONSTER_SCARAB6,
+	MONSTER_SCARAB7,
+	MONSTER_FETISH6,
+	MONSTER_FETISH7,
+	MONSTER_FETISH8,
+	MONSTER_FETISHBLOW6,
+	MONSTER_FETISHBLOW7,
+	MONSTER_FETISHBLOW8,
+	MONSTER_FETISHSHAMAN6,
+	MONSTER_FETISHSHAMAN7,
+	MONSTER_FETISHSHAMAN8,
+	MONSTER_BABOON7,
+	MONSTER_BABOON8,
+	MONSTER_UNRAVELER6,
+	MONSTER_UNRAVELER7,
+	MONSTER_UNRAVELER8,
+	MONSTER_UNRAVELER9,
+	MONSTER_ZEALOT4,
+	MONSTER_ZEALOT5,
+	MONSTER_CANTOR5,
+	MONSTER_CANTOR6,
+	MONSTER_VILEMOTHER4,
+	MONSTER_VILEMOTHER5,
+	MONSTER_VILECHILD4,
+	MONSTER_VILECHILD5,
+	MONSTER_SANDMAGGOT6,
+	MONSTER_MAGGOTBABY6,
+	MONSTER_MAGGOTEGG6,
+	MONSTER_MINION9,
+	MONSTER_MINION10,
+	MONSTER_MINION11,
+	MONSTER_ARACH6,
+	MONSTER_MEGADEMON4,
+	MONSTER_MEGADEMON5,
+	MONSTER_IMP6,
+	MONSTER_IMP7,
+	MONSTER_BONEFETISH6,
+	MONSTER_BONEFETISH7,
+	MONSTER_FINGERMAGE4,
+	MONSTER_FINGERMAGE5,
+	MONSTER_REGURGITATOR4,
+	MONSTER_VAMPIRE6,
+	MONSTER_VAMPIRE7,
+	MONSTER_VAMPIRE8,
+	MONSTER_REANIMATEDHORDE6,
+	MONSTER_DKFIG1,
+	MONSTER_DKFIG2,
+	MONSTER_DKMAG1,
+	MONSTER_DKMAG2,
+	MONSTER_MUMMY6,
+	MONSTER_UBERMEPHISTO,
+	MONSTER_UBERDIABLO,
+	MONSTER_UBERIZUAL,
+	MONSTER_UBERANDARIEL,
+	MONSTER_UBERDURIEL,
+	MONSTER_UBERBAAL,
+	MONSTER_DEMONSPAWNER,
+	MONSTER_DEMONHOLE,
+	MONSTER_MEGADEMON6,
+	MONSTER_DKMAG3,
+	MONSTER_IMP8,
+	MONSTER_SWARM5,
+	MONSTER_SANDMAGGOT7,
+	MONSTER_ARACH7,
+	MONSTER_SCARAB8,
+	MONSTER_SUCCUBUS8,
+	MONSTER_SUCCUBUSWITCH9,
+	MONSTER_CORRUPTROGUE6,
+	MONSTER_CR_ARCHER8,
+	MONSTER_CR_LANCER9,
+	MONSTER_OVERSEER6,
+	MONSTER_SKELETON8,
+	MONSTER_SK_ARCHER11,
+	MONSTER_SKMAGE_FIRE7,
+	MONSTER_SKMAGE_LTNG7,
+	MONSTER_SKMAGE_COLD6,
+	MONSTER_SKMAGE_POIS7,
+	MONSTER_VAMPIRE9,
+	MONSTER_WRAITH9,
+	MONSTER_WILLOWISP8,
+};
+
+enum D2C_SuperUniques
+{
+	SUPERUNIQUE_BISHIBOSH,
+	SUPERUNIQUE_BONEBREAK,
+	SUPERUNIQUE_COLDCROW,
+	SUPERUNIQUE_RAKANISHU,
+	SUPERUNIQUE_TREEHEAD_WOODFIST,
+	SUPERUNIQUE_GRISWOLD,
+	SUPERUNIQUE_THE_COUNTESS,
+	SUPERUNIQUE_PITSPAWN_FOULDOG,
+	SUPERUNIQUE_FLAMESPIKE_THE_CRAWLER,
+	SUPERUNIQUE_BONEASH,
+	SUPERUNIQUE_RADAMENT,
+	SUPERUNIQUE_BLOODWITCH_THE_WILD,
+	SUPERUNIQUE_FANGSKIN,
+	SUPERUNIQUE_BEETLEBURST,
+	SUPERUNIQUE_LEATHERARM,
+	SUPERUNIQUE_COLDWORM_THE_BURROWER,
+	SUPERUNIQUE_FIRE_EYE,
+	SUPERUNIQUE_DARK_ELDER,
+	SUPERUNIQUE_THE_SUMMONER,
+	SUPERUNIQUE_ANCIENT_KAA_THE_SOULLESS,
+	SUPERUNIQUE_THE_SMITH,
+	SUPERUNIQUE_WEB_MAGE_THE_BURNING,
+	SUPERUNIQUE_WITCH_DOCTOR_ENDUGU,
+	SUPERUNIQUE_STORMTREE,
+	SUPERUNIQUE_SARINA_THE_BATTLEMAID,
+	SUPERUNIQUE_ICEHAWK_RIFTWING,
+	SUPERUNIQUE_ISMAIL_VILEHAND,
+	SUPERUNIQUE_GELEB_FLAMEFINGER,
+	SUPERUNIQUE_BREMM_SPARKFIST,
+	SUPERUNIQUE_TOORC_ICEFIST,
+	SUPERUNIQUE_WYAND_VOIDFINGER,
+	SUPERUNIQUE_MAFFER_DRAGONHAND,
+	SUPERUNIQUE_WINGED_DEATH,
+	SUPERUNIQUE_THE_TORMENTOR,
+	SUPERUNIQUE_TAINTBREEDER,
+	SUPERUNIQUE_RIFTWRAITH_THE_CANNIBAL,
+	SUPERUNIQUE_INFECTOR_OF_SOULS,
+	SUPERUNIQUE_LORD_DE_SEIS,
+	SUPERUNIQUE_GRAND_VIZIER_OF_CHAOS,
+	SUPERUNIQUE_THE_COW_KING,
+	SUPERUNIQUE_CORPSEFIRE,
+	SUPERUNIQUE_THE_FEATURE_CREEP,
+	SUPERUNIQUE_SIEGE_BOSS,
+	SUPERUNIQUE_ANCIENT_BARBARIAN_1,
+	SUPERUNIQUE_ANCIENT_BARBARIAN_2,
+	SUPERUNIQUE_ANCIENT_BARBARIAN_3,
+	SUPERUNIQUE_AXE_DWELLER,
+	SUPERUNIQUE_BONESAW_BREAKER,
+	SUPERUNIQUE_DAC_FARREN,
+	SUPERUNIQUE_MEGAFLOW_RECTIFIER,
+	SUPERUNIQUE_EYEBACK_UNLEASHED,
+	SUPERUNIQUE_THREASH_SOCKET,
+	SUPERUNIQUE_PINDLESKIN,
+	SUPERUNIQUE_SNAPCHIP_SHATTER,
+	SUPERUNIQUE_ANODIZED_ELITE,
+	SUPERUNIQUE_VINVEAR_MOLECH,
+	SUPERUNIQUE_SHARP_TOOTH_SAYER,
+	SUPERUNIQUE_MAGMA_TORQUER,
+	SUPERUNIQUE_BLAZE_RIPPER,
+	SUPERUNIQUE_FROZENSTEIN,
+	SUPERUNIQUE_NIHLATHAK_BOSS,
+	SUPERUNIQUE_BAAL_SUBJECT_1,
+	SUPERUNIQUE_BAAL_SUBJECT_2,
+	SUPERUNIQUE_BAAL_SUBJECT_3,
+	SUPERUNIQUE_BAAL_SUBJECT_4,
+	SUPERUNIQUE_BAAL_SUBJECT_5,
+};
+
