+++ source/D2Common/src/DataTbls/DataTbls.cpp
@@ -431,7 +431,7 @@ void __fastcall DATATBLS_LoadStatesTxt(void* pMemPool)
 		{ "clteventfunc", TXTFIELD_WORD, 0, 50, NULL },
 		{ "cltactivefunc", TXTFIELD_WORD, 0, 52, NULL },
 		{ "srvactivefunc", TXTFIELD_WORD, 0, 54, NULL },
-		{ "end", 0, 0, 0, NULL },
+		{ "end", TXTFIELD_NONE, 0, 0, NULL },
 	};
 
 	sgptDataTables->pStatesLinker = (D2TxtLinkStrc*)FOG_AllocLinker(__FILE__, __LINE__);
@@ -579,7 +579,7 @@ void __fastcall DATATBLS_LoadPetTypeTxt(void* pMemPool)
 		{ "mclass4", TXTFIELD_WORD, 0, 184, NULL },
 		{ "micon4", TXTFIELD_ASCII, 32, 143, NULL },
 		{ "name", TXTFIELD_KEYTOWORD, 0, 12, DATATBLS_GetStringIdFromReferenceString },
-		{ "end", 0, 0, 0, NULL },
+		{ "end", TXTFIELD_NONE, 0, 0, NULL },
 	};
 
 	sgptDataTables->pPetTypeLinker = (D2TxtLinkStrc*)FOG_AllocLinker(__FILE__, __LINE__);
@@ -927,7 +927,7 @@ void __stdcall DATATBLS_LoadAllTxts(void* pMemPool, int a2, int a3)
 		{ "Druid", TXTFIELD_DWORD, 0, 20, NULL },
 		{ "Assassin", TXTFIELD_DWORD, 0, 24, NULL },
 		{ "ExpRatio", TXTFIELD_DWORD, 0, 28, NULL },
-		{ "end", 0, 0, 0, NULL },
+		{ "end", TXTFIELD_NONE, 0, 0, NULL },
 	};
 
 	DATATBLS_LoadSomeTxts(pMemPool);
@@ -999,82 +999,82 @@ void __fastcall DATATBLS_LoadSomeTxts(void* pMemPool)
 	D2BinFieldStrc pHireDescTbl[] =
 	{
 		{ "code", TXTFIELD_ASCIITOCODE, 0, 0, &sgptDataTables->pHireDescLinker },
-		{ "end", 0, 0, 0, NULL },
+		{ "end", TXTFIELD_NONE, 0, 0, NULL },
 	};
 	D2BinFieldStrc pMonModeTbl[] =
 	{
 		{ "code", TXTFIELD_ASCIITOCODE, 0, 0, &sgptDataTables->pMonModeLinker },
-		{ "end", 0, 0, 0, NULL },
+		{ "end", TXTFIELD_NONE, 0, 0, NULL },
 	};
 	D2BinFieldStrc pPlayerClassTbl[] =
 	{
 		{ "code", TXTFIELD_ASCIITOCODE, 0, 0, &sgptDataTables->pPlayerClassLinker },
-		{ "end", 0, 0, 0, NULL },
+		{ "end", TXTFIELD_NONE, 0, 0, NULL },
 	};
 	D2BinFieldStrc pPlrModeTbl[] =
 	{
 		{ "code", TXTFIELD_ASCIITOCODE, 0, 0, &sgptDataTables->pPlrModeLinker },
-		{ "end", 0, 0, 0, NULL },
+		{ "end", TXTFIELD_NONE, 0, 0, NULL },
 	};
 	D2BinFieldStrc pStorePageTbl[] =
 	{
 		{ "code", TXTFIELD_ASCIITOCODE, 0, 0, &sgptDataTables->pStorePageLinker },
-		{ "end", 0, 0, 0, NULL },
+		{ "end", TXTFIELD_NONE, 0, 0, NULL },
 	};
 	D2BinFieldStrc pMonAiTbl[] =
 	{
 		{ "AI", TXTFIELD_NAMETOINDEX, 0, 0, &sgptDataTables->pMonAiLinker },
-		{ "end", 0, 0, 0, NULL },
+		{ "end", TXTFIELD_NONE, 0, 0, NULL },
 	};
 	D2BinFieldStrc pHitClassTbl[] =
 	{
 		{ "code", TXTFIELD_ASCIITOCODE, 0, 0, &sgptDataTables->pHitClassLinker },
-		{ "end", 0, 0, 0, NULL },
+		{ "end", TXTFIELD_NONE, 0, 0, NULL },
 	};
 	D2BinFieldStrc pMonPlaceTbl[] =
 	{
 		{ "code", TXTFIELD_NAMETOINDEX, 0, 0, &sgptDataTables->pMonPlaceLinker },
-		{ "end", 0, 0, 0, NULL },
+		{ "end", TXTFIELD_NONE, 0, 0, NULL },
 	};
 	D2BinFieldStrc pCompCodeTbl[] =
 	{
 		{ "code", TXTFIELD_ASCIITOCODE, 0, 0, &sgptDataTables->pCompCodeLinker },
-		{ "end", 0, 0, 0, NULL },
+		{ "end", TXTFIELD_NONE, 0, 0, NULL },
 	};
 	D2BinFieldStrc pSkillCalcTbl[] =
 	{
 		{ "code", TXTFIELD_ASCIITOCODE, 0, 0, &sgptDataTables->pSkillCalcLinker },
-		{ "end", 0, 0, 0, NULL },
+		{ "end", TXTFIELD_NONE, 0, 0, NULL },
 	};
 	D2BinFieldStrc pElemTypesTbl[] =
 	{
 		{ "code", TXTFIELD_ASCIITOCODE, 0, 0, &sgptDataTables->pElemTypesLinker },
-		{ "end", 0, 0, 0, NULL },
+		{ "end", TXTFIELD_NONE, 0, 0, NULL },
 	};
 	D2BinFieldStrc pMissCalcTbl[] =
 	{
 		{ "code", TXTFIELD_ASCIITOCODE, 0, 0, &sgptDataTables->pMissileCalcLinker },
-		{ "end", 0, 0, 0, NULL },
+		{ "end", TXTFIELD_NONE, 0, 0, NULL },
 	};
 	D2BinFieldStrc pBodyLocsTbl[] =
 	{
 		{ "code", TXTFIELD_ASCIITOCODE, 0, 0, &sgptDataTables->pBodyLocsLinker },
-		{ "end", 0, 0, 0, NULL },
+		{ "end", TXTFIELD_NONE, 0, 0, NULL },
 	};
 	D2BinFieldStrc pSkillCodeTbl[] =
 	{
 		{ "skill", TXTFIELD_NAMETOINDEX, 0, 0, &sgptDataTables->iSkillCode },
-		{ "end", 0, 0, 0, NULL },
+		{ "end", TXTFIELD_NONE, 0, 0, NULL },
 	};
 	D2BinFieldStrc pColorsTbl[] =
 	{
 		{ "code", TXTFIELD_ASCIITOCODE, 0, 0, &sgptDataTables->pColorsLinker },
-		{ "end", 0, 0, 0, NULL },
+		{ "end", TXTFIELD_NONE, 0, 0, NULL },
 	};
 	D2BinFieldStrc pEventsTbl[] =
 	{
 		{ "event", TXTFIELD_NAMETOINDEX, 0, 0, &sgptDataTables->pEventsLinker },
-		{ "end", 0, 0, 0, NULL },
+		{ "end", TXTFIELD_NONE, 0, 0, NULL },
 	};
 
 	sgptDataTables->pCompCodeLinker = (D2TxtLinkStrc*)FOG_AllocLinker(__FILE__, __LINE__);
@@ -1205,7 +1205,7 @@ void __fastcall DATATBLS_LoadCharStatsTxt(void* pMemPool)
 		{ "item9count", TXTFIELD_BYTE, 0, 161, NULL },
 		{ "item10", TXTFIELD_RAW, 0, 164, NULL },
 		{ "item10count", TXTFIELD_BYTE, 0, 169, NULL },
-		{ "end", 0, 0, 0, NULL },
+		{ "end", TXTFIELD_NONE, 0, 0, NULL },
 	};
 
 	sgptDataTables->pCharStatsTxt = (D2CharStatsTxt*)DATATBLS_CompileTxt(pMemPool, "charstats", pTbl, &sgptDataTables->nCharStatsTxtRecordCount, sizeof(D2CharStatsTxt));
@@ -1240,7 +1240,7 @@ void __fastcall DATATBLS_LoadDifficultyLevelsTxt(void* pMemPool)
 		{ "GambleUnique", TXTFIELD_DWORD, 0, 76, NULL },
 		{ "GambleUber", TXTFIELD_DWORD, 0, 80, NULL },
 		{ "GambleUltra", TXTFIELD_DWORD, 0, 84, NULL },
-		{ "end", 0, 0, 0, NULL },
+		{ "end", TXTFIELD_NONE, 0, 0, NULL },
 	};
 
 	sgptDataTables->pDifficultyLevelsTxt = (D2DifficultyLevelsTxt*)DATATBLS_CompileTxt(pMemPool, "difficultylevels", pTbl, &sgptDataTables->nDifficultyLevelsTxtRecordCount, sizeof(D2DifficultyLevelsTxt));
