+++ source/D2Common/src/Drlg/DrlgDrlg.cpp
@@ -13,6 +13,7 @@
 #include "D2Dungeon.h"
 #include "D2Seed.h"
 #include <D2CMP.h>
+#include <DataTbls/LevelsIds.h>
 
 
 //D2Common.0x6FD74120 (#10014)
