+++ source/D2Common/src/Monsters/Monsters.cpp
@@ -11,6 +11,7 @@
 #include "Units/Units.h"
 #include <D2Lang.h>
 #include <D2BitManip.h>
+#include <DataTbls/MonsterIds.h>
 
 
 //TODO: Check names
