+++ source/D2CommonDefinitions/include/D2Constants.h
@@ -109,1576 +109,9 @@ enum D2C_InventoryRecords
 	INVENTORYRECORD_ASSASSIN2
 };
 
-#ifndef CONSTANTS_ITEMS //Pseudo-Macro to group all constants
-enum D2C_ItemQualities
-{
-	ITEMQUAL_INFERIOR = 0x01, 	//0x01 Inferior
-	ITEMQUAL_NORMAL = 0x02, 	//0x02 Normal
-	ITEMQUAL_SUPERIOR = 0x03, 	//0x03 Superior
-	ITEMQUAL_MAGIC = 0x04, 		//0x04 Magic
-	ITEMQUAL_SET = 0x05, 		//0x05 Set
-	ITEMQUAL_RARE = 0x06, 		//0x06 Rare
-	ITEMQUAL_UNIQUE = 0x07, 	//0x07 Unique
-	ITEMQUAL_CRAFT = 0x08, 		//0x08 Crafted
-	ITEMQUAL_TEMPERED = 0x09	//0x09 Tempered
-};
-
-enum D2C_ItemTypes
-{
-	ITEMTYPE_NONE_1 = 0,
-	ITEMTYPE_NONE_2,
-	ITEMTYPE_SHIELD,
-	ITEMTYPE_ARMOR,
-	ITEMTYPE_GOLD,
-	ITEMTYPE_BOW_QUIVER,
-	ITEMTYPE_CROSSBOW_QUIVER,
-	ITEMTYPE_PLAYER_BODY_PART,
-	ITEMTYPE_HERB,
-	ITEMTYPE_POTION,
-	ITEMTYPE_RING,
-	ITEMTYPE_ELIXIR,
-	ITEMTYPE_AMULET,
-	ITEMTYPE_CHARM,
-	ITEMTYPE_NONE_3,
-	ITEMTYPE_BOOTS,
-	ITEMTYPE_GLOVES,
-	ITEMTYPE_NONE_4,
-	ITEMTYPE_BOOK,
-	ITEMTYPE_BELT,
-	ITEMTYPE_GEM,
-	ITEMTYPE_TORCH,
-	ITEMTYPE_SCROLL,
-	ITEMTYPE_NONE_5,
-	ITEMTYPE_SCEPTER,
-	ITEMTYPE_WAND,
-	ITEMTYPE_STAFF,
-	ITEMTYPE_BOW,
-	ITEMTYPE_AXE,
-	ITEMTYPE_CLUB,
-	ITEMTYPE_SWORD,
-	ITEMTYPE_HAMMER,
-	ITEMTYPE_KNIFE,
-	ITEMTYPE_SPEAR,
-	ITEMTYPE_POLEARM,
-	ITEMTYPE_CROSSBOW,
-	ITEMTYPE_MACE,
-	ITEMTYPE_HELM,
-	ITEMTYPE_MISSILE_POTION,
-	ITEMTYPE_QUEST,
-	ITEMTYPE_BODY_PART,
-	ITEMTYPE_KEY,
-	ITEMTYPE_THROWING_KNIFE,
-	ITEMTYPE_THROWING_AXE,
-	ITEMTYPE_JAVELIN,
-	ITEMTYPE_WEAPON,
-	ITEMTYPE_MELEE_WEAPON,
-	ITEMTYPE_MISSILE_WEAPON,
-	ITEMTYPE_THROWN_WEAPON,
-	ITEMTYPE_COMBO_WEAPON,
-	ITEMTYPE_ANY_ARMOR,
-	ITEMTYPE_ANY_SHIELD,
-	ITEMTYPE_MISCELLANEOUS,
-	ITEMTYPE_SOCKET_FILLER,
-	ITEMTYPE_SECOND_HAND,
-	ITEMTYPE_STAVES_AND_RODS,
-	ITEMTYPE_MISSILE,
-	ITEMTYPE_BLUNT,
-	ITEMTYPE_JEWEL,
-	ITEMTYPE_CLASS_SPECIFIC,
-	ITEMTYPE_AMAZON_ITEM,
-	ITEMTYPE_BARBARIAN_ITEM,
-	ITEMTYPE_NECROMANCER_ITEM,
-	ITEMTYPE_PALADIN_ITEM,
-	ITEMTYPE_SORCERESS_ITEM,
-	ITEMTYPE_ASSASSIN_ITEM,
-	ITEMTYPE_DRUID_ITEM,
-	ITEMTYPE_HAND_TO_HAND,
-	ITEMTYPE_ORB,
-	ITEMTYPE_VOODOO_HEADS,
-	ITEMTYPE_AURIC_SHIELDS,
-	ITEMTYPE_PRIMAL_HELM,
-	ITEMTYPE_PELT,
-	ITEMTYPE_CLOAK,
-	ITEMTYPE_RUNE,
-	ITEMTYPE_CIRCLET,
-	ITEMTYPE_HEALING_POTION,
-	ITEMTYPE_MANA_POTION,
-	ITEMTYPE_REJUV_POTION,
-	ITEMTYPE_STAMINA_POTION,
-	ITEMTYPE_ANTIDOTE_POTION,
-	ITEMTYPE_THAWING_POTION,
-	ITEMTYPE_SMALL_CHARM,
-	ITEMTYPE_MEDIUM_CHARM,
-	ITEMTYPE_LARGE_CHARM,
-	ITEMTYPE_AMAZON_BOW,
-	ITEMTYPE_AMAZON_SPEAR,
-	ITEMTYPE_AMAZON_JAVELIN,
-	ITEMTYPE_HAND_TO_HAND_2,
-	ITEMTYPE_MAGIC_BOW_QUIV,
-	ITEMTYPE_UNK,
-	ITEMTYPE_CHIPPED_GEM,
-	ITEMTYPE_FLAWED_GEM,
-	ITEMTYPE_STANDARD_GEM,
-	ITEMTYPE_FLAWLESS_GEM,
-	ITEMTYPE_PERFECT_GEM,
-	ITEMTYPE_AMETHYST,
-	ITEMTYPE_DIAMOND,
-	ITEMTYPE_EMERALD,
-	ITEMTYPE_RUBY,
-	ITEMTYPE_SAPPHIRE,
-	ITEMTYPE_TOPAZ,
-	ITEMTYPE_SKULL,
-	// Own ItemTypes
-	ITEMTYPE_TOME,
-	ITEMTYPE_CLUESCROLL
-};
-
-enum D2C_ItemModes
-{
-	IMODE_STORED, 		//Item is in Storage (inventory, cube, Stash?)
-	IMODE_EQUIP, 		//Item is Equippped
-	IMODE_INBELT, 		//Item is in Belt Rows
-	IMODE_ONGROUND, 	//Item is on Ground
-	IMODE_ONCURSOR, 	//Item is on Cursor
-	IMODE_DROPPING, 	//Item is Being Dropped
-	IMODE_SOCKETED		//Item is Socketed in another Item
-};
-
-enum D2C_ItemFlags
-{
-	IFLAG_NEWITEM = 0x00000001,
-	IFLAG_TARGET = 0x00000002,
-	IFLAG_TARGETING = 0x00000004,
-	IFLAG_DELETED = 0x00000008,
-	IFLAG_IDENTIFIED = 0x00000010,
-	IFLAG_QUANTITY = 0x00000020,
-	IFLAG_SWITCHIN = 0x00000040,
-	IFLAG_SWITCHOUT = 0x00000080,
-	IFLAG_BROKEN = 0x00000100,
-	IFLAG_REPAIRED = 0x00000200,
-	IFLAG_UNK1 = 0x00000400,
-	IFLAG_SOCKETED = 0x00000800,
-	IFLAG_NOSELL = 0x00001000,
-	IFLAG_INSTORE = 0x00002000,
-	IFLAG_NOEQUIP = 0x00004000,
-	IFLAG_NAMED = 0x00008000,
-	IFLAG_ISEAR = 0x00010000,
-	IFLAG_STARTITEM = 0x00020000,
-	IFLAG_UNK2 = 0x00040000,
-	IFLAG_INIT = 0x00080000,
-	IFLAG_UNK3 = 0x00100000,
-	IFLAG_COMPACTSAVE = 0x00200000,
-	IFLAG_ETHEREAL = 0x00400000,
-	IFLAG_JUSTSAVED = 0x00800000,
-	IFLAG_PERSONALIZED = 0x01000000,
-	IFLAG_LOWQUALITY = 0x02000000,
-	IFLAG_RUNEWORD = 0x04000000,
-	IFLAG_ITEM = 0x08000000
-};
-
-enum D2C_UniqueItemFlags
-{
-	UNIQUEITEM_ENABLED = 0x00000001,
-	UNIQUEITEM_NOLIMIT = 0x00000002,
-	UNIQUEITEM_CARRY1 = 0x00000004,
-	UNIQUEITEM_LADDER = 0x00000008,
-};
-
-enum D2C_ItemDropFlags
-{
-	ITEMDROPFLAG_ALWAYSETH = 0x00000004,
-};
-
-enum D2C_ItemCmdFlags
-{
-	ITEMCMDFLAG_DELETE = 0x00000020,
-};
-
-#define D2C_InventoryHeader 0x1020304
-
-enum D2C_ItemInvPage
-{
-	INVPAGE_INVENTORY = 0,
-	INVPAGE_EQUIP = 1,
-	INVPAGE_TRADE = 2,
-	INVPAGE_CUBE = 3,
-	INVPAGE_STASH = 4,
-	INVPAGE_BELT = 5,
-	INVPAGE_NULL = 255
-};
-
-enum D2C_PlayerBodyLocs
-{
-	BODYLOC_NONE,		//Not Equipped
-	BODYLOC_HEAD,		//Helm
-	BODYLOC_NECK,		//Amulet
-	BODYLOC_TORSO,		//Body Armor
-	BODYLOC_RARM,		//Right-Hand
-	BODYLOC_LARM,		//Left-Hand
-	BODYLOC_RRIN,		//Right Ring
-	BODYLOC_LRIN,		//Left Ring
-	BODYLOC_BELT,		//Belt
-	BODYLOC_FEET,		//Boots
-	BODYLOC_GLOVES,		//Gloves
-	BODYLOC_SWRARM,		//Right-Hand on Switch
-	BODYLOC_SWLARM		//Left-Hand on Switch
-};
-
-enum D2C_NodePages
-{
-	NODEPAGE_STORAGE = 1,
-	NODEPAGE_BELTSLOTS = 2,
-	NODEPAGE_EQUIP = 3
-};
-
-enum D2C_InventoryGrids
-{
-	INVGRID_BODYLOC,
-	INVGRID_BELT,
-	INVGRID_INVENTORY,
-	INVGRID_MAX_NUM
-};
-
-enum D2C_CubeOps
-{
-	CUBEOP_COWPORTAL = 1,
-	CUBEOP_UBERDUNGEON = 2,
-	CUBEOP_UBERTRISTRAM = 3,
-	CUBEOP_SKILLPTS = 4,
-	CUBEOP_STATPTS = 5,
-	CUBEOP_REDPORTAL = 128,
-	CUBEOP_ADDSTAT = 129,
-	CUBEOP_ITEMCODE = 252,
-	CUBEOP_ITEMTYPE = 253,
-	CUBEOP_USEITEM = 254,
-	CUBEOP_USETYPE = 255,
-};
-
-enum D2C_CubeOutFlags
-{
-	CUBEFLAG_OUT_COPYMODS = 0x01,
-	CUBEFLAG_OUT_SOCKET = 0x02,
-	CUBEFLAG_OUT_ETHEREAL = 0x04,
-	CUBEFLAG_OUT_SPECIAL = 0x08,
-	CUBEFLAG_OUT_UNSOCKET = 0x10,
-	CUBEFLAG_OUT_REMOVE = 0x20,
-	CUBEFLAG_OUT_NORMAL = 0x40,
-	CUBEFLAG_OUT_EXCEPTIONAL = 0x80,
-	CUBEFLAG_OUT_ELITE = 0x100,
-	CUBEFLAG_OUT_REPAIR = 0x200,
-	CUBEFLAG_OUT_RECHARGE = 0x400,
-
-	//Custom Flags
-	CUBEFLAG_OUT_KEEPONLYSOCKETABLES = 0x800,
-	CUBEFLAG_OUT_DUPE = 0x1000,
-};
-
-enum D2C_CubeInFlags
-{
-	CUBEFLAG_IN_USEANY = 0x01,
-	CUBEFLAG_IN_ITEMCODE = 0x02,
-	CUBEFLAG_IN_NOSOCKET = 0x04,
-	CUBEFLAG_IN_SOCKETED = 0x08,
-	CUBEFLAG_IN_ETHEREAL = 0x10,
-	CUBEFLAG_IN_NOETHEREAL = 0x20,
-	CUBEFLAG_IN_SPECIAL = 0x40,
-	CUBEFLAG_IN_UPGRADED = 0x80,
-	CUBEFLAG_IN_NORMAL = 0x100,
-	CUBEFLAG_IN_EXCEPTIONAL = 0x200,
-	CUBEFLAG_IN_ELITE = 0x400,
-	CUBEFLAG_IN_NORUNES = 0x800,
-};
-#endif // CONSTANTS_ITEMS
 
 #ifndef CONSTANTS_LEVELS //Pseudo-Macro to group all constants
-enum D2C_Levels
-{
-	LEVEL_NONE = 0,
-	LEVEL_ROGUEENCAMPMENT = 1,
-	LEVEL_BLOODMOOR,
-	LEVEL_COLDPLAINS,
-	LEVEL_STONYFIELD,
-	LEVEL_DARKWOOD,
-	LEVEL_BLACKMARSH,
-	LEVEL_TAMOEHIGHLAND,
-	LEVEL_DENOFEVIL,
-	LEVEL_CAVELEV1,
-	LEVEL_UNDERGROUNDPASSAGELEV1,
-	LEVEL_HOLELEV1,
-	LEVEL_PITLEV1,
-	LEVEL_CAVELEV2,
-	LEVEL_UNDERGROUNDPASSAGELEV2,
-	LEVEL_HOLELEV2,
-	LEVEL_PITLEV2,
-	LEVEL_BURIALGROUNDS,
-	LEVEL_CRYPT,
-	LEVEL_MAUSOLEUM,
-	LEVEL_FORGOTTENTOWER,
-	LEVEL_TOWERCELLARLEV1,
-	LEVEL_TOWERCELLARLEV2,
-	LEVEL_TOWERCELLARLEV3,
-	LEVEL_TOWERCELLARLEV4,
-	LEVEL_TOWERCELLARLEV5,
-	LEVEL_MONASTERYGATE,
-	LEVEL_OUTERCLOISTER,
-	LEVEL_BARRACKS,
-	LEVEL_JAILLEV1,
-	LEVEL_JAILLEV2,
-	LEVEL_JAILLEV3,
-	LEVEL_INNERCLOISTER,
-	LEVEL_CATHEDRAL,
-	LEVEL_CATACOMBSLEV1,
-	LEVEL_CATACOMBSLEV2,
-	LEVEL_CATACOMBSLEV3,
-	LEVEL_CATACOMBSLEV4,
-	LEVEL_TRISTRAM,
-	LEVEL_MOOMOOFARM,
-	LEVEL_LUTGHOLEIN,
-	LEVEL_ROCKYWASTE,
-	LEVEL_DRYHILLS,
-	LEVEL_FAROASIS,
-	LEVEL_LOSTCITY,
-	LEVEL_VALLEYOFSNAKES,
-	LEVEL_CANYONOFTHEMAGI,
-	LEVEL_SEWERSLEV1,
-	LEVEL_SEWERSLEV2,
-	LEVEL_SEWERSLEV3,
-	LEVEL_HAREMLEV1,
-	LEVEL_HAREMLEV2,
-	LEVEL_PALACECELLARLEV1,
-	LEVEL_PALACECELLARLEV2,
-	LEVEL_PALACECELLARLEV3,
-	LEVEL_STONYTOMBLEV1,
-	LEVEL_HALLSOFTHEDEADLEV1,
-	LEVEL_HALLSOFTHEDEADLEV2,
-	LEVEL_CLAWVIPERTEMPLELEV1,
-	LEVEL_STONYTOMBLEV2,
-	LEVEL_HALLSOFTHEDEADLEV3,
-	LEVEL_CLAWVIPERTEMPLELEV2,
-	LEVEL_MAGGOTLAIRLEV1,
-	LEVEL_MAGGOTLAIRLEV2,
-	LEVEL_MAGGOTLAIRLEV3,
-	LEVEL_ANCIENTTUNNELS,
-	LEVEL_TALRASHASTOMB1,
-	LEVEL_TALRASHASTOMB2,
-	LEVEL_TALRASHASTOMB3,
-	LEVEL_TALRASHASTOMB4,
-	LEVEL_TALRASHASTOMB5,
-	LEVEL_TALRASHASTOMB6,
-	LEVEL_TALRASHASTOMB7,
-	LEVEL_DURIELSLAIR,
-	LEVEL_ARCANESANCTUARY,
-	LEVEL_KURASTDOCKTOWN,
-	LEVEL_SPIDERFOREST,
-	LEVEL_GREATMARSH,
-	LEVEL_FLAYERJUNGLE,
-	LEVEL_LOWERKURAST,
-	LEVEL_KURASTBAZAAR,
-	LEVEL_UPPERKURAST,
-	LEVEL_KURASTCAUSEWAY,
-	LEVEL_TRAVINCAL,
-	LEVEL_SPIDERCAVE,
-	LEVEL_SPIDERCAVERN,
-	LEVEL_SWAMPYPITLEV1,
-	LEVEL_SWAMPYPITLEV2,
-	LEVEL_FLAYERDUNGEONLEV1,
-	LEVEL_FLAYERDUNGEONLEV2,
-	LEVEL_SWAMPYPITLEV3,
-	LEVEL_FLAYERDUNGEONLEV3,
-	LEVEL_SEWERSA3LEV1,
-	LEVEL_SEWERSA3LEV2,
-	LEVEL_RUINEDTEMPLE,
-	LEVEL_DISUSEDFANE,
-	LEVEL_FORGOTTENRELIQUARY,
-	LEVEL_FORGOTTENTEMPLE,
-	LEVEL_RUINEDFANE,
-	LEVEL_DISUSEDRELIQUARY,
-	LEVEL_DURANCEOFHATELEV1,
-	LEVEL_DURANCEOFHATELEV2,
-	LEVEL_DURANCEOFHATELEV3,
-	LEVEL_THEPANDEMONIUMFORTRESS,
-	LEVEL_OUTERSTEPPES,
-	LEVEL_PLAINSOFDESPAIR,
-	LEVEL_CITYOFTHEDAMNED,
-	LEVEL_RIVEROFFLAME,
-	LEVEL_CHAOSSANCTUM,
-	LEVEL_HARROGATH,
-	LEVEL_BLOODYFOOTHILLS,
-	LEVEL_ID_ACT5_BARRICADE_1,
-	LEVEL_ARREATPLATEAU,
-	LEVEL_CRYSTALIZEDCAVERNLEV1,
-	LEVEL_CELLAROFPITY,
-	LEVEL_CRYSTALIZEDCAVERNLEV2,
-	LEVEL_ECHOCHAMBER,
-	LEVEL_TUNDRAWASTELANDS,
-	LEVEL_GLACIALCAVESLEV1,
-	LEVEL_GLACIALCAVESLEV2,
-	LEVEL_ROCKYSUMMIT,
-	LEVEL_NIHLATHAKSTEMPLE,
-	LEVEL_HALLSOFANGUISH,
-	LEVEL_HALLSOFDEATHSCALLING,
-	LEVEL_HALLSOFVAUGHT,
-	LEVEL_HELL1,
-	LEVEL_HELL2,
-	LEVEL_HELL3,
-	LEVEL_THEWORLDSTONEKEEPLEV1,
-	LEVEL_THEWORLDSTONEKEEPLEV2,
-	LEVEL_THEWORLDSTONEKEEPLEV3,
-	LEVEL_THRONEOFDESTRUCTION,
-	LEVEL_THEWORLDSTONECHAMBER
-};
 
-enum D2C_LvlPrestIds
-{
-	LVLPREST_NONE,
-	LVLPREST_ACT1_TOWN_1,
-	LVLPREST_ACT1_TOWN_1_TRANSITION_E,
-	LVLPREST_ACT1_TOWN_1_TRANSITION_S,
-	LVLPREST_ACT1_WILD_BORDER_1,
-	LVLPREST_ACT1_WILD_BORDER_2,
-	LVLPREST_ACT1_WILD_BORDER_3,
-	LVLPREST_ACT1_WILD_BORDER_4,
-	LVLPREST_ACT1_WILD_BORDER_5,
-	LVLPREST_ACT1_WILD_BORDER_6,
-	LVLPREST_ACT1_WILD_BORDER_7,
-	LVLPREST_ACT1_WILD_BORDER_8,
-	LVLPREST_ACT1_WILD_BORDER_9,
-	LVLPREST_ACT1_WILD_BORDER_10,
-	LVLPREST_ACT1_WILD_BORDER_11,
-	LVLPREST_ACT1_WILD_BORDER_12,
-	LVLPREST_ACT1_WILD_CLIFF_BORDER_2,
-	LVLPREST_ACT1_WILD_CLIFF_BORDER_3,
-	LVLPREST_ACT1_WILD_CLIFF_BORDER_5,
-	LVLPREST_ACT1_WILD_CLIFF_BORDER_6A,
-	LVLPREST_ACT1_WILD_CLIFF_BORDER_6B,
-	LVLPREST_ACT1_WILD_CLIFF_BORDER_6C,
-	LVLPREST_ACT1_WILD_CLIFF_BORDER_7,
-	LVLPREST_ACT1_WILD_CLIFF_BORDER_10,
-	LVLPREST_ACT1_WILD_CLIFF_CAVE_RIGHT,
-	LVLPREST_ACT1_WILD_CLIFF_CAVE_LEFT,
-	LVLPREST_ACT1_RIVER_UPPER,
-	LVLPREST_ACT1_RIVER_LOWER,
-	LVLPREST_ACT1_BRIDGE,
-	LVLPREST_ACT1_STONE_FILL_1,
-	LVLPREST_ACT1_STONE_FILL_2,
-	LVLPREST_ACT1_CORRAL_FILL,
-	LVLPREST_ACT1_FENCE_FILL_1,
-	LVLPREST_ACT1_FENCE_FILL_2,
-	LVLPREST_ACT1_FENCE_FILL_3,
-	LVLPREST_ACT1_FENCE_FILL_4,
-	LVLPREST_ACT1_FENCE_FILL_5,
-	LVLPREST_ACT1_FENCE_FILL_6,
-	LVLPREST_ACT1_SWAMP_FILL_1,
-	LVLPREST_ACT1_SWAMP_FILL_2,
-	LVLPREST_ACT1_TREE_FILL,
-	LVLPREST_ACT1_RUIN,
-	LVLPREST_ACT1_FALLEN_CAMP_1,
-	LVLPREST_ACT1_FALLEN_CAMP_2,
-	LVLPREST_ACT1_FALLEN_CAMP_BISHIBOSH,
-	LVLPREST_ACT1_CAMP,
-	LVLPREST_ACT1_POND,
-	LVLPREST_ACT1_COTTAGES_1,
-	LVLPREST_ACT1_COTTAGES_2,
-	LVLPREST_ACT1_COTTAGES_3,
-	LVLPREST_ACT1_BIVOUAC,
-	LVLPREST_ACT1_CAVE_ENTRANCE,
-	LVLPREST_ACT1_DOE_ENTRANCE,
-	LVLPREST_ACT1_CAVE_W,
-	LVLPREST_ACT1_CAVE_E,
-	LVLPREST_ACT1_CAVE_EW,
-	LVLPREST_ACT1_CAVE_S,
-	LVLPREST_ACT1_CAVE_SW,
-	LVLPREST_ACT1_CAVE_SE,
-	LVLPREST_ACT1_CAVE_SEW,
-	LVLPREST_ACT1_CAVE_N,
-	LVLPREST_ACT1_CAVE_NW,
-	LVLPREST_ACT1_CAVE_NE,
-	LVLPREST_ACT1_CAVE_NEW,
-	LVLPREST_ACT1_CAVE_NS,
-	LVLPREST_ACT1_CAVE_NSW,
-	LVLPREST_ACT1_CAVE_NSE,
-	LVLPREST_ACT1_CAVE_NSEW,
-	LVLPREST_ACT1_CAVE_THEME_W,
-	LVLPREST_ACT1_CAVE_THEME_E,
-	LVLPREST_ACT1_CAVE_THEME_EW,
-	LVLPREST_ACT1_CAVE_THEME_S,
-	LVLPREST_ACT1_CAVE_THEME_SW,
-	LVLPREST_ACT1_CAVE_THEME_SE,
-	LVLPREST_ACT1_CAVE_THEME_SEW,
-	LVLPREST_ACT1_CAVE_THEME_N,
-	LVLPREST_ACT1_CAVE_THEME_NW,
-	LVLPREST_ACT1_CAVE_THEME_NE,
-	LVLPREST_ACT1_CAVE_THEME_NEW,
-	LVLPREST_ACT1_CAVE_THEME_NS,
-	LVLPREST_ACT1_CAVE_THEME_NSW,
-	LVLPREST_ACT1_CAVE_THEME_NSE,
-	LVLPREST_ACT1_CAVE_THEME_NSEW,
-	LVLPREST_ACT1_CAVE_PREV_W,
-	LVLPREST_ACT1_CAVE_PREV_E,
-	LVLPREST_ACT1_CAVE_PREV_S,
-	LVLPREST_ACT1_CAVE_PREV_N,
-	LVLPREST_ACT1_CAVE_NEXT_W,
-	LVLPREST_ACT1_CAVE_NEXT_E,
-	LVLPREST_ACT1_CAVE_NEXT_S,
-	LVLPREST_ACT1_CAVE_NEXT_N,
-	LVLPREST_ACT1_CAVE_DOWN_W,
-	LVLPREST_ACT1_CAVE_DOWN_E,
-	LVLPREST_ACT1_CAVE_DOWN_S,
-	LVLPREST_ACT1_CAVE_DOWN_N,
-	LVLPREST_ACT1_CAVE_DEN_OF_EVIL_W,
-	LVLPREST_ACT1_CAVE_DEN_OF_EVIL_E,
-	LVLPREST_ACT1_CAVE_DEN_OF_EVIL_S,
-	LVLPREST_ACT1_CAVE_DEN_OF_EVIL_N,
-	LVLPREST_ACT1_CAVE_COLDCROW_W,
-	LVLPREST_ACT1_CAVE_COLDCROW_E,
-	LVLPREST_ACT1_CAVE_COLDCROW_S,
-	LVLPREST_ACT1_CAVE_COLDCROW_N,
-	LVLPREST_ACT1_CAVE_TREASURE_1,
-	LVLPREST_ACT1_CAVE_TREASURE_2,
-	LVLPREST_ACT1_CAVE_TREASURE_3,
-	LVLPREST_ACT1_CAVE_TREASURE_4,
-	LVLPREST_ACT1_CAVE_TREASURE_5,
-	LVLPREST_ACT1_GRAVEYARD,
-	LVLPREST_ACT1_CRYPT_W,
-	LVLPREST_ACT1_CRYPT_E,
-	LVLPREST_ACT1_CRYPT_EW,
-	LVLPREST_ACT1_CRYPT_S,
-	LVLPREST_ACT1_CRYPT_SW,
-	LVLPREST_ACT1_CRYPT_SE,
-	LVLPREST_ACT1_CRYPT_SEW,
-	LVLPREST_ACT1_CRYPT_N,
-	LVLPREST_ACT1_CRYPT_NW,
-	LVLPREST_ACT1_CRYPT_NE,
-	LVLPREST_ACT1_CRYPT_NEW,
-	LVLPREST_ACT1_CRYPT_NS,
-	LVLPREST_ACT1_CRYPT_NSW,
-	LVLPREST_ACT1_CRYPT_NSE,
-	LVLPREST_ACT1_CRYPT_NSEW,
-	LVLPREST_ACT1_CRYPT_THEME_W,
-	LVLPREST_ACT1_CRYPT_THEME_E,
-	LVLPREST_ACT1_CRYPT_THEME_EW,
-	LVLPREST_ACT1_CRYPT_THEME_S,
-	LVLPREST_ACT1_CRYPT_THEME_SW,
-	LVLPREST_ACT1_CRYPT_THEME_SE,
-	LVLPREST_ACT1_CRYPT_THEME_SEW,
-	LVLPREST_ACT1_CRYPT_THEME_N,
-	LVLPREST_ACT1_CRYPT_THEME_NW,
-	LVLPREST_ACT1_CRYPT_THEME_NE,
-	LVLPREST_ACT1_CRYPT_THEME_NEW,
-	LVLPREST_ACT1_CRYPT_THEME_NS,
-	LVLPREST_ACT1_CRYPT_THEME_NSW,
-	LVLPREST_ACT1_CRYPT_THEME_NSE,
-	LVLPREST_ACT1_CRYPT_THEME_NSEW,
-	LVLPREST_ACT1_CRYPT_PREV_W,
-	LVLPREST_ACT1_CRYPT_PREV_E,
-	LVLPREST_ACT1_CRYPT_PREV_S,
-	LVLPREST_ACT1_CRYPT_PREV_N,
-	LVLPREST_ACT1_CRYPT_NEXT_W,
-	LVLPREST_ACT1_CRYPT_NEXT_E,
-	LVLPREST_ACT1_CRYPT_NEXT_S,
-	LVLPREST_ACT1_CRYPT_NEXT_N,
-	LVLPREST_ACT1_CRYPT_BONEBREAK_W,
-	LVLPREST_ACT1_CRYPT_BONEBREAK_E,
-	LVLPREST_ACT1_CRYPT_BONEBREAK_S,
-	LVLPREST_ACT1_CRYPT_BONEBREAK_N,
-	LVLPREST_ACT1_CRYPT_CHEST_W,
-	LVLPREST_ACT1_CRYPT_CHEST_E,
-	LVLPREST_ACT1_CRYPT_CHEST_S,
-	LVLPREST_ACT1_CRYPT_CHEST_N,
-	LVLPREST_ACT1_CRYPT_PORTAL_W,
-	LVLPREST_ACT1_CRYPT_PORTAL_E,
-	LVLPREST_ACT1_CRYPT_PORTAL_S,
-	LVLPREST_ACT1_CRYPT_PORTAL_N,
-	LVLPREST_ACT1_CRYPT_COUNTESS_X,
-	LVLPREST_ACT1_CAIRN_STONES,
-	LVLPREST_ACT1_INIFUS,
-	LVLPREST_ACT1_TOWER_TOME,
-	LVLPREST_ACT1_TOWER_1,
-	LVLPREST_ACT1_TOWER_2,
-	LVLPREST_ACT1_MON_FRONT,
-	LVLPREST_ACT1_COURTYARD_1,
-	LVLPREST_ACT1_BARRACKS_COURT_CONNECT,
-	LVLPREST_ACT1_BARRACKS_W,
-	LVLPREST_ACT1_BARRACKS_E,
-	LVLPREST_ACT1_BARRACKS_EW,
-	LVLPREST_ACT1_BARRACKS_S,
-	LVLPREST_ACT1_BARRACKS_SW,
-	LVLPREST_ACT1_BARRACKS_SE,
-	LVLPREST_ACT1_BARRACKS_SEW,
-	LVLPREST_ACT1_BARRACKS_N,
-	LVLPREST_ACT1_BARRACKS_NW,
-	LVLPREST_ACT1_BARRACKS_NE,
-	LVLPREST_ACT1_BARRACKS_NEW,
-	LVLPREST_ACT1_BARRACKS_NS,
-	LVLPREST_ACT1_BARRACKS_NSW,
-	LVLPREST_ACT1_BARRACKS_NSE,
-	LVLPREST_ACT1_BARRACKS_NSEW,
-	LVLPREST_ACT1_BARRACKS_THEME_W,
-	LVLPREST_ACT1_BARRACKS_THEME_E,
-	LVLPREST_ACT1_BARRACKS_THEME_EW,
-	LVLPREST_ACT1_BARRACKS_THEME_S,
-	LVLPREST_ACT1_BARRACKS_THEME_SW,
-	LVLPREST_ACT1_BARRACKS_THEME_SE,
-	LVLPREST_ACT1_BARRACKS_THEME_SEW,
-	LVLPREST_ACT1_BARRACKS_THEME_N,
-	LVLPREST_ACT1_BARRACKS_THEME_NW,
-	LVLPREST_ACT1_BARRACKS_THEME_NE,
-	LVLPREST_ACT1_BARRACKS_THEME_NEW,
-	LVLPREST_ACT1_BARRACKS_THEME_NS,
-	LVLPREST_ACT1_BARRACKS_THEME_NSW,
-	LVLPREST_ACT1_BARRACKS_THEME_NSE,
-	LVLPREST_ACT1_BARRACKS_THEME_NSEW,
-	LVLPREST_ACT1_BARRACKS_NEXT_W,
-	LVLPREST_ACT1_BARRACKS_NEXT_E,
-	LVLPREST_ACT1_BARRACKS_NEXT_S,
-	LVLPREST_ACT1_BARRACKS_NEXT_N,
-	LVLPREST_ACT1_BARRACKS_FORGE_W,
-	LVLPREST_ACT1_BARRACKS_FORGE_E,
-	LVLPREST_ACT1_BARRACKS_FORGE_S,
-	LVLPREST_ACT1_BARRACKS_FORGE_N,
-	LVLPREST_ACT1_JAIL_W,
-	LVLPREST_ACT1_JAIL_E,
-	LVLPREST_ACT1_JAIL_EW,
-	LVLPREST_ACT1_JAIL_S,
-	LVLPREST_ACT1_JAIL_SW,
-	LVLPREST_ACT1_JAIL_SE,
-	LVLPREST_ACT1_JAIL_SEW,
-	LVLPREST_ACT1_JAIL_N,
-	LVLPREST_ACT1_JAIL_NW,
-	LVLPREST_ACT1_JAIL_NE,
-	LVLPREST_ACT1_JAIL_NEW,
-	LVLPREST_ACT1_JAIL_NS,
-	LVLPREST_ACT1_JAIL_NSW,
-	LVLPREST_ACT1_JAIL_NSE,
-	LVLPREST_ACT1_JAIL_NSEW,
-	LVLPREST_ACT1_JAIL_THEME_W,
-	LVLPREST_ACT1_JAIL_THEME_E,
-	LVLPREST_ACT1_JAIL_THEME_EW,
-	LVLPREST_ACT1_JAIL_THEME_S,
-	LVLPREST_ACT1_JAIL_THEME_SW,
-	LVLPREST_ACT1_JAIL_THEME_SE,
-	LVLPREST_ACT1_JAIL_THEME_SEW,
-	LVLPREST_ACT1_JAIL_THEME_N,
-	LVLPREST_ACT1_JAIL_THEME_NW,
-	LVLPREST_ACT1_JAIL_THEME_NE,
-	LVLPREST_ACT1_JAIL_THEME_NEW,
-	LVLPREST_ACT1_JAIL_THEME_NS,
-	LVLPREST_ACT1_JAIL_THEME_NSW,
-	LVLPREST_ACT1_JAIL_THEME_NSE,
-	LVLPREST_ACT1_JAIL_THEME_NSEW,
-	LVLPREST_ACT1_JAIL_PREV_W,
-	LVLPREST_ACT1_JAIL_PREV_E,
-	LVLPREST_ACT1_JAIL_PREV_S,
-	LVLPREST_ACT1_JAIL_PREV_N,
-	LVLPREST_ACT1_JAIL_NEXT_W,
-	LVLPREST_ACT1_JAIL_NEXT_E,
-	LVLPREST_ACT1_JAIL_NEXT_S,
-	LVLPREST_ACT1_JAIL_NEXT_N,
-	LVLPREST_ACT1_JAIL_CATH_W,
-	LVLPREST_ACT1_JAIL_CATH_E,
-	LVLPREST_ACT1_JAIL_CATH_S,
-	LVLPREST_ACT1_JAIL_CATH_N,
-	LVLPREST_ACT1_JAIL_WAYPOINT_W,
-	LVLPREST_ACT1_JAIL_WAYPOINT_E,
-	LVLPREST_ACT1_JAIL_WAYPOINT_S,
-	LVLPREST_ACT1_JAIL_WAYPOINT_N,
-	LVLPREST_ACT1_JAIL_PITSPAWN_W,
-	LVLPREST_ACT1_JAIL_PITSPAWN_E,
-	LVLPREST_ACT1_JAIL_PITSPAWN_S,
-	LVLPREST_ACT1_JAIL_PITSPAWN_N,
-	LVLPREST_ACT1_COURTYARD_2,
-	LVLPREST_ACT1_CATHEDRAL,
-	LVLPREST_ACT1_CATACOMBS_W,
-	LVLPREST_ACT1_CATACOMBS_E,
-	LVLPREST_ACT1_CATACOMBS_EW,
-	LVLPREST_ACT1_CATACOMBS_S,
-	LVLPREST_ACT1_CATACOMBS_SW,
-	LVLPREST_ACT1_CATACOMBS_SE,
-	LVLPREST_ACT1_CATACOMBS_SEW,
-	LVLPREST_ACT1_CATACOMBS_N,
-	LVLPREST_ACT1_CATACOMBS_NW,
-	LVLPREST_ACT1_CATACOMBS_NE,
-	LVLPREST_ACT1_CATACOMBS_NEW,
-	LVLPREST_ACT1_CATACOMBS_NS,
-	LVLPREST_ACT1_CATACOMBS_NSW,
-	LVLPREST_ACT1_CATACOMBS_NSE,
-	LVLPREST_ACT1_CATACOMBS_NSEW,
-	LVLPREST_ACT1_CATACOMBS_THEME_W,
-	LVLPREST_ACT1_CATACOMBS_THEME_E,
-	LVLPREST_ACT1_CATACOMBS_THEME_EW,
-	LVLPREST_ACT1_CATACOMBS_THEME_S,
-	LVLPREST_ACT1_CATACOMBS_THEME_SW,
-	LVLPREST_ACT1_CATACOMBS_THEME_SE,
-	LVLPREST_ACT1_CATACOMBS_THEME_SEW,
-	LVLPREST_ACT1_CATACOMBS_THEME_N,
-	LVLPREST_ACT1_CATACOMBS_THEME_NW,
-	LVLPREST_ACT1_CATACOMBS_THEME_NE,
-	LVLPREST_ACT1_CATACOMBS_THEME_NEW,
-	LVLPREST_ACT1_CATACOMBS_THEME_NS,
-	LVLPREST_ACT1_CATACOMBS_THEME_NSW,
-	LVLPREST_ACT1_CATACOMBS_THEME_NSE,
-	LVLPREST_ACT1_CATACOMBS_THEME_NSEW,
-	LVLPREST_ACT1_CATACOMBS_PREV_EW,
-	LVLPREST_ACT1_CATACOMBS_PREV_NS,
-	LVLPREST_ACT1_CATACOMBS_PREV_NSEW,
-	LVLPREST_ACT1_CATACOMBS_NEXT_W,
-	LVLPREST_ACT1_CATACOMBS_NEXT_E,
-	LVLPREST_ACT1_CATACOMBS_NEXT_S,
-	LVLPREST_ACT1_CATACOMBS_NEXT_N,
-	LVLPREST_ACT1_CATACOMBS_WAYPOINT_W,
-	LVLPREST_ACT1_CATACOMBS_WAYPOINT_E,
-	LVLPREST_ACT1_CATACOMBS_WAYPOINT_S,
-	LVLPREST_ACT1_CATACOMBS_WAYPOINT_N,
-	LVLPREST_ACT1_ANDARIEL,
-	LVLPREST_ACT1_TRISTRAM,
-	LVLPREST_ACT2_TOWN,
-	LVLPREST_ACT2_SEWER_W,
-	LVLPREST_ACT2_SEWER_E,
-	LVLPREST_ACT2_SEWER_EW,
-	LVLPREST_ACT2_SEWER_S,
-	LVLPREST_ACT2_SEWER_SW,
-	LVLPREST_ACT2_SEWER_SE,
-	LVLPREST_ACT2_SEWER_SEW,
-	LVLPREST_ACT2_SEWER_N,
-	LVLPREST_ACT2_SEWER_NW,
-	LVLPREST_ACT2_SEWER_NE,
-	LVLPREST_ACT2_SEWER_NEW,
-	LVLPREST_ACT2_SEWER_NS,
-	LVLPREST_ACT2_SEWER_NSW,
-	LVLPREST_ACT2_SEWER_NSE,
-	LVLPREST_ACT2_SEWER_NSEW,
-	LVLPREST_ACT2_SEWER_THEME_W,
-	LVLPREST_ACT2_SEWER_THEME_E,
-	LVLPREST_ACT2_SEWER_THEME_EW,
-	LVLPREST_ACT2_SEWER_THEME_S,
-	LVLPREST_ACT2_SEWER_THEME_SW,
-	LVLPREST_ACT2_SEWER_THEME_SE,
-	LVLPREST_ACT2_SEWER_THEME_SEW,
-	LVLPREST_ACT2_SEWER_THEME_N,
-	LVLPREST_ACT2_SEWER_THEME_NW,
-	LVLPREST_ACT2_SEWER_THEME_NE,
-	LVLPREST_ACT2_SEWER_THEME_NEW,
-	LVLPREST_ACT2_SEWER_THEME_NS,
-	LVLPREST_ACT2_SEWER_THEME_NSW,
-	LVLPREST_ACT2_SEWER_THEME_NSE,
-	LVLPREST_ACT2_SEWER_THEME_NSEW,
-	LVLPREST_ACT2_SEWER_PREV_W,
-	LVLPREST_ACT2_SEWER_PREV_E,
-	LVLPREST_ACT2_SEWER_PREV_S,
-	LVLPREST_ACT2_SEWER_PREV_N,
-	LVLPREST_ACT2_SEWER_PREV_NS,
-	LVLPREST_ACT2_SEWER_NEXT_W,
-	LVLPREST_ACT2_SEWER_NEXT_E,
-	LVLPREST_ACT2_SEWER_NEXT_S,
-	LVLPREST_ACT2_SEWER_NEXT_N,
-	LVLPREST_ACT2_SEWER_RADAMENTS_LAIR_W,
-	LVLPREST_ACT2_SEWER_RADAMENTS_LAIR_E,
-	LVLPREST_ACT2_SEWER_RADAMENTS_LAIR_S,
-	LVLPREST_ACT2_SEWER_RADAMENTS_LAIR_N,
-	LVLPREST_ACT2_SEWER_WAYPOINT_W,
-	LVLPREST_ACT2_SEWER_WAYPOINT_E,
-	LVLPREST_ACT2_SEWER_WAYPOINT_S,
-	LVLPREST_ACT2_SEWER_WAYPOINT_N,
-	LVLPREST_ACT2_SEWER_CHEST_W,
-	LVLPREST_ACT2_SEWER_CHEST_E,
-	LVLPREST_ACT2_SEWER_CHEST_S,
-	LVLPREST_ACT2_SEWER_CHEST_N,
-	LVLPREST_ACT2_HAREM,
-	LVLPREST_ACT2_CORRUPT_HAREM_NE,
-	LVLPREST_ACT2_CORRUPT_HAREM_SE,
-	LVLPREST_ACT2_CORRUPT_HAREM_SW,
-	LVLPREST_ACT2_CORRUPT_HAREM_NW,
-	LVLPREST_ACT2_BASEMENT_NE,
-	LVLPREST_ACT2_BASEMENT_SE,
-	LVLPREST_ACT2_BASEMENT_SW,
-	LVLPREST_ACT2_BASEMENT_NW,
-	LVLPREST_ACT2_DESERT_TRANSITION_W,
-	LVLPREST_ACT2_DESERT_TRANSITION_N,
-	LVLPREST_ACT2_DESERT_BORDER_1,
-	LVLPREST_ACT2_DESERT_BORDER_2,
-	LVLPREST_ACT2_DESERT_BORDER_3,
-	LVLPREST_ACT2_DESERT_BORDER_4,
-	LVLPREST_ACT2_DESERT_BORDER_5,
-	LVLPREST_ACT2_DESERT_BORDER_6,
-	LVLPREST_ACT2_DESERT_BORDER_7,
-	LVLPREST_ACT2_DESERT_BORDER_8,
-	LVLPREST_ACT2_DESERT_BORDER_9,
-	LVLPREST_ACT2_DESERT_BORDER_10,
-	LVLPREST_ACT2_DESERT_BORDER_11,
-	LVLPREST_ACT2_DESERT_BORDER_12,
-	LVLPREST_ACT2_DESERT_CLIFF_RIGHT_ENDS,
-	LVLPREST_ACT2_DESERT_CLIFF_RIGHT_WALL,
-	LVLPREST_ACT2_DESERT_CLIFF_RIGHT_PATH,
-	LVLPREST_ACT2_DESERT_CLIFF_LEFT_ENDS,
-	LVLPREST_ACT2_DESERT_CLIFF_LEFT_WALL,
-	LVLPREST_ACT2_DESERT_CLIFF_LEFT_PATH,
-	LVLPREST_ACT2_DESERT_CLIFF_TOP,
-	LVLPREST_ACT2_DESERT_CLIFF_RIGHT_KING_TOMB,
-	LVLPREST_ACT2_DESERT_CLIFF_RIGHT_KING_ENDS,
-	LVLPREST_ACT2_DESERT_CLIFF_LEFT_KING_TOMB,
-	LVLPREST_ACT2_DESERT_CLIFF_LEFT_KING_ENDS,
-	LVLPREST_ACT2_DESERT_CLIFF_TOP_KING_TOMB,
-	LVLPREST_ACT2_DESERT_TOMB_1,
-	LVLPREST_ACT2_DESERT_TOMB_2,
-	LVLPREST_ACT2_DESERT_LAIR_1,
-	LVLPREST_ACT2_DESERT_LAIR_2,
-	LVLPREST_ACT2_DESERT_VALLEY_RUIN_1,
-	LVLPREST_ACT2_DESERT_VALLEY_RUIN_2,
-	LVLPREST_ACT2_DESERT_VALLEY_WARP,
-	LVLPREST_ACT2_DESERT_OASIS_1,
-	LVLPREST_ACT2_DESERT_OASIS_2,
-	LVLPREST_ACT2_DESERT_OASIS_3,
-	LVLPREST_ACT2_DESERT_FILL_MESA_1,
-	LVLPREST_ACT2_DESERT_FILL_HEAD_1,
-	LVLPREST_ACT2_DESERT_FILL_HEAD_2,
-	LVLPREST_ACT2_DESERT_FILL_BONE_1,
-	LVLPREST_ACT2_DESERT_FILL_BONE_2,
-	LVLPREST_ACT2_DESERT_FILL_WAGON_1,
-	LVLPREST_ACT2_DESERT_FILL_BERMS_1,
-	LVLPREST_ACT2_DESERT_FILL_BERMS_2,
-	LVLPREST_ACT2_DESERT_FILL_BERMS_3,
-	LVLPREST_ACT2_DESERT_FILL_BERMS_4,
-	LVLPREST_ACT2_DESERT_RUINS_16X16,
-	LVLPREST_ACT2_DESERT_RUINS_16X08,
-	LVLPREST_ACT2_DESERT_RUINS_08X16,
-	LVLPREST_ACT2_DESERT_RUINS_08X08,
-	LVLPREST_ACT2_DESERT_RUINS_SEWER,
-	LVLPREST_ACT2_DESERT_RUINS_ELDER,
-	LVLPREST_ACT2_TOMB_W,
-	LVLPREST_ACT2_TOMB_E,
-	LVLPREST_ACT2_TOMB_EW,
-	LVLPREST_ACT2_TOMB_S,
-	LVLPREST_ACT2_TOMB_SW,
-	LVLPREST_ACT2_TOMB_SE,
-	LVLPREST_ACT2_TOMB_SEW,
-	LVLPREST_ACT2_TOMB_N,
-	LVLPREST_ACT2_TOMB_NW,
-	LVLPREST_ACT2_TOMB_NE,
-	LVLPREST_ACT2_TOMB_NEW,
-	LVLPREST_ACT2_TOMB_NS,
-	LVLPREST_ACT2_TOMB_NSW,
-	LVLPREST_ACT2_TOMB_NSE,
-	LVLPREST_ACT2_TOMB_NSEW,
-	LVLPREST_ACT2_TOMB_THEME_W,
-	LVLPREST_ACT2_TOMB_THEME_E,
-	LVLPREST_ACT2_TOMB_THEME_EW,
-	LVLPREST_ACT2_TOMB_THEME_S,
-	LVLPREST_ACT2_TOMB_THEME_SW,
-	LVLPREST_ACT2_TOMB_THEME_SE,
-	LVLPREST_ACT2_TOMB_THEME_SEW,
-	LVLPREST_ACT2_TOMB_THEME_N,
-	LVLPREST_ACT2_TOMB_THEME_NW,
-	LVLPREST_ACT2_TOMB_THEME_NE,
-	LVLPREST_ACT2_TOMB_THEME_NEW,
-	LVLPREST_ACT2_TOMB_THEME_NS,
-	LVLPREST_ACT2_TOMB_THEME_NSW,
-	LVLPREST_ACT2_TOMB_THEME_NSE,
-	LVLPREST_ACT2_TOMB_THEME_NSEW,
-	LVLPREST_ACT2_TOMB_PREV_SEW,
-	LVLPREST_ACT2_TOMB_PREV_NEW,
-	LVLPREST_ACT2_TOMB_PREV_NSW,
-	LVLPREST_ACT2_TOMB_PREV_NSE,
-	LVLPREST_ACT2_TOMB_NEXT_W,
-	LVLPREST_ACT2_TOMB_NEXT_E,
-	LVLPREST_ACT2_TOMB_NEXT_S,
-	LVLPREST_ACT2_TOMB_NEXT_N,
-	LVLPREST_ACT2_TOMB_TREASURE_W,
-	LVLPREST_ACT2_TOMB_TREASURE_E,
-	LVLPREST_ACT2_TOMB_TREASURE_S,
-	LVLPREST_ACT2_TOMB_TREASURE_N,
-	LVLPREST_ACT2_TOMB_CUBE_W,
-	LVLPREST_ACT2_TOMB_CUBE_E,
-	LVLPREST_ACT2_TOMB_CUBE_S,
-	LVLPREST_ACT2_TOMB_CUBE_N,
-	LVLPREST_ACT2_TOMB_TALRASHA_W,
-	LVLPREST_ACT2_TOMB_TALRASHA_E,
-	LVLPREST_ACT2_TOMB_TALRASHA_S,
-	LVLPREST_ACT2_TOMB_TALRASHA_N,
-	LVLPREST_ACT2_TOMB_LEATHERARM_W,
-	LVLPREST_ACT2_TOMB_LEATHERARM_E,
-	LVLPREST_ACT2_TOMB_LEATHERARM_S,
-	LVLPREST_ACT2_TOMB_LEATHERARM_N,
-	LVLPREST_ACT2_TOMB_KAA_W,
-	LVLPREST_ACT2_TOMB_KAA_E,
-	LVLPREST_ACT2_TOMB_KAA_S,
-	LVLPREST_ACT2_TOMB_KAA_N,
-	LVLPREST_ACT2_TOMB_CHEST_W,
-	LVLPREST_ACT2_TOMB_CHEST_E,
-	LVLPREST_ACT2_TOMB_CHEST_S,
-	LVLPREST_ACT2_TOMB_CHEST_N,
-	LVLPREST_ACT2_TOMB_WAYPOINT_W,
-	LVLPREST_ACT2_TOMB_WAYPOINT_E,
-	LVLPREST_ACT2_TOMB_WAYPOINT_S,
-	LVLPREST_ACT2_TOMB_WAYPOINT_N,
-	LVLPREST_ACT2_TOMB_TAINTED_SUN_X,
-	LVLPREST_ACT2_DURIELS_LAIR,
-	LVLPREST_ACT2_LAIR_W,
-	LVLPREST_ACT2_LAIR_E,
-	LVLPREST_ACT2_LAIR_EW,
-	LVLPREST_ACT2_LAIR_S,
-	LVLPREST_ACT2_LAIR_SW,
-	LVLPREST_ACT2_LAIR_SE,
-	LVLPREST_ACT2_LAIR_SEW,
-	LVLPREST_ACT2_LAIR_N,
-	LVLPREST_ACT2_LAIR_NW,
-	LVLPREST_ACT2_LAIR_NE,
-	LVLPREST_ACT2_LAIR_NEW,
-	LVLPREST_ACT2_LAIR_NS,
-	LVLPREST_ACT2_LAIR_NSW,
-	LVLPREST_ACT2_LAIR_NSE,
-	LVLPREST_ACT2_LAIR_NSEW,
-	LVLPREST_ACT2_LAIR_PREV_W,
-	LVLPREST_ACT2_LAIR_PREV_E,
-	LVLPREST_ACT2_LAIR_PREV_S,
-	LVLPREST_ACT2_LAIR_PREV_N,
-	LVLPREST_ACT2_LAIR_NEXT_W,
-	LVLPREST_ACT2_LAIR_NEXT_E,
-	LVLPREST_ACT2_LAIR_NEXT_S,
-	LVLPREST_ACT2_LAIR_NEXT_N,
-	LVLPREST_ACT2_LAIR_TREASURE_W,
-	LVLPREST_ACT2_LAIR_TREASURE_E,
-	LVLPREST_ACT2_LAIR_TREASURE_S,
-	LVLPREST_ACT2_LAIR_TREASURE_N,
-	LVLPREST_ACT2_LAIR_TIGHT_SPOT_S,
-	LVLPREST_ACT2_ARCANE_W,
-	LVLPREST_ACT2_ARCANE_E,
-	LVLPREST_ACT2_ARCANE_EW,
-	LVLPREST_ACT2_ARCANE_S,
-	LVLPREST_ACT2_ARCANE_SW,
-	LVLPREST_ACT2_ARCANE_SE,
-	LVLPREST_ACT2_ARCANE_SEW,
-	LVLPREST_ACT2_ARCANE_N,
-	LVLPREST_ACT2_ARCANE_NW,
-	LVLPREST_ACT2_ARCANE_NE,
-	LVLPREST_ACT2_ARCANE_NEW,
-	LVLPREST_ACT2_ARCANE_NS,
-	LVLPREST_ACT2_ARCANE_NSW,
-	LVLPREST_ACT2_ARCANE_NSE,
-	LVLPREST_ACT2_ARCANE_NSEW,
-	LVLPREST_ACT2_ARCANE_SUMMONER_W,
-	LVLPREST_ACT2_ARCANE_SUMMONER_E,
-	LVLPREST_ACT2_ARCANE_SUMMONER_S,
-	LVLPREST_ACT2_ARCANE_SUMMONER_N,
-	LVLPREST_ACT3_TOWN,
-	LVLPREST_ACT3_JUNGLE_W,
-	LVLPREST_ACT3_JUNGLE_E,
-	LVLPREST_ACT3_JUNGLE_EW,
-	LVLPREST_ACT3_JUNGLE_S,
-	LVLPREST_ACT3_JUNGLE_SW,
-	LVLPREST_ACT3_JUNGLE_SE,
-	LVLPREST_ACT3_JUNGLE_SEW,
-	LVLPREST_ACT3_JUNGLE_N,
-	LVLPREST_ACT3_JUNGLE_NW,
-	LVLPREST_ACT3_JUNGLE_NE,
-	LVLPREST_ACT3_JUNGLE_NEW,
-	LVLPREST_ACT3_JUNGLE_NS,
-	LVLPREST_ACT3_JUNGLE_NSW,
-	LVLPREST_ACT3_JUNGLE_NSE,
-	LVLPREST_ACT3_JUNGLE_NSEW,
-	LVLPREST_ACT3_JUNGLE_W_E,
-	LVLPREST_ACT3_JUNGLE_W_S,
-	LVLPREST_ACT3_JUNGLE_W_N,
-	LVLPREST_ACT3_JUNGLE_E_W,
-	LVLPREST_ACT3_JUNGLE_E_S,
-	LVLPREST_ACT3_JUNGLE_E_N,
-	LVLPREST_ACT3_JUNGLE_EW_S,
-	LVLPREST_ACT3_JUNGLE_EW_N,
-	LVLPREST_ACT3_JUNGLE_S_W,
-	LVLPREST_ACT3_JUNGLE_S_E,
-	LVLPREST_ACT3_JUNGLE_S_N,
-	LVLPREST_ACT3_JUNGLE_SW_E,
-	LVLPREST_ACT3_JUNGLE_SW_N,
-	LVLPREST_ACT3_JUNGLE_SE_W,
-	LVLPREST_ACT3_JUNGLE_SE_N,
-	LVLPREST_ACT3_JUNGLE_SEW_N,
-	LVLPREST_ACT3_JUNGLE_N_W,
-	LVLPREST_ACT3_JUNGLE_N_E,
-	LVLPREST_ACT3_JUNGLE_N_S,
-	LVLPREST_ACT3_JUNGLE_NW_E,
-	LVLPREST_ACT3_JUNGLE_NW_S,
-	LVLPREST_ACT3_JUNGLE_NE_W,
-	LVLPREST_ACT3_JUNGLE_NE_S,
-	LVLPREST_ACT3_JUNGLE_NEW_S,
-	LVLPREST_ACT3_JUNGLE_NS_W,
-	LVLPREST_ACT3_JUNGLE_NS_E,
-	LVLPREST_ACT3_JUNGLE_NSW_E,
-	LVLPREST_ACT3_JUNGLE_NSE_W,
-	LVLPREST_ACT3_JUNGLE_HEAD,
-	LVLPREST_ACT3_JUNGLE_TAIL,
-	LVLPREST_ACT3_CLEARING_WEBBY_W,
-	LVLPREST_ACT3_CLEARING_WEBBY_E,
-	LVLPREST_ACT3_CLEARING_WEBBY_EW,
-	LVLPREST_ACT3_CLEARING_WEBBY_S,
-	LVLPREST_ACT3_CLEARING_WEBBY_SW,
-	LVLPREST_ACT3_CLEARING_WEBBY_SE,
-	LVLPREST_ACT3_CLEARING_WEBBY_N,
-	LVLPREST_ACT3_CLEARING_WEBBY_NW,
-	LVLPREST_ACT3_CLEARING_WEBBY_NE,
-	LVLPREST_ACT3_CLEARING_WEBBY_NS,
-	LVLPREST_ACT3_CLEARING_BOGGY_W,
-	LVLPREST_ACT3_CLEARING_BOGGY_E,
-	LVLPREST_ACT3_CLEARING_BOGGY_EW,
-	LVLPREST_ACT3_CLEARING_BOGGY_S,
-	LVLPREST_ACT3_CLEARING_BOGGY_SW,
-	LVLPREST_ACT3_CLEARING_BOGGY_SE,
-	LVLPREST_ACT3_CLEARING_BOGGY_N,
-	LVLPREST_ACT3_CLEARING_BOGGY_NW,
-	LVLPREST_ACT3_CLEARING_BOGGY_NE,
-	LVLPREST_ACT3_CLEARING_BOGGY_NS,
-	LVLPREST_ACT3_CLEARING_PYGMY_W,
-	LVLPREST_ACT3_CLEARING_PYGMY_E,
-	LVLPREST_ACT3_CLEARING_PYGMY_EW,
-	LVLPREST_ACT3_CLEARING_PYGMY_S,
-	LVLPREST_ACT3_CLEARING_PYGMY_SW,
-	LVLPREST_ACT3_CLEARING_PYGMY_SE,
-	LVLPREST_ACT3_CLEARING_PYGMY_N,
-	LVLPREST_ACT3_CLEARING_PYGMY_NW,
-	LVLPREST_ACT3_CLEARING_PYGMY_NE,
-	LVLPREST_ACT3_CLEARING_PYGMY_NS,
-	LVLPREST_ACT3_SLUMS_BORDER_N,
-	LVLPREST_ACT3_SLUMS_BORDER_S,
-	LVLPREST_ACT3_SLUMS_BORDER_E,
-	LVLPREST_ACT3_SLUMS_BORDER_W,
-	LVLPREST_ACT3_SLUMS_BORDER_NE,
-	LVLPREST_ACT3_SLUMS_BORDER_NW,
-	LVLPREST_ACT3_SLUMS_BORDER_SE,
-	LVLPREST_ACT3_SLUMS_BORDER_SW,
-	LVLPREST_ACT3_SLUMS_GATE_N,
-	LVLPREST_ACT3_SLUMS_GATE_S,
-	LVLPREST_ACT3_SLUMS_08X08,
-	LVLPREST_ACT3_SLUMS_08X16,
-	LVLPREST_ACT3_SLUMS_16X08,
-	LVLPREST_ACT3_SLUMS_16X16,
-	LVLPREST_ACT3_BURBS_BORDER_N,
-	LVLPREST_ACT3_BURBS_BORDER_S,
-	LVLPREST_ACT3_BURBS_BORDER_E,
-	LVLPREST_ACT3_BURBS_BORDER_W,
-	LVLPREST_ACT3_BURBS_BORDER_NE,
-	LVLPREST_ACT3_BURBS_BORDER_NW,
-	LVLPREST_ACT3_BURBS_BORDER_SE,
-	LVLPREST_ACT3_BURBS_BORDER_SW,
-	LVLPREST_ACT3_BURBS_GATE_N,
-	LVLPREST_ACT3_BURBS_GATE_S,
-	LVLPREST_ACT3_BURBS_SEWER,
-	LVLPREST_ACT3_BURBS_TEMPLE,
-	LVLPREST_ACT3_BURBS_WAYPOINT,
-	LVLPREST_ACT3_BURBS_08X08,
-	LVLPREST_ACT3_BURBS_08X16,
-	LVLPREST_ACT3_BURBS_16X08,
-	LVLPREST_ACT3_BURBS_16X16,
-	LVLPREST_ACT3_METRO_BORDER_N,
-	LVLPREST_ACT3_METRO_BORDER_S,
-	LVLPREST_ACT3_METRO_BORDER_E,
-	LVLPREST_ACT3_METRO_BORDER_W,
-	LVLPREST_ACT3_METRO_BORDER_NE,
-	LVLPREST_ACT3_METRO_BORDER_NW,
-	LVLPREST_ACT3_METRO_BORDER_SE,
-	LVLPREST_ACT3_METRO_BORDER_SW,
-	LVLPREST_ACT3_METRO_GATE_N,
-	LVLPREST_ACT3_METRO_GATE_S,
-	LVLPREST_ACT3_METRO_SEWER,
-	LVLPREST_ACT3_METROTEMPLE,
-	LVLPREST_ACT3_METRO_08X08,
-	LVLPREST_ACT3_METRO_08X16,
-	LVLPREST_ACT3_METRO_16X08,
-	LVLPREST_ACT3_METRO_16X16,
-	LVLPREST_ACT3_BRIDGE,
-	LVLPREST_ACT3_TRAVINCAL_NW,
-	LVLPREST_ACT3_TRAVINCAL_N,
-	LVLPREST_ACT3_TRAVINCAL_NE,
-	LVLPREST_ACT3_TRAVINCAL_SW,
-	LVLPREST_ACT3_TRAVINCAL_S,
-	LVLPREST_ACT3_TRAVINCAL_SE,
-	LVLPREST_ACT3_SPIDER_SW,
-	LVLPREST_ACT3_SPIDER_SE,
-	LVLPREST_ACT3_SPIDER_NW,
-	LVLPREST_ACT3_SPIDER_NE,
-	LVLPREST_ACT3_SPIDER_CHEST_NW,
-	LVLPREST_ACT3_SPIDER_CHEST_NE,
-	LVLPREST_ACT3_DUNGEON_W,
-	LVLPREST_ACT3_DUNGEON_E,
-	LVLPREST_ACT3_DUNGEON_EW,
-	LVLPREST_ACT3_DUNGEON_S,
-	LVLPREST_ACT3_DUNGEON_SW,
-	LVLPREST_ACT3_DUNGEON_SE,
-	LVLPREST_ACT3_DUNGEON_SEW,
-	LVLPREST_ACT3_DUNGEON_N,
-	LVLPREST_ACT3_DUNGEON_NW,
-	LVLPREST_ACT3_DUNGEON_NE,
-	LVLPREST_ACT3_DUNGEON_NEW,
-	LVLPREST_ACT3_DUNGEON_NS,
-	LVLPREST_ACT3_DUNGEON_NSW,
-	LVLPREST_ACT3_DUNGEON_NSE,
-	LVLPREST_ACT3_DUNGEON_NSEW,
-	LVLPREST_ACT3_DUNGEON_THEME_W,
-	LVLPREST_ACT3_DUNGEON_THEME_E,
-	LVLPREST_ACT3_DUNGEON_THEME_EW,
-	LVLPREST_ACT3_DUNGEON_THEME_S,
-	LVLPREST_ACT3_DUNGEON_THEME_SW,
-	LVLPREST_ACT3_DUNGEON_THEME_SE,
-	LVLPREST_ACT3_DUNGEON_THEME_SEW,
-	LVLPREST_ACT3_DUNGEON_THEME_N,
-	LVLPREST_ACT3_DUNGEON_THEME_NW,
-	LVLPREST_ACT3_DUNGEON_THEME_NE,
-	LVLPREST_ACT3_DUNGEON_THEME_NEW,
-	LVLPREST_ACT3_DUNGEON_THEME_NS,
-	LVLPREST_ACT3_DUNGEON_THEME_NSW,
-	LVLPREST_ACT3_DUNGEON_THEME_NSE,
-	LVLPREST_ACT3_DUNGEON_THEME_NSEW,
-	LVLPREST_ACT3_DUNGEON_PREV_W,
-	LVLPREST_ACT3_DUNGEON_PREV_E,
-	LVLPREST_ACT3_DUNGEON_PREV_S,
-	LVLPREST_ACT3_DUNGEON_PREV_N,
-	LVLPREST_ACT3_DUNGEON_NEXT_W,
-	LVLPREST_ACT3_DUNGEON_NEXT_E,
-	LVLPREST_ACT3_DUNGEON_NEXT_S,
-	LVLPREST_ACT3_DUNGEON_NEXT_N,
-	LVLPREST_ACT3_DUNGEON_TREASURE_1,
-	LVLPREST_ACT3_DUNGEON_TREASURE_2,
-	LVLPREST_ACT3_SEWER_W,
-	LVLPREST_ACT3_SEWER_E,
-	LVLPREST_ACT3_SEWER_EW,
-	LVLPREST_ACT3_SEWER_S,
-	LVLPREST_ACT3_SEWER_SW,
-	LVLPREST_ACT3_SEWER_SE,
-	LVLPREST_ACT3_SEWER_SEW,
-	LVLPREST_ACT3_SEWER_N,
-	LVLPREST_ACT3_SEWER_NW,
-	LVLPREST_ACT3_SEWER_NE,
-	LVLPREST_ACT3_SEWER_NEW,
-	LVLPREST_ACT3_SEWER_NS,
-	LVLPREST_ACT3_SEWER_NSW,
-	LVLPREST_ACT3_SEWER_NSE,
-	LVLPREST_ACT3_SEWER_NSEW,
-	LVLPREST_ACT3_SEWER_THEME_W,
-	LVLPREST_ACT3_SEWER_THEME_E,
-	LVLPREST_ACT3_SEWER_THEME_EW,
-	LVLPREST_ACT3_SEWER_THEME_S,
-	LVLPREST_ACT3_SEWER_THEME_SW,
-	LVLPREST_ACT3_SEWER_THEME_SE,
-	LVLPREST_ACT3_SEWER_THEME_SEW,
-	LVLPREST_ACT3_SEWER_THEME_N,
-	LVLPREST_ACT3_SEWER_THEME_NW,
-	LVLPREST_ACT3_SEWER_THEME_NE,
-	LVLPREST_ACT3_SEWER_THEME_NEW,
-	LVLPREST_ACT3_SEWER_THEME_NS,
-	LVLPREST_ACT3_SEWER_THEME_NSW,
-	LVLPREST_ACT3_SEWER_THEME_NSE,
-	LVLPREST_ACT3_SEWER_THEME_NSEW,
-	LVLPREST_ACT3_SEWER_PREV_SW,
-	LVLPREST_ACT3_SEWER_PREV_SE,
-	LVLPREST_ACT3_SEWER_PREV_NW,
-	LVLPREST_ACT3_SEWER_PREV_NE,
-	LVLPREST_ACT3_SEWER_DRAIN_W,
-	LVLPREST_ACT3_SEWER_DRAIN_E,
-	LVLPREST_ACT3_SEWER_DRAIN_S,
-	LVLPREST_ACT3_SEWER_DRAIN_N,
-	LVLPREST_ACT3_SEWER_CHEST_W,
-	LVLPREST_ACT3_SEWER_CHEST_E,
-	LVLPREST_ACT3_SEWER_CHEST_S,
-	LVLPREST_ACT3_SEWER_CHEST_N,
-	LVLPREST_ACT3_SEWER_TREASURE_X,
-	LVLPREST_ACT3_TEMPLE_1,
-	LVLPREST_ACT3_TEMPLE_2,
-	LVLPREST_ACT3_TEMPLE_3,
-	LVLPREST_ACT3_TEMPLE_4,
-	LVLPREST_ACT3_TEMPLE_5,
-	LVLPREST_ACT3_TEMPLE_6,
-	LVLPREST_ACT3_MEPHISTO_W,
-	LVLPREST_ACT3_MEPHISTO_E,
-	LVLPREST_ACT3_MEPHISTO_EW,
-	LVLPREST_ACT3_MEPHISTO_S,
-	LVLPREST_ACT3_MEPHISTO_SW,
-	LVLPREST_ACT3_MEPHISTO_SE,
-	LVLPREST_ACT3_MEPHISTO_SEW,
-	LVLPREST_ACT3_MEPHISTO_N,
-	LVLPREST_ACT3_MEPHISTO_NW,
-	LVLPREST_ACT3_MEPHISTO_NE,
-	LVLPREST_ACT3_MEPHISTO_NEW,
-	LVLPREST_ACT3_MEPHISTO_NS,
-	LVLPREST_ACT3_MEPHISTO_NSW,
-	LVLPREST_ACT3_MEPHISTO_NSE,
-	LVLPREST_ACT3_MEPHISTO_NSEW,
-	LVLPREST_ACT3_MEPHISTO_THEME_W,
-	LVLPREST_ACT3_MEPHISTO_THEME_E,
-	LVLPREST_ACT3_MEPHISTO_THEME_EW,
-	LVLPREST_ACT3_MEPHISTO_THEME_S,
-	LVLPREST_ACT3_MEPHISTO_THEME_SW,
-	LVLPREST_ACT3_MEPHISTO_THEME_SE,
-	LVLPREST_ACT3_MEPHISTO_THEME_SEW,
-	LVLPREST_ACT3_MEPHISTO_THEME_N,
-	LVLPREST_ACT3_MEPHISTO_THEME_NW,
-	LVLPREST_ACT3_MEPHISTO_THEME_NE,
-	LVLPREST_ACT3_MEPHISTO_THEME_NEW,
-	LVLPREST_ACT3_MEPHISTO_THEME_NS,
-	LVLPREST_ACT3_MEPHISTO_THEME_NSW,
-	LVLPREST_ACT3_MEPHISTO_THEME_NSE,
-	LVLPREST_ACT3_MEPHISTO_THEME_NSEW,
-	LVLPREST_ACT3_MEPHISTO_PREV_W,
-	LVLPREST_ACT3_MEPHISTO_PREV_E,
-	LVLPREST_ACT3_MEPHISTO_PREV_S,
-	LVLPREST_ACT3_MEPHISTO_PREV_N,
-	LVLPREST_ACT3_MEPHISTO_NEXT_W,
-	LVLPREST_ACT3_MEPHISTO_NEXT_E,
-	LVLPREST_ACT3_MEPHISTO_NEXT_S,
-	LVLPREST_ACT3_MEPHISTO_NEXT_N,
-	LVLPREST_ACT3_MEPHISTO_WAYPOINT_W,
-	LVLPREST_ACT3_MEPHISTO_WAYPOINT_E,
-	LVLPREST_ACT3_MEPHISTO_WAYPOINT_S,
-	LVLPREST_ACT3_MEPHISTO_WAYPOINT_N,
-	LVLPREST_ACT3_MEPHISTO_COMPLEX,
-	LVLPREST_ACT4_FORTRESS,
-	LVLPREST_ACT4_FORTRESS_TRANSITION,
-	LVLPREST_ACT4_MESA_BORDER_1,
-	LVLPREST_ACT4_MESA_BORDER_2,
-	LVLPREST_ACT4_MESA_BORDER_3,
-	LVLPREST_ACT4_MESA_BORDER_4,
-	LVLPREST_ACT4_MESA_BORDER_5,
-	LVLPREST_ACT4_MESA_BORDER_6,
-	LVLPREST_ACT4_MESA_BORDER_7,
-	LVLPREST_ACT4_MESA_BORDER_8,
-	LVLPREST_ACT4_MESA_BORDER_9,
-	LVLPREST_ACT4_MESA_BORDER_10,
-	LVLPREST_ACT4_MESA_BORDER_11,
-	LVLPREST_ACT4_MESA_BORDER_12,
-	LVLPREST_ACT4_MESA_WARP,
-	LVLPREST_ACT4_MESA_1_24X24,
-	LVLPREST_ACT4_MESA_1_16X16,
-	LVLPREST_ACT4_MESA_1_16X08,
-	LVLPREST_ACT4_MESA_1_08X16,
-	LVLPREST_ACT4_MESA_1_08X08,
-	LVLPREST_ACT4_MESA_2_24X24,
-	LVLPREST_ACT4_MESA_2_16X16,
-	LVLPREST_ACT4_MESA_2_16X08,
-	LVLPREST_ACT4_MESA_2_08X16,
-	LVLPREST_ACT4_MESA_2_08X08,
-	LVLPREST_ACT4_MESA_2_IZUAL,
-	LVLPREST_ACT4_MESA_3_24X24,
-	LVLPREST_ACT4_MESA_3_16X16,
-	LVLPREST_ACT4_MESA_3_16X08,
-	LVLPREST_ACT4_MESA_3_08X16,
-	LVLPREST_ACT4_MESA_3_08X08,
-	LVLPREST_ACT4_PITS_1_16X16,
-	LVLPREST_ACT4_PITS_1_16X08,
-	LVLPREST_ACT4_PITS_1_08X16,
-	LVLPREST_ACT4_PITS_1_08X08,
-	LVLPREST_ACT4_PITS_2_16X16,
-	LVLPREST_ACT4_PITS_2_16X08,
-	LVLPREST_ACT4_PITS_2_08X16,
-	LVLPREST_ACT4_PITS_2_08X08,
-	LVLPREST_ACT4_LAVA_X,
-	LVLPREST_ACT4_LAVA_W,
-	LVLPREST_ACT4_LAVA_E,
-	LVLPREST_ACT4_LAVA_EW,
-	LVLPREST_ACT4_LAVA_S,
-	LVLPREST_ACT4_LAVA_SW,
-	LVLPREST_ACT4_LAVA_SE,
-	LVLPREST_ACT4_LAVA_SEW,
-	LVLPREST_ACT4_LAVA_N,
-	LVLPREST_ACT4_LAVA_NW,
-	LVLPREST_ACT4_LAVA_NE,
-	LVLPREST_ACT4_LAVA_NEW,
-	LVLPREST_ACT4_LAVA_NS,
-	LVLPREST_ACT4_LAVA_NSW,
-	LVLPREST_ACT4_LAVA_NSE,
-	LVLPREST_ACT4_LAVA_NSEW,
-	LVLPREST_ACT4_LAVA_WARP_N,
-	LVLPREST_ACT4_LAVA_FORGE_W,
-	LVLPREST_ACT4_LAVA_FORGE_E,
-	LVLPREST_ACT4_BRIDGE_1,
-	LVLPREST_ACT4_BRIDGE_2,
-	LVLPREST_ACT4_DIABLO_ENTRY,
-	LVLPREST_ACT4_DIABLO_ARM_W,
-	LVLPREST_ACT4_DIABLO_ARM_E,
-	LVLPREST_ACT4_DIABLO_ARM_S,
-	LVLPREST_ACT4_DIABLO_ARM_N,
-	LVLPREST_ACT4_DIABLO_HEART,
-	LVLPREST_ACT5_TOWN,
-	LVLPREST_ACT5_TEMPLE_FINAL_ROOM,
-	LVLPREST_ACT5_SIEGE_TO_TOWN,
-	LVLPREST_ACT5_SIEGE_STRIP_1,
-	LVLPREST_ACT5_SIEGE_STRIP_2,
-	LVLPREST_ACT5_SIEGE_STRIP_3,
-	LVLPREST_ACT5_SIEGE_STRIP_4,
-	LVLPREST_ACT5_SIEGE_STRIP_5,
-	LVLPREST_ACT5_SIEGE_STRIP_6,
-	LVLPREST_ACT5_SIEGE_STRIP_7,
-	LVLPREST_ACT5_SIEGE_STRIP_8,
-	LVLPREST_ACT5_SIEGE_STRIP_9,
-	LVLPREST_ACT5_SIEGE_STRIP_10,
-	LVLPREST_ACT5_SIEGE_STRIP_11,
-	LVLPREST_ACT5_SIEGE_STRIP_12,
-	LVLPREST_ACT5_SIEGE_STRIP_13,
-	LVLPREST_ACT5_SIEGE_TO_BARRICADE,
-	LVLPREST_ACT5_BARRICADE_TO_SIEGE,
-	LVLPREST_ACT5_BARRICADE_CLIFF_BORDER_1,
-	LVLPREST_ACT5_BARRICADE_CLIFF_BORDER_2,
-	LVLPREST_ACT5_BARRICADE_CLIFF_BORDER_3,
-	LVLPREST_ACT5_BARRICADE_CLIFF_BORDER_4,
-	LVLPREST_ACT5_BARRICADE_CLIFF_BORDER_5,
-	LVLPREST_ACT5_BARRICADE_CLIFF_BORDER_6,
-	LVLPREST_ACT5_BARRICADE_CLIFF_BORDER_7,
-	LVLPREST_ACT5_BARRICADE_CLIFF_BORDER_8,
-	LVLPREST_ACT5_BARRICADE_CLIFF_BORDER_9,
-	LVLPREST_ACT5_BARRICADE_CLIFF_BORDER_10,
-	LVLPREST_ACT5_BARRICADE_CLIFF_BORDER_11,
-	LVLPREST_ACT5_BARRICADE_CLIFF_BORDER_12,
-	LVLPREST_ACT5_BARRICADE_RAVINE_BORDER_1,
-	LVLPREST_ACT5_BARRICADE_RAVINE_BORDER_2,
-	LVLPREST_ACT5_BARRICADE_RAVINE_BORDER_3,
-	LVLPREST_ACT5_BARRICADE_RAVINE_BORDER_4,
-	LVLPREST_ACT5_BARRICADE_RAVINE_BORDER_5,
-	LVLPREST_ACT5_BARRICADE_RAVINE_BORDER_6,
-	LVLPREST_ACT5_BARRICADE_RAVINE_BORDER_7,
-	LVLPREST_ACT5_BARRICADE_RAVINE_BORDER_8,
-	LVLPREST_ACT5_BARRICADE_RAVINE_BORDER_9,
-	LVLPREST_ACT5_BARRICADE_RAVINE_BORDER_10,
-	LVLPREST_ACT5_BARRICADE_RAVINE_BORDER_11,
-	LVLPREST_ACT5_BARRICADE_RAVINE_BORDER_12,
-	LVLPREST_ACT5_BARRICADE_RAVINE_CLIFF_BORDER_5,
-	LVLPREST_ACT5_BARRICADE_CLIFF_RAVINE_BORDER_7,
-	LVLPREST_ACT5_BARRICADE_ENTRANCE_16X32,
-	LVLPREST_ACT5_BARRICADE_ENTRANCE_32X16,
-	LVLPREST_ACT5_BARRICADE_EXIT_32X16,
-	LVLPREST_ACT5_BARRICADE_EXIT_16X32,
-	LVLPREST_ACT5_BARRICADE_FROM_CAVE_32X16,
-	LVLPREST_ACT5_BARRICADE_FROM_CAVE_16X32,
-	LVLPREST_ACT5_BARRICADE_TO_CAVE_32X16,
-	LVLPREST_ACT5_BARRICADE_TO_CAVE_16X32,
-	LVLPREST_ACT5_BARRICADE_1,
-	LVLPREST_ACT5_BARRICADE_2,
-	LVLPREST_ACT5_BARRICADE_3,
-	LVLPREST_ACT5_BARRICADE_4,
-	LVLPREST_ACT5_BARRICADE_5,
-	LVLPREST_ACT5_BARRICADE_6,
-	LVLPREST_ACT5_BARRICADE_7,
-	LVLPREST_ACT5_BARRICADE_8,
-	LVLPREST_ACT5_BARRICADE_9,
-	LVLPREST_ACT5_BARRICADE_10,
-	LVLPREST_ACT5_BARRICADE_11,
-	LVLPREST_ACT5_BARRICADE_12,
-	LVLPREST_ACT5_BARRICADE_13,
-	LVLPREST_ACT5_BARRICADE_14,
-	LVLPREST_ACT5_BARRICADE_15,
-	LVLPREST_ACT5_BARRICADE_16,
-	LVLPREST_ACT5_BARRICADE_PRISON_1,
-	LVLPREST_ACT5_BARRICADE_PRISON_2,
-	LVLPREST_ACT5_BARRICADE_PRISON_3,
-	LVLPREST_ACT5_BARRICADE_PRISON_4,
-	LVLPREST_ACT5_BARRICADE_PRISON_5,
-	LVLPREST_ACT5_BARRICADE_PRISON_6,
-	LVLPREST_ACT5_BARRICADE_PRISON_7,
-	LVLPREST_ACT5_BARRICADE_PRISON_8,
-	LVLPREST_ACT5_BARRICADE_BUILDING,
-	LVLPREST_ACT5_BARRICADE_FILLER,
-	LVLPREST_ACT5_BARRICADE_FILLER_TREASURE,
-	LVLPREST_ACT5_BARRICADE_RUINS_N_1,
-	LVLPREST_ACT5_BARRICADE_RUINS_N_2,
-	LVLPREST_ACT5_BARRICADE_RUINS_N_TREASURE,
-	LVLPREST_ACT5_BARRICADE_RUINS_W_1,
-	LVLPREST_ACT5_BARRICADE_RUINS_W_2,
-	LVLPREST_ACT5_BARRICADE_RUINS_W_TREASURE,
-	LVLPREST_ACT5_BARRICADE_SNOW_LAKE__1,
-	LVLPREST_ACT5_BARRICADE_SNOW_LAKE__2,
-	LVLPREST_ACT5_BARRICADE_SNOW_OTHER,
-	LVLPREST_ACT5_BARRICADE_SNOW_TREASURE,
-	LVLPREST_ACT5_BARRICADE_SNOW_UNIQUE,
-	LVLPREST_ACT5_BARRICADE_WAYPOINT_DIRT,
-	LVLPREST_ACT5_BARRICADE_WAYPOINT_SNOW,
-	LVLPREST_ACT5_BARRICADE_HELL_PORTAL_N,
-	LVLPREST_ACT5_BARRICADE_HELL_PORTAL_W,
-	LVLPREST_ACT5_BARRICADE_CLIFF_BORDER_1_SNOW,
-	LVLPREST_ACT5_BARRICADE_CLIFF_BORDER_2_SNOW,
-	LVLPREST_ACT5_BARRICADE_CLIFF_BORDER_3_SNOW,
-	LVLPREST_ACT5_BARRICADE_CLIFF_BORDER_4_SNOW,
-	LVLPREST_ACT5_BARRICADE_CLIFF_BORDER_5_SNOW,
-	LVLPREST_ACT5_BARRICADE_CLIFF_BORDER_6_SNOW,
-	LVLPREST_ACT5_BARRICADE_CLIFF_BORDER_7_SNOW,
-	LVLPREST_ACT5_BARRICADE_CLIFF_BORDER_8_SNOW,
-	LVLPREST_ACT5_BARRICADE_CLIFF_BORDER_9_SNOW,
-	LVLPREST_ACT5_BARRICADE_CLIFF_BORDER_10_SNOW,
-	LVLPREST_ACT5_BARRICADE_CLIFF_BORDER_11_SNOW,
-	LVLPREST_ACT5_BARRICADE_CLIFF_BORDER_12_SNOW,
-	LVLPREST_ACT5_BARRICADE_RAVINE_BORDER_1_SNOW,
-	LVLPREST_ACT5_BARRICADE_RAVINE_BORDER_2_SNOW,
-	LVLPREST_ACT5_BARRICADE_RAVINE_BORDER_3_SNOW,
-	LVLPREST_ACT5_BARRICADE_RAVINE_BORDER_4_SNOW,
-	LVLPREST_ACT5_BARRICADE_RAVINE_BORDER_5_SNOW,
-	LVLPREST_ACT5_BARRICADE_RAVINE_BORDER_6_SNOW,
-	LVLPREST_ACT5_BARRICADE_RAVINE_BORDER_7_SNOW,
-	LVLPREST_ACT5_BARRICADE_RAVINE_BORDER_8_SNOW,
-	LVLPREST_ACT5_BARRICADE_RAVINE_BORDER_9_SNOW,
-	LVLPREST_ACT5_BARRICADE_RAVINE_BORDER_10_SNOW,
-	LVLPREST_ACT5_BARRICADE_RAVINE_BORDER_11_SNOW,
-	LVLPREST_ACT5_BARRICADE_RAVINE_BORDER_12_SNOW,
-	LVLPREST_ACT5_BARRICADE_RAVINE_CLIFF_BORDER_5_SNOW,
-	LVLPREST_ACT5_BARRICADE_CLIFF_RAVINE_BORDER_7_SNOW,
-	LVLPREST_ACT5_BARRICADE_FROM_CAVE_32X16_SNOW,
-	LVLPREST_ACT5_BARRICADE_FROM_CAVE_16X32_SNOW,
-	LVLPREST_ACT5_BARRICADE_TO_CAVE_32X16_SNOW,
-	LVLPREST_ACT5_BARRICADE_TO_CAVE_16X32_SNOW,
-	LVLPREST_ACT5_BARRICADE_1_SNOW,
-	LVLPREST_ACT5_BARRICADE_2_SNOW,
-	LVLPREST_ACT5_BARRICADE_3_SNOW,
-	LVLPREST_ACT5_BARRICADE_4_SNOW,
-	LVLPREST_ACT5_BARRICADE_5_SNOW,
-	LVLPREST_ACT5_BARRICADE_6_SNOW,
-	LVLPREST_ACT5_BARRICADE_7_SNOW,
-	LVLPREST_ACT5_BARRICADE_8_SNOW,
-	LVLPREST_ACT5_BARRICADE_9_SNOW,
-	LVLPREST_ACT5_BARRICADE_10_SNOW,
-	LVLPREST_ACT5_BARRICADE_11_SNOW,
-	LVLPREST_ACT5_BARRICADE_12_SNOW,
-	LVLPREST_ACT5_BARRICADE_13_SNOW,
-	LVLPREST_ACT5_BARRICADE_14_SNOW,
-	LVLPREST_ACT5_BARRICADE_15_SNOW,
-	LVLPREST_ACT5_BARRICADE_16_SNOW,
-	LVLPREST_ACT5_ICE_W,
-	LVLPREST_ACT5_ICE_E,
-	LVLPREST_ACT5_ICE_EW,
-	LVLPREST_ACT5_ICE_S,
-	LVLPREST_ACT5_ICE_SW,
-	LVLPREST_ACT5_ICE_SE,
-	LVLPREST_ACT5_ICE_SEW,
-	LVLPREST_ACT5_ICE_N,
-	LVLPREST_ACT5_ICE_NW,
-	LVLPREST_ACT5_ICE_NE,
-	LVLPREST_ACT5_ICE_NEW,
-	LVLPREST_ACT5_ICE_NS,
-	LVLPREST_ACT5_ICE_NSW,
-	LVLPREST_ACT5_ICE_NSE,
-	LVLPREST_ACT5_ICE_NSEW,
-	LVLPREST_ACT5_ICE_PREV_W,
-	LVLPREST_ACT5_ICE_PREV_E,
-	LVLPREST_ACT5_ICE_PREV_S,
-	LVLPREST_ACT5_ICE_PREV_N,
-	LVLPREST_ACT5_ICE_NEXT_W,
-	LVLPREST_ACT5_ICE_NEXT_E,
-	LVLPREST_ACT5_ICE_NEXT_S,
-	LVLPREST_ACT5_ICE_NEXT_N,
-	LVLPREST_ACT5_ICE_DOWN_W,
-	LVLPREST_ACT5_ICE_DOWN_E,
-	LVLPREST_ACT5_ICE_DOWN_S,
-	LVLPREST_ACT5_ICE_DOWN_N,
-	LVLPREST_ACT5_ICE_THEME_W,
-	LVLPREST_ACT5_ICE_THEME_E,
-	LVLPREST_ACT5_ICE_THEME_S,
-	LVLPREST_ACT5_ICE_THEME_N,
-	LVLPREST_ACT5_ICE_WAYPOINT_N,
-	LVLPREST_ACT5_ICE_WAYPOINT_S,
-	LVLPREST_ACT5_ICE_WAYPOINT_E,
-	LVLPREST_ACT5_ICE_WAYPOINT_W,
-	LVLPREST_ACT5_ICE_RIVER_A,
-	LVLPREST_ACT5_ICE_RIVER_B,
-	LVLPREST_ACT5_ICE_POOL_A,
-	LVLPREST_ACT5_ICE_POOL_B,
-	LVLPREST_ACT5_TEMPLE_NE,
-	LVLPREST_ACT5_TEMPLE_NW,
-	LVLPREST_ACT5_TEMPLE_SE_UP,
-	LVLPREST_ACT5_TEMPLE_SW,
-	LVLPREST_ACT5_TEMPLE_NE_DOWN,
-	LVLPREST_ACT5_TEMPLE_NW_DOWN,
-	LVLPREST_ACT5_TEMPLE_SW_DOWN,
-	LVLPREST_ACT5_TEMPLE_NE_WAYPOINT,
-	LVLPREST_ACT5_TEMPLE_NW_WAYPOINT,
-	LVLPREST_ACT5_TEMPLE_SE_WAYPOINT,
-	LVLPREST_ACT5_TEMPLE_SW_WAYPOINT,
-	LVLPREST_ACT5_LAVA_N,
-	LVLPREST_ACT5_LAVA_S,
-	LVLPREST_ACT5_LAVA_E,
-	LVLPREST_ACT5_LAVA_W,
-	LVLPREST_ACT5_LAVA_EW,
-	LVLPREST_ACT5_LAVA_NS,
-	LVLPREST_ACT5_BAAL_W,
-	LVLPREST_ACT5_BAAL_E,
-	LVLPREST_ACT5_BAAL_EW,
-	LVLPREST_ACT5_BAAL_S,
-	LVLPREST_ACT5_BAAL_SW,
-	LVLPREST_ACT5_BAAL_SE,
-	LVLPREST_ACT5_BAAL_SEW,
-	LVLPREST_ACT5_BAAL_N,
-	LVLPREST_ACT5_BAAL_NW,
-	LVLPREST_ACT5_BAAL_NE,
-	LVLPREST_ACT5_BAAL_NEW,
-	LVLPREST_ACT5_BAAL_NS,
-	LVLPREST_ACT5_BAAL_NSW,
-	LVLPREST_ACT5_BAAL_NSE,
-	LVLPREST_ACT5_BAAL_NSEW,
-	LVLPREST_ACT5_BAAL_PREV_NEW,
-	LVLPREST_ACT5_BAAL_PREV_NSE,
-	LVLPREST_ACT5_BAAL_PREV_NSW,
-	LVLPREST_ACT5_BAAL_PREV_SEW,
-	LVLPREST_ACT5_BAAL_NEXT_N,
-	LVLPREST_ACT5_BAAL_NEXT_S,
-	LVLPREST_ACT5_BAAL_NEXT_E,
-	LVLPREST_ACT5_BAAL_NEXT_W,
-	LVLPREST_ACT5_BAAL_WAYPOINT_N,
-	LVLPREST_ACT5_BAAL_WAYPOINT_S,
-	LVLPREST_ACT5_BAAL_WAYPOINT_E,
-	LVLPREST_ACT5_BAAL_WAYPOINT_W,
-	LVLPREST_ACT5_THRONEROOM,
-	LVLPREST_ACT5_WORLDSTONE,
-	LVLPREST_ACT5_TEMPLE_ENTRANCE,
-	LVLPREST_ACT5_BAAL_ENTRANCE,
-};
-
-enum D2C_LvlSubIds
-{
-	LVLSUB_ACT1_BORDER_CLIFFS,
-	LVLSUB_ACT1_BORDER_MIDDLE,
-	LVLSUB_ACT1_BORDER_CORNER,
-	LVLSUB_ACT1_BORDER_BORDER,
-	LVLSUB_ACT1_WAYPOINT,
-	LVLSUB_ACT1_WELL_SHRINE,
-	LVLSUB_ACT1_WILD_OBJECTS,
-	LVLSUB_ACT2_WAYPOINT,
-	LVLSUB_ACT2_WELL_SHRINE,
-	LVLSUB_ACT2_DESERT_OBJECTS,
-	LVLSUB_ACT5_SNOW__TREES_ROCKS__SMALL_STUFF,
-	LVLSUB_ACT5_DIRT_PATCHES__SNOWY_OBJECTS,
-	LVLSUB_ACT5_BARRICADE,
-};
-
-enum D2C_LvlTypes
-{
-	LVLTYPE_NONE,
-	LVLTYPE_ACT1_TOWN,
-	LVLTYPE_ACT1_WILDERNESS,
-	LVLTYPE_ACT1_CAVE,
-	LVLTYPE_ACT1_CRYPT,
-	LVLTYPE_ACT1_MONASTERY,
-	LVLTYPE_ACT1_COURTYARD,
-	LVLTYPE_ACT1_BARRACKS,
-	LVLTYPE_ACT1_JAIL,
-	LVLTYPE_ACT1_CATHEDRAL,
-	LVLTYPE_ACT1_CATACOMBS,
-	LVLTYPE_ACT1_TRISTRAM,
-	LVLTYPE_ACT2_TOWN,
-	LVLTYPE_ACT2_SEWER,
-	LVLTYPE_ACT2_HAREM,
-	LVLTYPE_ACT2_BASEMENT,
-	LVLTYPE_ACT2_DESERT,
-	LVLTYPE_ACT2_TOMB,
-	LVLTYPE_ACT2_LAIR,
-	LVLTYPE_ACT2_ARCANE,
-	LVLTYPE_ACT3_TOWN,
-	LVLTYPE_ACT3_JUNGLE,
-	LVLTYPE_ACT3_KURAST,
-	LVLTYPE_ACT3_SPIDER,
-	LVLTYPE_ACT3_DUNGEON,
-	LVLTYPE_ACT3_SEWER,
-	LVLTYPE_ACT4_TOWN,
-	LVLTYPE_ACT4_MESA,
-	LVLTYPE_ACT4_LAVA,
-	LVLTYPE_ACT5_TOWN,
-	LVLTYPE_ACT5_SIEGE,
-	LVLTYPE_ACT5_BARRICADE,
-	LVLTYPE_ACT5_TEMPLE,
-	LVLTYPE_ACT5_ICE_CAVES,
-	LVLTYPE_ACT5_BAAL,
-	LVLTYPE_ACT5_LAVA,
-	NUM_LVLTPYES
-};
 
 enum D2C_WarpTileTypes
 {
@@ -1837,744 +270,6 @@ enum D2InactiveMonsterFlags
 	INACTIVEMONSTERFLAG_TYPE_SUPERUNIQUE = 0x00000800,
 };
 
-enum D2C_MonsterIds
-{
-	MONSTER_SKELETON1 = 0,
-	MONSTER_SKELETON2,
-	MONSTER_SKELETON3,
-	MONSTER_SKELETON4,
-	MONSTER_SKELETON5,
-	MONSTER_ZOMBIE1,
-	MONSTER_ZOMBIE2,
-	MONSTER_ZOMBIE3,
-	MONSTER_ZOMBIE4,
-	MONSTER_ZOMBIE5,
-	MONSTER_BIGHEAD1,
-	MONSTER_BIGHEAD2,
-	MONSTER_BIGHEAD3,
-	MONSTER_BIGHEAD4,
-	MONSTER_BIGHEAD5,
-	MONSTER_FOULCROW1,
-	MONSTER_FOULCROW2,
-	MONSTER_FOULCROW3,
-	MONSTER_FOULCROW4,
-	MONSTER_FALLEN1,
-	MONSTER_FALLEN2,
-	MONSTER_FALLEN3,
-	MONSTER_FALLEN4,
-	MONSTER_FALLEN5,
-	MONSTER_BRUTE2,
-	MONSTER_BRUTE3,
-	MONSTER_BRUTE4,
-	MONSTER_BRUTE5,
-	MONSTER_BRUTE1,
-	MONSTER_SANDRAIDER1,
-	MONSTER_SANDRAIDER2,
-	MONSTER_SANDRAIDER3,
-	MONSTER_SANDRAIDER4,
-	MONSTER_SANDRAIDER5,
-	MONSTER_GORGON1,
-	MONSTER_GORGON2,
-	MONSTER_GORGON3,
-	MONSTER_GORGON4,
-	MONSTER_WRAITH1,
-	MONSTER_WRAITH2,
-	MONSTER_WRAITH3,
-	MONSTER_WRAITH4,
-	MONSTER_WRAITH5,
-	MONSTER_CORRUPTROGUE1,
-	MONSTER_CORRUPTROGUE2,
-	MONSTER_CORRUPTROGUE3,
-	MONSTER_CORRUPTROGUE4,
-	MONSTER_CORRUPTROGUE5,
-	MONSTER_BABOON1,
-	MONSTER_BABOON2,
-	MONSTER_BABOON3,
-	MONSTER_BABOON4,
-	MONSTER_BABOON5,
-	MONSTER_GOATMAN1,
-	MONSTER_GOATMAN2,
-	MONSTER_GOATMAN3,
-	MONSTER_GOATMAN4,
-	MONSTER_GOATMAN5,
-	MONSTER_FALLENSHAMAN1,
-	MONSTER_FALLENSHAMAN2,
-	MONSTER_FALLENSHAMAN3,
-	MONSTER_FALLENSHAMAN4,
-	MONSTER_FALLENSHAMAN5,
-	MONSTER_QUILLRAT1,
-	MONSTER_QUILLRAT2,
-	MONSTER_QUILLRAT3,
-	MONSTER_QUILLRAT4,
-	MONSTER_QUILLRAT5,
-	MONSTER_SANDMAGGOT1,
-	MONSTER_SANDMAGGOT2,
-	MONSTER_SANDMAGGOT3,
-	MONSTER_SANDMAGGOT4,
-	MONSTER_SANDMAGGOT5,
-	MONSTER_CLAWVIPER1,
-	MONSTER_CLAWVIPER2,
-	MONSTER_CLAWVIPER3,
-	MONSTER_CLAWVIPER4,
-	MONSTER_CLAWVIPER5,
-	MONSTER_SANDLEAPER1,
-	MONSTER_SANDLEAPER2,
-	MONSTER_SANDLEAPER3,
-	MONSTER_SANDLEAPER4,
-	MONSTER_SANDLEAPER5,
-	MONSTER_PANTHERWOMAN1,
-	MONSTER_PANTHERWOMAN2,
-	MONSTER_PANTHERWOMAN3,
-	MONSTER_PANTHERWOMAN4,
-	MONSTER_SWARM1,
-	MONSTER_SWARM2,
-	MONSTER_SWARM3,
-	MONSTER_SWARM4,
-	MONSTER_SCARAB1,
-	MONSTER_SCARAB2,
-	MONSTER_SCARAB3,
-	MONSTER_SCARAB4,
-	MONSTER_SCARAB5,
-	MONSTER_MUMMY1,
-	MONSTER_MUMMY2,
-	MONSTER_MUMMY3,
-	MONSTER_MUMMY4,
-	MONSTER_MUMMY5,
-	MONSTER_UNRAVELER1,
-	MONSTER_UNRAVELER2,
-	MONSTER_UNRAVELER3,
-	MONSTER_UNRAVELER4,
-	MONSTER_UNRAVELER5,
-	MONSTER_CHAOSHORDE1,
-	MONSTER_CHAOSHORDE2,
-	MONSTER_CHAOSHORDE3,
-	MONSTER_CHAOSHORDE4,
-	MONSTER_VULTURE1,
-	MONSTER_VULTURE2,
-	MONSTER_VULTURE3,
-	MONSTER_VULTURE4,
-	MONSTER_MOSQUITO1,
-	MONSTER_MOSQUITO2,
-	MONSTER_MOSQUITO3,
-	MONSTER_MOSQUITO4,
-	MONSTER_WILLOWISP1,
-	MONSTER_WILLOWISP2,
-	MONSTER_WILLOWISP3,
-	MONSTER_WILLOWISP4,
-	MONSTER_ARACH1,
-	MONSTER_ARACH2,
-	MONSTER_ARACH3,
-	MONSTER_ARACH4,
-	MONSTER_ARACH5,
-	MONSTER_THORNHULK1,
-	MONSTER_THORNHULK2,
-	MONSTER_THORNHULK3,
-	MONSTER_THORNHULK4,
-	MONSTER_VAMPIRE1,
-	MONSTER_VAMPIRE2,
-	MONSTER_VAMPIRE3,
-	MONSTER_VAMPIRE4,
-	MONSTER_VAMPIRE5,
-	MONSTER_BATDEMON1,
-	MONSTER_BATDEMON2,
-	MONSTER_BATDEMON3,
-	MONSTER_BATDEMON4,
-	MONSTER_BATDEMON5,
-	MONSTER_FETISH1,
-	MONSTER_FETISH2,
-	MONSTER_FETISH3,
-	MONSTER_FETISH4,
-	MONSTER_FETISH5,
-	MONSTER_CAIN1,
-	MONSTER_GHEED,
-	MONSTER_AKARA,
-	MONSTER_CHICKEN,
-	MONSTER_KASHYA,
-	MONSTER_RAT,
-	MONSTER_ROGUE1,
-	MONSTER_HELLMETEOR,
-	MONSTER_CHARSI,
-	MONSTER_WARRIV1,
-	MONSTER_ANDARIEL,
-	MONSTER_BIRD1,
-	MONSTER_BIRD2,
-	MONSTER_BAT,
-	MONSTER_CR_ARCHER1,
-	MONSTER_CR_ARCHER2,
-	MONSTER_CR_ARCHER3,
-	MONSTER_CR_ARCHER4,
-	MONSTER_CR_ARCHER5,
-	MONSTER_CR_LANCER1,
-	MONSTER_CR_LANCER2,
-	MONSTER_CR_LANCER3,
-	MONSTER_CR_LANCER4,
-	MONSTER_CR_LANCER5,
-	MONSTER_SK_ARCHER1,
-	MONSTER_SK_ARCHER2,
-	MONSTER_SK_ARCHER3,
-	MONSTER_SK_ARCHER4,
-	MONSTER_SK_ARCHER5,
-	MONSTER_WARRIV2,
-	MONSTER_ATMA,
-	MONSTER_DROGNAN,
-	MONSTER_FARA,
-	MONSTER_COW,
-	MONSTER_MAGGOTBABY1,
-	MONSTER_MAGGOTBABY2,
-	MONSTER_MAGGOTBABY3,
-	MONSTER_MAGGOTBABY4,
-	MONSTER_MAGGOTBABY5,
-	MONSTER_CAMEL,
-	MONSTER_BLUNDERBORE1,
-	MONSTER_BLUNDERBORE2,
-	MONSTER_BLUNDERBORE3,
-	MONSTER_BLUNDERBORE4,
-	MONSTER_MAGGOTEGG1,
-	MONSTER_MAGGOTEGG2,
-	MONSTER_MAGGOTEGG3,
-	MONSTER_MAGGOTEGG4,
-	MONSTER_MAGGOTEGG5,
-	MONSTER_ACT2MALE,
-	MONSTER_ACT2FEMALE,
-	MONSTER_ACT2CHILD,
-	MONSTER_GREIZ,
-	MONSTER_ELZIX,
-	MONSTER_GEGLASH,
-	MONSTER_JERHYN,
-	MONSTER_LYSANDER,
-	MONSTER_ACT2GUARD1,
-	MONSTER_ACT2VENDOR1,
-	MONSTER_ACT2VENDOR2,
-	MONSTER_CROWNEST1,
-	MONSTER_CROWNEST2,
-	MONSTER_CROWNEST3,
-	MONSTER_CROWNEST4,
-	MONSTER_MESHIF1,
-	MONSTER_DURIEL,
-	MONSTER_BONEFETISH1,
-	MONSTER_BONEFETISH2,
-	MONSTER_BONEFETISH3,
-	MONSTER_BONEFETISH4,
-	MONSTER_BONEFETISH5,
-	MONSTER_DARKGUARD1,
-	MONSTER_DARKGUARD2,
-	MONSTER_DARKGUARD3,
-	MONSTER_DARKGUARD4,
-	MONSTER_DARKGUARD5,
-	MONSTER_BLOODMAGE1,
-	MONSTER_BLOODMAGE2,
-	MONSTER_BLOODMAGE3,
-	MONSTER_BLOODMAGE4,
-	MONSTER_BLOODMAGE5,
-	MONSTER_MAGGOT,
-	MONSTER_SARCOPHAGUS,
-	MONSTER_RADAMENT,
-	MONSTER_FIREBEAST,
-	MONSTER_ICEGLOBE,
-	MONSTER_LIGHTNINGBEAST,
-	MONSTER_POISONORB,
-	MONSTER_FLYINGSCIMITAR,
-	MONSTER_ZEALOT1,
-	MONSTER_ZEALOT2,
-	MONSTER_ZEALOT3,
-	MONSTER_CANTOR1,
-	MONSTER_CANTOR2,
-	MONSTER_CANTOR3,
-	MONSTER_CANTOR4,
-	MONSTER_MEPHISTO,
-	MONSTER_DIABLO,
-	MONSTER_CAIN2,
-	MONSTER_CAIN3,
-	MONSTER_CAIN4,
-	MONSTER_FROGDEMON1,
-	MONSTER_FROGDEMON2,
-	MONSTER_FROGDEMON3,
-	MONSTER_SUMMONER,
-	MONSTER_TYRAEL1,
-	MONSTER_ASHEARA,
-	MONSTER_HRATLI,
-	MONSTER_ALKOR,
-	MONSTER_ORMUS,
-	MONSTER_IZUAL,
-	MONSTER_HALBU,
-	MONSTER_TENTACLE1,
-	MONSTER_TENTACLE2,
-	MONSTER_TENTACLE3,
-	MONSTER_TENTACLEHEAD1,
-	MONSTER_TENTACLEHEAD2,
-	MONSTER_TENTACLEHEAD3,
-	MONSTER_MESHIF2,
-	MONSTER_CAIN5,
-	MONSTER_NAVI,
-	MONSTER_BLOODRAVEN,
-	MONSTER_BUG,
-	MONSTER_SCORPION,
-	MONSTER_ROGUE2,
-	MONSTER_ROGUEHIRE,
-	MONSTER_ROGUE3,
-	MONSTER_GARGOYLETRAP,
-	MONSTER_SKMAGE_POIS1,
-	MONSTER_SKMAGE_POIS2,
-	MONSTER_SKMAGE_POIS3,
-	MONSTER_SKMAGE_POIS4,
-	MONSTER_FETISHSHAMAN1,
-	MONSTER_FETISHSHAMAN2,
-	MONSTER_FETISHSHAMAN3,
-	MONSTER_FETISHSHAMAN4,
-	MONSTER_FETISHSHAMAN5,
-	MONSTER_LARVA,
-	MONSTER_MAGGOTQUEEN1,
-	MONSTER_MAGGOTQUEEN2,
-	MONSTER_MAGGOTQUEEN3,
-	MONSTER_MAGGOTQUEEN4,
-	MONSTER_MAGGOTQUEEN5,
-	MONSTER_CLAYGOLEM,
-	MONSTER_BLOODGOLEM,
-	MONSTER_IRONGOLEM,
-	MONSTER_FIREGOLEM,
-	MONSTER_FAMILIAR,
-	MONSTER_ACT3MALE,
-	MONSTER_BABOON6,
-	MONSTER_ACT3FEMALE,
-	MONSTER_NATALYA,
-	MONSTER_VILEMOTHER1,
-	MONSTER_VILEMOTHER2,
-	MONSTER_VILEMOTHER3,
-	MONSTER_VILECHILD1,
-	MONSTER_VILECHILD2,
-	MONSTER_VILECHILD3,
-	MONSTER_FINGERMAGE1,
-	MONSTER_FINGERMAGE2,
-	MONSTER_FINGERMAGE3,
-	MONSTER_REGURGITATOR1,
-	MONSTER_REGURGITATOR2,
-	MONSTER_REGURGITATOR3,
-	MONSTER_DOOMKNIGHT1,
-	MONSTER_DOOMKNIGHT2,
-	MONSTER_DOOMKNIGHT3,
-	MONSTER_QUILLBEAR1,
-	MONSTER_QUILLBEAR2,
-	MONSTER_QUILLBEAR3,
-	MONSTER_QUILLBEAR4,
-	MONSTER_QUILLBEAR5,
-	MONSTER_SNAKE,
-	MONSTER_PARROT,
-	MONSTER_FISH,
-	MONSTER_EVILHOLE1,
-	MONSTER_EVILHOLE2,
-	MONSTER_EVILHOLE3,
-	MONSTER_EVILHOLE4,
-	MONSTER_EVILHOLE5,
-	MONSTER_TRAP_FIREBOLT,
-	MONSTER_TRAP_HORZMISSILE,
-	MONSTER_TRAP_VERTMISSILE,
-	MONSTER_TRAP_POISONCLOUD,
-	MONSTER_TRAP_LIGHTNING,
-	MONSTER_ACT2GUARD2,
-	MONSTER_INVISOSPAWNER,
-	MONSTER_DIABLOCLONE,
-	MONSTER_SUCKERNEST1,
-	MONSTER_SUCKERNEST2,
-	MONSTER_SUCKERNEST3,
-	MONSTER_SUCKERNEST4,
-	MONSTER_ACT2HIRE,
-	MONSTER_MINISPIDER,
-	MONSTER_BONEPRISON1,
-	MONSTER_BONEPRISON2,
-	MONSTER_BONEPRISON3,
-	MONSTER_BONEPRISON4,
-	MONSTER_BONEWALL,
-	MONSTER_COUNCILMEMBER1,
-	MONSTER_COUNCILMEMBER2,
-	MONSTER_COUNCILMEMBER3,
-	MONSTER_TURRET1,
-	MONSTER_TURRET2,
-	MONSTER_TURRET3,
-	MONSTER_HYDRA1,
-	MONSTER_HYDRA2,
-	MONSTER_HYDRA3,
-	MONSTER_TRAP_MELEE,
-	MONSTER_SEVENTOMBS,
-	MONSTER_DOPPLEZON,
-	MONSTER_VALKYRIE,
-	MONSTER_ACT2GUARD3,
-	MONSTER_ACT3HIRE,
-	MONSTER_MEGADEMON1,
-	MONSTER_MEGADEMON2,
-	MONSTER_MEGADEMON3,
-	MONSTER_NECROSKELETON,
-	MONSTER_NECROMAGE,
-	MONSTER_GRISWOLD,
-	MONSTER_COMPELLINGORB,
-	MONSTER_TYRAEL2,
-	MONSTER_DARKWANDERER,
-	MONSTER_TRAP_NOVA,
-	MONSTER_SPIRITMUMMY,
-	MONSTER_LIGHTNINGSPIRE,
-	MONSTER_FIRETOWER,
-	MONSTER_SLINGER1,
-	MONSTER_SLINGER2,
-	MONSTER_SLINGER3,
-	MONSTER_SLINGER4,
-	MONSTER_ACT2GUARD4,
-	MONSTER_ACT2GUARD5,
-	MONSTER_SKMAGE_COLD1,
-	MONSTER_SKMAGE_COLD2,
-	MONSTER_SKMAGE_COLD3,
-	MONSTER_SKMAGE_COLD4,
-	MONSTER_SKMAGE_FIRE1,
-	MONSTER_SKMAGE_FIRE2,
-	MONSTER_SKMAGE_FIRE3,
-	MONSTER_SKMAGE_FIRE4,
-	MONSTER_SKMAGE_LTNG1,
-	MONSTER_SKMAGE_LTNG2,
-	MONSTER_SKMAGE_LTNG3,
-	MONSTER_SKMAGE_LTNG4,
-	MONSTER_HELLBOVINE,
-	MONSTER_WINDOW1,
-	MONSTER_WINDOW2,
-	MONSTER_SLINGER5,
-	MONSTER_SLINGER6,
-	MONSTER_FETISHBLOW1,
-	MONSTER_FETISHBLOW2,
-	MONSTER_FETISHBLOW3,
-	MONSTER_FETISHBLOW4,
-	MONSTER_FETISHBLOW5,
-	MONSTER_MEPHISTOSPIRIT,
-	MONSTER_SMITH,
-	MONSTER_TRAPPEDSOUL1,
-	MONSTER_TRAPPEDSOUL2,
-	MONSTER_JAMELLA,
-	MONSTER_IZUALGHOST,
-	MONSTER_FETISH11,
-	MONSTER_MALACHAI,
-	MONSTER_HEPHASTO,
-	MONSTER_WAKEOFDESTRUCTION,
-	MONSTER_CHARGEBOLTSENTRY,
-	MONSTER_LIGHTNINGSENTRY,
-	MONSTER_BLADECREEPER,
-	MONSTER_INVISOPET,
-	MONSTER_INFERNOSENTRY,
-	MONSTER_DEATHSENTRY,
-	MONSTER_SHADOWWARRIOR,
-	MONSTER_SHADOWMASTER,
-	MONSTER_DRUIDHAWK,
-	MONSTER_SPIRITWOLF,
-	MONSTER_FENRIS,
-	MONSTER_SPIRITOFBARBS,
-	MONSTER_HEARTOFWOLVERINE,
-	MONSTER_OAKSAGE,
-	MONSTER_PLAGUEPOPPY,
-	MONSTER_CYCLEOFLIFE,
-	MONSTER_VINECREATURE,
-	MONSTER_DRUIDBEAR,
-	MONSTER_EAGLE,
-	MONSTER_WOLF,
-	MONSTER_BEAR,
-	MONSTER_BARRICADEDOOR1,
-	MONSTER_BARRICADEDOOR2,
-	MONSTER_PRISONDOOR,
-	MONSTER_BARRICADETOWER,
-	MONSTER_REANIMATEDHORDE1,
-	MONSTER_REANIMATEDHORDE2,
-	MONSTER_REANIMATEDHORDE3,
-	MONSTER_REANIMATEDHORDE4,
-	MONSTER_REANIMATEDHORDE5,
-	MONSTER_SIEGEBEAST1,
-	MONSTER_SIEGEBEAST2,
-	MONSTER_SIEGEBEAST3,
-	MONSTER_SIEGEBEAST4,
-	MONSTER_SIEGEBEAST5,
-	MONSTER_SNOWYETI1,
-	MONSTER_SNOWYETI2,
-	MONSTER_SNOWYETI3,
-	MONSTER_SNOWYETI4,
-	MONSTER_WOLFRIDER1,
-	MONSTER_WOLFRIDER2,
-	MONSTER_WOLFRIDER3,
-	MONSTER_MINION1,
-	MONSTER_MINION2,
-	MONSTER_MINION3,
-	MONSTER_MINION4,
-	MONSTER_MINION5,
-	MONSTER_MINION6,
-	MONSTER_MINION7,
-	MONSTER_MINION8,
-	MONSTER_SUICIDEMINION1,
-	MONSTER_SUICIDEMINION2,
-	MONSTER_SUICIDEMINION3,
-	MONSTER_SUICIDEMINION4,
-	MONSTER_SUICIDEMINION5,
-	MONSTER_SUICIDEMINION6,
-	MONSTER_SUICIDEMINION7,
-	MONSTER_SUICIDEMINION8,
-	MONSTER_SUCCUBUS1,
-	MONSTER_SUCCUBUS2,
-	MONSTER_SUCCUBUS3,
-	MONSTER_SUCCUBUS4,
-	MONSTER_SUCCUBUS5,
-	MONSTER_SUCCUBUSWITCH1,
-	MONSTER_SUCCUBUSWITCH2,
-	MONSTER_SUCCUBUSWITCH3,
-	MONSTER_SUCCUBUSWITCH4,
-	MONSTER_SUCCUBUSWITCH5,
-	MONSTER_OVERSEER1,
-	MONSTER_OVERSEER2,
-	MONSTER_OVERSEER3,
-	MONSTER_OVERSEER4,
-	MONSTER_OVERSEER5,
-	MONSTER_MINIONSPAWNER1,
-	MONSTER_MINIONSPAWNER2,
-	MONSTER_MINIONSPAWNER3,
-	MONSTER_MINIONSPAWNER4,
-	MONSTER_MINIONSPAWNER5,
-	MONSTER_MINIONSPAWNER6,
-	MONSTER_MINIONSPAWNER7,
-	MONSTER_MINIONSPAWNER8,
-	MONSTER_IMP1,
-	MONSTER_IMP2,
-	MONSTER_IMP3,
-	MONSTER_IMP4,
-	MONSTER_IMP5,
-	MONSTER_CATAPULT1,
-	MONSTER_CATAPULT2,
-	MONSTER_CATAPULT3,
-	MONSTER_CATAPULT4,
-	MONSTER_FROZENHORROR1,
-	MONSTER_FROZENHORROR2,
-	MONSTER_FROZENHORROR3,
-	MONSTER_FROZENHORROR4,
-	MONSTER_FROZENHORROR5,
-	MONSTER_BLOODLORD1,
-	MONSTER_BLOODLORD2,
-	MONSTER_BLOODLORD3,
-	MONSTER_BLOODLORD4,
-	MONSTER_BLOODLORD5,
-	MONSTER_LARZUK,
-	MONSTER_DREHYA,
-	MONSTER_MALAH,
-	MONSTER_NIHLATHAK,
-	MONSTER_QUAL_KEHK,
-	MONSTER_CATAPULTSPOTTER1,
-	MONSTER_CATAPULTSPOTTER2,
-	MONSTER_CATAPULTSPOTTER3,
-	MONSTER_CATAPULTSPOTTER4,
-	MONSTER_CAIN6,
-	MONSTER_TYRAEL3,
-	MONSTER_ACT5BARB1,
-	MONSTER_ACT5BARB2,
-	MONSTER_BARRICADEWALL1,
-	MONSTER_BARRICADEWALL2,
-	MONSTER_NIHLATHAKBOSS,
-	MONSTER_DREHYAICED,
-	MONSTER_EVILHUT,
-	MONSTER_DEATHMAULER1,
-	MONSTER_DEATHMAULER2,
-	MONSTER_DEATHMAULER3,
-	MONSTER_DEATHMAULER4,
-	MONSTER_DEATHMAULER5,
-	MONSTER_ACT5POW,
-	MONSTER_ACT5BARB3,
-	MONSTER_ACT5BARB4,
-	MONSTER_ANCIENTSTATUE1,
-	MONSTER_ANCIENTSTATUE2,
-	MONSTER_ANCIENTSTATUE3,
-	MONSTER_ANCIENTBARB1,
-	MONSTER_ANCIENTBARB2,
-	MONSTER_ANCIENTBARB3,
-	MONSTER_BAALTHRONE,
-	MONSTER_BAALCRAB,
-	MONSTER_BAALTAUNT,
-	MONSTER_PUTRIDDEFILER1,
-	MONSTER_PUTRIDDEFILER2,
-	MONSTER_PUTRIDDEFILER3,
-	MONSTER_PUTRIDDEFILER4,
-	MONSTER_PUTRIDDEFILER5,
-	MONSTER_PAINWORM1,
-	MONSTER_PAINWORM2,
-	MONSTER_PAINWORM3,
-	MONSTER_PAINWORM4,
-	MONSTER_PAINWORM5,
-	MONSTER_BUNNY,
-	MONSTER_BAALHIGHPRIEST,
-	MONSTER_VENOMLORD,
-	MONSTER_BAALCRABSTAIRS,
-	MONSTER_ACT5HIRE1,
-	MONSTER_ACT5HIRE2,
-	MONSTER_BAALTENTACLE1,
-	MONSTER_BAALTENTACLE2,
-	MONSTER_BAALTENTACLE3,
-	MONSTER_BAALTENTACLE4,
-	MONSTER_BAALTENTACLE5,
-	MONSTER_INJUREDBARB1,
-	MONSTER_INJUREDBARB2,
-	MONSTER_INJUREDBARB3,
-	MONSTER_BAALCLONE,
-	MONSTER_BAALMINION1,
-	MONSTER_BAALMINION2,
-	MONSTER_BAALMINION3,
-	MONSTER_WORLDSTONEEFFECT,
-	MONSTER_SK_ARCHER6,
-	MONSTER_SK_ARCHER7,
-	MONSTER_SK_ARCHER8,
-	MONSTER_SK_ARCHER9,
-	MONSTER_SK_ARCHER10,
-	MONSTER_BIGHEAD6,
-	MONSTER_BIGHEAD7,
-	MONSTER_BIGHEAD8,
-	MONSTER_BIGHEAD9,
-	MONSTER_BIGHEAD10,
-	MONSTER_GOATMAN6,
-	MONSTER_GOATMAN7,
-	MONSTER_GOATMAN8,
-	MONSTER_GOATMAN9,
-	MONSTER_GOATMAN10,
-	MONSTER_FOULCROW5,
-	MONSTER_FOULCROW6,
-	MONSTER_FOULCROW7,
-	MONSTER_FOULCROW8,
-	MONSTER_CLAWVIPER6,
-	MONSTER_CLAWVIPER7,
-	MONSTER_CLAWVIPER8,
-	MONSTER_CLAWVIPER9,
-	MONSTER_CLAWVIPER10,
-	MONSTER_SANDRAIDER6,
-	MONSTER_SANDRAIDER7,
-	MONSTER_SANDRAIDER8,
-	MONSTER_SANDRAIDER9,
-	MONSTER_SANDRAIDER10,
-	MONSTER_DEATHMAULER6,
-	MONSTER_QUILLRAT6,
-	MONSTER_QUILLRAT7,
-	MONSTER_QUILLRAT8,
-	MONSTER_VULTURE5,
-	MONSTER_THORNHULK5,
-	MONSTER_SLINGER7,
-	MONSTER_SLINGER8,
-	MONSTER_SLINGER9,
-	MONSTER_CR_ARCHER6,
-	MONSTER_CR_ARCHER7,
-	MONSTER_CR_LANCER6,
-	MONSTER_CR_LANCER7,
-	MONSTER_CR_LANCER8,
-	MONSTER_BLUNDERBORE5,
-	MONSTER_BLUNDERBORE6,
-	MONSTER_SKMAGE_FIRE5,
-	MONSTER_SKMAGE_FIRE6,
-	MONSTER_SKMAGE_LTNG5,
-	MONSTER_SKMAGE_LTNG6,
-	MONSTER_SKMAGE_COLD5,
-	MONSTER_SKMAGE_POIS5,
-	MONSTER_SKMAGE_POIS6,
-	MONSTER_PANTHERWOMAN5,
-	MONSTER_PANTHERWOMAN6,
-	MONSTER_SANDLEAPER6,
-	MONSTER_SANDLEAPER7,
-	MONSTER_WRAITH6,
-	MONSTER_WRAITH7,
-	MONSTER_WRAITH8,
-	MONSTER_SUCCUBUS6,
-	MONSTER_SUCCUBUS7,
-	MONSTER_SUCCUBUSWITCH6,
-	MONSTER_SUCCUBUSWITCH7,
-	MONSTER_SUCCUBUSWITCH8,
-	MONSTER_WILLOWISP5,
-	MONSTER_WILLOWISP6,
-	MONSTER_WILLOWISP7,
-	MONSTER_FALLEN6,
-	MONSTER_FALLEN7,
-	MONSTER_FALLEN8,
-	MONSTER_FALLENSHAMAN6,
-	MONSTER_FALLENSHAMAN7,
-	MONSTER_FALLENSHAMAN8,
-	MONSTER_SKELETON6,
-	MONSTER_SKELETON7,
-	MONSTER_BATDEMON6,
-	MONSTER_BATDEMON7,
-	MONSTER_BLOODLORD6,
-	MONSTER_BLOODLORD7,
-	MONSTER_SCARAB6,
-	MONSTER_SCARAB7,
-	MONSTER_FETISH6,
-	MONSTER_FETISH7,
-	MONSTER_FETISH8,
-	MONSTER_FETISHBLOW6,
-	MONSTER_FETISHBLOW7,
-	MONSTER_FETISHBLOW8,
-	MONSTER_FETISHSHAMAN6,
-	MONSTER_FETISHSHAMAN7,
-	MONSTER_FETISHSHAMAN8,
-	MONSTER_BABOON7,
-	MONSTER_BABOON8,
-	MONSTER_UNRAVELER6,
-	MONSTER_UNRAVELER7,
-	MONSTER_UNRAVELER8,
-	MONSTER_UNRAVELER9,
-	MONSTER_ZEALOT4,
-	MONSTER_ZEALOT5,
-	MONSTER_CANTOR5,
-	MONSTER_CANTOR6,
-	MONSTER_VILEMOTHER4,
-	MONSTER_VILEMOTHER5,
-	MONSTER_VILECHILD4,
-	MONSTER_VILECHILD5,
-	MONSTER_SANDMAGGOT6,
-	MONSTER_MAGGOTBABY6,
-	MONSTER_MAGGOTEGG6,
-	MONSTER_MINION9,
-	MONSTER_MINION10,
-	MONSTER_MINION11,
-	MONSTER_ARACH6,
-	MONSTER_MEGADEMON4,
-	MONSTER_MEGADEMON5,
-	MONSTER_IMP6,
-	MONSTER_IMP7,
-	MONSTER_BONEFETISH6,
-	MONSTER_BONEFETISH7,
-	MONSTER_FINGERMAGE4,
-	MONSTER_FINGERMAGE5,
-	MONSTER_REGURGITATOR4,
-	MONSTER_VAMPIRE6,
-	MONSTER_VAMPIRE7,
-	MONSTER_VAMPIRE8,
-	MONSTER_REANIMATEDHORDE6,
-	MONSTER_DKFIG1,
-	MONSTER_DKFIG2,
-	MONSTER_DKMAG1,
-	MONSTER_DKMAG2,
-	MONSTER_MUMMY6,
-	MONSTER_UBERMEPHISTO,
-	MONSTER_UBERDIABLO,
-	MONSTER_UBERIZUAL,
-	MONSTER_UBERANDARIEL,
-	MONSTER_UBERDURIEL,
-	MONSTER_UBERBAAL,
-	MONSTER_DEMONSPAWNER,
-	MONSTER_DEMONHOLE,
-	MONSTER_MEGADEMON6,
-	MONSTER_DKMAG3,
-	MONSTER_IMP8,
-	MONSTER_SWARM5,
-	MONSTER_SANDMAGGOT7,
-	MONSTER_ARACH7,
-	MONSTER_SCARAB8,
-	MONSTER_SUCCUBUS8,
-	MONSTER_SUCCUBUSWITCH9,
-	MONSTER_CORRUPTROGUE6,
-	MONSTER_CR_ARCHER8,
-	MONSTER_CR_LANCER9,
-	MONSTER_OVERSEER6,
-	MONSTER_SKELETON8,
-	MONSTER_SK_ARCHER11,
-	MONSTER_SKMAGE_FIRE7,
-	MONSTER_SKMAGE_LTNG7,
-	MONSTER_SKMAGE_COLD6,
-	MONSTER_SKMAGE_POIS7,
-	MONSTER_VAMPIRE9,
-	MONSTER_WRAITH9,
-	MONSTER_WILLOWISP8,
-};
-
 enum D2C_MonTypeFlags
 {
 	MONTYPEFLAG_OTHER = 0x01,
@@ -2672,75 +367,6 @@ enum D2C_MonUMods
 	MONUMOD_NOVADEATH				//nova on death
 };
 
-enum D2C_SuperUniques
-{
-	SUPERUNIQUE_BISHIBOSH,
-	SUPERUNIQUE_BONEBREAK,
-	SUPERUNIQUE_COLDCROW,
-	SUPERUNIQUE_RAKANISHU,
-	SUPERUNIQUE_TREEHEAD_WOODFIST,
-	SUPERUNIQUE_GRISWOLD,
-	SUPERUNIQUE_THE_COUNTESS,
-	SUPERUNIQUE_PITSPAWN_FOULDOG,
-	SUPERUNIQUE_FLAMESPIKE_THE_CRAWLER,
-	SUPERUNIQUE_BONEASH,
-	SUPERUNIQUE_RADAMENT,
-	SUPERUNIQUE_BLOODWITCH_THE_WILD,
-	SUPERUNIQUE_FANGSKIN,
-	SUPERUNIQUE_BEETLEBURST,
-	SUPERUNIQUE_LEATHERARM,
-	SUPERUNIQUE_COLDWORM_THE_BURROWER,
-	SUPERUNIQUE_FIRE_EYE,
-	SUPERUNIQUE_DARK_ELDER,
-	SUPERUNIQUE_THE_SUMMONER,
-	SUPERUNIQUE_ANCIENT_KAA_THE_SOULLESS,
-	SUPERUNIQUE_THE_SMITH,
-	SUPERUNIQUE_WEB_MAGE_THE_BURNING,
-	SUPERUNIQUE_WITCH_DOCTOR_ENDUGU,
-	SUPERUNIQUE_STORMTREE,
-	SUPERUNIQUE_SARINA_THE_BATTLEMAID,
-	SUPERUNIQUE_ICEHAWK_RIFTWING,
-	SUPERUNIQUE_ISMAIL_VILEHAND,
-	SUPERUNIQUE_GELEB_FLAMEFINGER,
-	SUPERUNIQUE_BREMM_SPARKFIST,
-	SUPERUNIQUE_TOORC_ICEFIST,
-	SUPERUNIQUE_WYAND_VOIDFINGER,
-	SUPERUNIQUE_MAFFER_DRAGONHAND,
-	SUPERUNIQUE_WINGED_DEATH,
-	SUPERUNIQUE_THE_TORMENTOR,
-	SUPERUNIQUE_TAINTBREEDER,
-	SUPERUNIQUE_RIFTWRAITH_THE_CANNIBAL,
-	SUPERUNIQUE_INFECTOR_OF_SOULS,
-	SUPERUNIQUE_LORD_DE_SEIS,
-	SUPERUNIQUE_GRAND_VIZIER_OF_CHAOS,
-	SUPERUNIQUE_THE_COW_KING,
-	SUPERUNIQUE_CORPSEFIRE,
-	SUPERUNIQUE_THE_FEATURE_CREEP,
-	SUPERUNIQUE_SIEGE_BOSS,
-	SUPERUNIQUE_ANCIENT_BARBARIAN_1,
-	SUPERUNIQUE_ANCIENT_BARBARIAN_2,
-	SUPERUNIQUE_ANCIENT_BARBARIAN_3,
-	SUPERUNIQUE_AXE_DWELLER,
-	SUPERUNIQUE_BONESAW_BREAKER,
-	SUPERUNIQUE_DAC_FARREN,
-	SUPERUNIQUE_MEGAFLOW_RECTIFIER,
-	SUPERUNIQUE_EYEBACK_UNLEASHED,
-	SUPERUNIQUE_THREASH_SOCKET,
-	SUPERUNIQUE_PINDLESKIN,
-	SUPERUNIQUE_SNAPCHIP_SHATTER,
-	SUPERUNIQUE_ANODIZED_ELITE,
-	SUPERUNIQUE_VINVEAR_MOLECH,
-	SUPERUNIQUE_SHARP_TOOTH_SAYER,
-	SUPERUNIQUE_MAGMA_TORQUER,
-	SUPERUNIQUE_BLAZE_RIPPER,
-	SUPERUNIQUE_FROZENSTEIN,
-	SUPERUNIQUE_NIHLATHAK_BOSS,
-	SUPERUNIQUE_BAAL_SUBJECT_1,
-	SUPERUNIQUE_BAAL_SUBJECT_2,
-	SUPERUNIQUE_BAAL_SUBJECT_3,
-	SUPERUNIQUE_BAAL_SUBJECT_4,
-	SUPERUNIQUE_BAAL_SUBJECT_5,
-};
 
 enum D2MapAIActions
 {
@@ -2751,152 +377,6 @@ enum D2MapAIActions
 #endif // CONSTANTS_MONSTERS
 
 #ifndef CONSTANTS_OBJECTS //Pseudo-Macro to group all constants
-enum D2C_ObjModes
-{
-	OBJMODE_NEUTRAL,	// NU
-	OBJMODE_OPERATING,	// OP
-	OBJMODE_OPENED,		// ON
-	OBJMODE_SPECIAL1,	// S1
-	OBJMODE_SPECIAL2,	// S2
-	OBJMODE_SPECIAL3,	// S3
-	OBJMODE_SPECIAL4,	// S4
-	OBJMODE_SPECIAL5	// S5
-};
-
-enum D2Objects
-{
-	OBJECT_TOWERTOME = 8,
-	OBJECT_ROGUEFOUNTAIN = 12,
-	OBJECT_DOORGATELEFT = 13,
-	OBJECT_DOORGATERIGHT = 14,
-	OBJECT_DOORWOODENLEFT = 15,
-	OBJECT_DOORWOODENRIGHT = 16,
-	OBJECT_STONEALPHA = 17,
-	OBJECT_STONEBETA = 18,
-	OBJECT_STONEGAMMA = 19,
-	OBJECT_STONEDELTA = 20,
-	OBJECT_STONELAMBDA = 21,
-	OBJECT_STONETHETA = 22,
-	OBJECT_DOORCOURTYARDLEFT = 23,
-	OBJECT_DOORCOURTYARDRIGHT = 24,
-	OBJECT_DOORCATHEDRALDOUBLE = 25,
-	OBJECT_CAINGIBBET = 26,
-	OBJECT_DOORMONASTERYDOUBLERIGHT = 27,
-	OBJECT_INIFUSS_TREE = 30,
-	OBJECT_ROGUEBONFIRE = 39,
-	OBJECT_RIVER1 = 40,
-	OBJECT_RIVER2 = 41,
-	OBJECT_RIVER3 = 42,
-	OBJECT_RIVER4 = 43,
-	OBJECT_RIVER5 = 44,
-	OBJECT_ANDARIELS_DOOR = 47,
-	OBJECT_TOWN_PORTAL = 59,
-	OBJECT_PERMANENT_TOWN_PORTAL = 60,
-	OBJECT_INVISIBLE_OBJECT = 61,
-	OBJECT_DOORCATHEDRALLEFT = 62,
-	OBJECT_DOORCATHEDRALRIGHT = 63,
-	OBJECT_DOORWOODENLEFT2 = 64,
-	OBJECT_INVISIBLE_RIVER_SOUND1 = 65,
-	OBJECT_INVISIBLE_RIVER_SOUND2 = 66,
-	OBJECT_TRAP_DOOR = 74,
-	OBJECT_OBELISK = 80,
-	OBJECT_DESERT_OBELISK = 90,
-	OBJECT_TOMBDOORLEFT = 91,
-	OBJECT_TOMBDOORRIGHT = 92,
-	OBJECT_PORTAL_TO_DURIELS_LAIR = 100,
-	OBJECT_MALUS = 108,
-	OBJECT_WAYPOINT_PORTAL = 119,
-	OBJECT_JERHYN_PLACEHOLDER1 = 121,
-	OBJECT_JERHYN_PLACEHOLDER2 = 122,
-	OBJECT_VILE_DOG_AFTERGLOW = 131,
-	OBJECT_TAINTEDSUNALTAR = 149,
-	OBJECT_ORIFICE = 152,
-	OBJECT_TYRAELS_DOOR = 153,
-	OBJECT_FIRE_SMALL = 160,
-	OBJECT_FIRE_MEDIUM = 161,
-	OBJECT_FIRE_LARGE = 162,
-	OBJECT_CAINPORTAL = 189,
-	OBJECT_TELEPORTATION_PAD = 192,
-	OBJECT_LAM_ESENS_TOME = 193,
-	OBJECT_FLOORTRAP = 196,
-	OBJECT_SLIMEDOOR1 = 229,
-	OBJECT_SLIMEDOOR2 = 230,
-	OBJECT_ACT3WAYPOINT_TOWN = 237,
-	OBJECT_EXPLODING_COW = 250,
-	OBJECT_GIDBINN_ALTAR = 251,
-	OBJECT_GIDBINN_DECOY = 252,
-	OBJECT_TOMBFLOORTRAP = 261,
-	OBJECT_STASH = 267,
-	OBJECT_WIRTSBODY = 268,
-	OBJECT_GOLD_PLACEHOLDER = 269,
-	OBJECT_IRONGRATEDOORLEFT = 290,
-	OBJECT_IRONGRATEDOORRIGHT = 291,
-	OBJECT_WOODENGRATEDOORLEFT = 292,
-	OBJECT_WOODENGRATEDOORRIGHT = 293,
-	OBJECT_WOODENDOORLEFT = 294,
-	OBJECT_WOODENDOORRIGHT = 295,
-	OBJECT_ARCANE_SANCTUARY_PORTAL = 298,
-	OBJECT_HAREM_GUARD_1 = 314,
-	OBJECT_HAREM_GUARD_2 = 315,
-	OBJECT_HAREM_GUARD_3 = 316,
-	OBJECT_HAREM_GUARD_4 = 317,
-	OBJECT_EUNUCH_HAREM_BLOCKER = 318,
-	OBJECT_STEEG_STONE = 337,
-	OBJECT_GUILD_VAULT = 338,
-	OBJECT_TROPHY_CASE = 339,
-	OBJECT_MESSAGE_BOARD = 340,
-	OBJECT_MEPHISTOBRIDGE = 341,
-	OBJECT_HELLGATE_PORTAL = 342,
-	OBJECT_HORADRIC_CUBE_CHEST = 354,
-	OBJECT_HORADRIC_SCROLL_CHEST = 355,
-	OBJECT_STAFF_OF_KINGS_CHEST = 356,
-	OBJECT_YET_ANOTHER_TOME = 357,
-	OBJECT_SEWER_STAIRS = 366,
-	OBJECT_ACT3SEWERLEVER = 367,
-	OBJECT_DARKWANDERER_STARTPOSITION = 368,
-	OBJECT_CHEST = 371,
-	OBJECT_HELLFORGE = 376,
-	OBJECT_GUILD_PORTAL = 377,
-	OBJECT_HRATLI_START = 378,
-	OBJECT_HRATLI_END = 379,
-	OBJECT_NATALYA_START = 382,
-	OBJECT_CAIN_START_POSITION = 385,
-	OBJECT_DIABLO_SEAL1 = 392,
-	OBJECT_DIABLO_SEAL2 = 393,
-	OBJECT_DIABLO_SEAL3 = 394,
-	OBJECT_DIABLO_SEAL4 = 395,
-	OBJECT_DIABLO_SEAL5 = 396,
-	OBJECT_COMPELLING_ORB = 404,
-	OBJECT_KHALIM_CHEST1 = 405,
-	OBJECT_KHALIM_CHEST2 = 406,
-	OBJECT_KHALIM_CHEST3 = 407,
-	OBJECT_SIEGE_CONTROL = 410,
-	OBJECT_HARROGATH_TOWN_MAIN_GATE = 449,
-	OBJECT_DREHYA_START_IN_TOWN = 459,
-	OBJECT_DREHYA_START_OUTSIDE_TOWN = 460,
-	OBJECT_NIHLATHAK_START_IN_TOWN = 461,
-	OBJECT_NIHLATHAK_START_OUTSIDE_TOWN = 462,
-	OBJECT_CAGEDWUSSIE1 = 473,
-	OBJECT_ANCIENTSTATUE3 = 474,
-	OBJECT_ANCIENTSTATUE1 = 475,
-	OBJECT_ANCIENTSTATUE2 = 476,
-	OBJECT_LARZUK_GREETING = 542,
-	OBJECT_LARZUK_STANDARD = 543,
-	OBJECT_ANCIENTSALTAR = 546,
-	OBJECT_ANCIENTSDOOR = 547,
-	OBJECT_BAALS_LAIR = 557,
-	OBJECT_FROZEN_ANYA = 558,
-	OBJECT_BBQ_BUNNY = 559,
-	OBJECT_INVISIBLE_ANCIENT = 561,
-	OBJECT_INVISIBLE_BASE = 562,
-	OBJECT_BAALS_PORTAL_WORLDSTONECHAMBER = 563,
-	OBJECT_SUMMIT_DOOR = 564,
-	OBJECT_LAST_PORTAL = 565,
-	OBJECT_HARROGATH_LAST_LAST_PORTAL = 566,
-	OBJECT_ZOO = 567,
-	OBJECT_KEEPER = 568,
-	OBJECT_BAALS_PORTAL_THRONEOFDESTRUCTION = 569,
-};
 
 enum D2ObjectSubClasses
 {
@@ -2910,396 +390,9 @@ enum D2ObjectSubClasses
 	OBJSUBCLASS_DOOR = 0x80,
 };
 
-enum D2C_ShrineTypes
-{
-	SHRINE_NONE,
-	SHRINE_REFILL,
-	SHRINE_HPBOOST,
-	SHRINE_MANABOOST,
-	SHRINE_HPXCHANGE,
-	SHRINE_MANAXCHANGE,
-	SHRINE_ARMOR,
-	SHRINE_COMBAT,
-	SHRINE_RESFIRE,
-	SHRINE_RESCOLD,
-	SHRINE_RESLITE,
-	SHRINE_RESPOIS,
-	SHRINE_SKILL,
-	SHRINE_MANAREGEN,
-	SHRINE_STAM,
-	SHRINE_EXP,
-	SHRINE_ENIRHS,
-	SHRINE_PORTAL,
-	SHRINE_GEM,
-	SHRINE_STORM,
-	SHRINE_WARP,
-	SHRINE_EXPLOSION,
-	SHRINE_POISON
-};
 #endif // CONSTANTS_OBJECTS
 
 #ifndef CONSTANTS_SKILLS //Pseudo-Macro to group all constants
-enum D2C_Skills
-{
-	SKILL_ATTACK = 0,
-	SKILL_KICK,
-	SKILL_THROW,
-	SKILL_UNSUMMON,
-	SKILL_LEFTHANDTHROW,
-	SKILL_LEFTHANDSWING,
-	SKILL_MAGICARROW,
-	SKILL_FIREARROW,
-	SKILL_INNERSIGHT,
-	SKILL_CRITICALSTRIKE,
-	SKILL_JAB,
-	SKILL_COLDARROW,
-	SKILL_MULTIPLESHOT,
-	SKILL_DODGE,
-	SKILL_POWERSTRIKE,
-	SKILL_POISONJAVELIN,
-	SKILL_EXPLODINGARROW,
-	SKILL_SLOWMISSILES,
-	SKILL_AVOID,
-	SKILL_IMPALE,
-	SKILL_LIGHTNINGBOLT,
-	SKILL_ICEARROW,
-	SKILL_GUIDEDARROW,
-	SKILL_PENETRATE,
-	SKILL_CHARGEDSTRIKE,
-	SKILL_PLAGUEJAVELIN,
-	SKILL_STRAFE,
-	SKILL_IMMOLATIONARROW,
-	SKILL_DOPPLEZON,
-	SKILL_EVADE,
-	SKILL_FEND,
-	SKILL_FREEZINGARROW,
-	SKILL_VALKYRIE,
-	SKILL_PIERCE,
-	SKILL_LIGHTNINGSTRIKE,
-	SKILL_LIGHTNINGFURY,
-	SKILL_FIREBOLT,
-	SKILL_WARMTH,
-	SKILL_CHARGEDBOLT,
-	SKILL_ICEBOLT,
-	SKILL_FROZENARMOR,
-	SKILL_INFERNO,
-	SKILL_STATICFIELD,
-	SKILL_TELEKINESIS,
-	SKILL_FROSTNOVA,
-	SKILL_ICEBLAST,
-	SKILL_BLAZE,
-	SKILL_FIREBALL,
-	SKILL_NOVA,
-	SKILL_LIGHTNING,
-	SKILL_SHIVERARMOR,
-	SKILL_FIREWALL,
-	SKILL_ENCHANT,
-	SKILL_CHAINLIGHTNING,
-	SKILL_TELEPORT,
-	SKILL_GLACIALSPIKE,
-	SKILL_METEOR,
-	SKILL_THUNDERSTORM,
-	SKILL_ENERGYSHIELD,
-	SKILL_BLIZZARD,
-	SKILL_CHILLINGARMOR,
-	SKILL_FIREMASTERY,
-	SKILL_HYDRA,
-	SKILL_LIGHTNINGMASTERY,
-	SKILL_FROZENORB,
-	SKILL_COLDMASTERY,
-	SKILL_AMPLIFYDAMAGE,
-	SKILL_TEETH,
-	SKILL_BONEARMOR,
-	SKILL_SKELETONMASTERY,
-	SKILL_RAISESKELETON,
-	SKILL_DIMVISION,
-	SKILL_WEAKEN,
-	SKILL_POISONDAGGER,
-	SKILL_CORPSEEXPLOSION,
-	SKILL_CLAYGOLEM,
-	SKILL_IRONMAIDEN,
-	SKILL_TERROR,
-	SKILL_BONEWALL,
-	SKILL_GOLEMMASTERY,
-	SKILL_RAISESKELETALMAGE,
-	SKILL_CONFUSE,
-	SKILL_LIFETAP,
-	SKILL_POISONEXPLOSION,
-	SKILL_BONESPEAR,
-	SKILL_BLOODGOLEM,
-	SKILL_ATTRACT,
-	SKILL_DECREPIFY,
-	SKILL_BONEPRISON,
-	SKILL_SUMMONRESIST,
-	SKILL_IRONGOLEM,
-	SKILL_LOWERRESIST,
-	SKILL_POISONNOVA,
-	SKILL_BONESPIRIT,
-	SKILL_FIREGOLEM,
-	SKILL_REVIVE,
-	SKILL_SACRIFICE,
-	SKILL_SMITE,
-	SKILL_MIGHT,
-	SKILL_PRAYER,
-	SKILL_RESISTFIRE,
-	SKILL_HOLYBOLT,
-	SKILL_HOLYFIRE,
-	SKILL_THORNS,
-	SKILL_DEFIANCE,
-	SKILL_RESISTCOLD,
-	SKILL_ZEAL,
-	SKILL_CHARGE,
-	SKILL_BLESSEDAIM,
-	SKILL_CLEANSING,
-	SKILL_RESISTLIGHTNING,
-	SKILL_VENGEANCE,
-	SKILL_BLESSEDHAMMER,
-	SKILL_CONCENTRATION,
-	SKILL_HOLYFREEZE,
-	SKILL_VIGOR,
-	SKILL_CONVERSION,
-	SKILL_HOLYSHIELD,
-	SKILL_HOLYSHOCK,
-	SKILL_SANCTUARY,
-	SKILL_MEDITATION,
-	SKILL_FISTOFTHEHEAVENS,
-	SKILL_FANATICISM,
-	SKILL_CONVICTION,
-	SKILL_REDEMPTION,
-	SKILL_SALVATION,
-	SKILL_BASH,
-	SKILL_SWORDMASTERY,
-	SKILL_AXEMASTERY,
-	SKILL_MACEMASTERY,
-	SKILL_HOWL,
-	SKILL_FINDPOTION,
-	SKILL_LEAP,
-	SKILL_DOUBLESWING,
-	SKILL_POLEARMMASTERY,
-	SKILL_THROWINGMASTERY,
-	SKILL_SPEARMASTERY,
-	SKILL_TAUNT,
-	SKILL_SHOUT,
-	SKILL_STUN,
-	SKILL_DOUBLETHROW,
-	SKILL_INCREASEDSTAMINA,
-	SKILL_FINDITEM,
-	SKILL_LEAPATTACK,
-	SKILL_CONCENTRATE,
-	SKILL_IRONSKIN,
-	SKILL_BATTLECRY,
-	SKILL_FRENZY,
-	SKILL_INCREASEDSPEED,
-	SKILL_BATTLEORDERS,
-	SKILL_GRIMWARD,
-	SKILL_WHIRLWIND,
-	SKILL_BERSERK,
-	SKILL_NATURALRESISTANCE,
-	SKILL_WARCRY,
-	SKILL_BATTLECOMMAND,
-	SKILL_FIREHIT,
-	SKILL_UNHOLYBOLT,
-	SKILL_SKELETONRAISE,
-	SKILL_MAGGOTEGG,
-	SKILL_SHAMANFIRE,
-	SKILL_MAGOTTUP,
-	SKILL_MAGOTTDOWN,
-	SKILL_MAGOTTLAY,
-	SKILL_ANDRIALSPRAY,
-	SKILL_JUMP,
-	SKILL_SWARMMOVE,
-	SKILL_NEST,
-	SKILL_QUICKSTRIKE,
-	SKILL_VAMPIREFIREBALL,
-	SKILL_VAMPIREFIREWALL,
-	SKILL_VAMPIREMETEOR,
-	SKILL_GARGOYLETRAP,
-	SKILL_SPIDERLAY,
-	SKILL_VAMPIREHEAL,
-	SKILL_VAMPIRERAISE,
-	SKILL_SUBMERGE,
-	SKILL_FETISHAURA,
-	SKILL_FETISHINFERNO,
-	SKILL_ZAKARUMHEAL,
-	SKILL_EMERGE,
-	SKILL_RESURRECT,
-	SKILL_BESTOW,
-	SKILL_MISSILESKILL1,
-	SKILL_MONTELEPORT,
-	SKILL_PRIMELIGHTNING,
-	SKILL_PRIMEBOLT,
-	SKILL_PRIMEBLAZE,
-	SKILL_PRIMEFIREWALL,
-	SKILL_PRIMESPIKE,
-	SKILL_PRIMEICENOVA,
-	SKILL_PRIMEPOISONBALL,
-	SKILL_PRIMEPOISONNOVA,
-	SKILL_DIABLIGHT,
-	SKILL_DIABCOLD,
-	SKILL_DIABFIRE,
-	SKILL_FINGERMAGESPIDER,
-	SKILL_DIABWALL,
-	SKILL_DIABRUN,
-	SKILL_DIABPRISON,
-	SKILL_POISONBALLTRAP,
-	SKILL_ANDYPOISONBOLT,
-	SKILL_HIREABLEMISSILE,
-	SKILL_DESERTTURRET,
-	SKILL_ARCANETOWER,
-	SKILL_MONBLIZZARD,
-	SKILL_MOSQUITO,
-	SKILL_CURSEDBALLTRAPRIGHT,
-	SKILL_CURSEDBALLTRAPLEFT,
-	SKILL_MONFROZENARMOR,
-	SKILL_MONBONEARMOR,
-	SKILL_MONBONESPIRIT,
-	SKILL_MONCURSECAST,
-	SKILL_HELLMETEOR,
-	SKILL_REGURGITATOREAT,
-	SKILL_MONFRENZY,
-	SKILL_QUEENDEATH,
-	SKILL_SCROLLOFIDENTIFY,
-	SKILL_BOOKOFIDENTIFY,
-	SKILL_SCROLLOFTOWNPORTAL,
-	SKILL_BOOKOFTOWNPORTAL,
-	SKILL_RAVEN,
-	SKILL_PLAGUEPOPPY,
-	SKILL_WEARWOLF,
-	SKILL_SHAPESHIFTING,
-	SKILL_FIRESTORM,
-	SKILL_OAKSAGE,
-	SKILL_SUMMONSPIRITWOLF,
-	SKILL_WEARBEAR,
-	SKILL_MOLTENBOULDER,
-	SKILL_ARCTICBLAST,
-	SKILL_CYCLEOFLIFE,
-	SKILL_FERALRAGE,
-	SKILL_MAUL,
-	SKILL_ERUPTION,
-	SKILL_CYCLONEARMOR,
-	SKILL_HEARTOFWOLVERINE,
-	SKILL_SUMMONFENRIS,
-	SKILL_RABIES,
-	SKILL_FIRECLAWS,
-	SKILL_TWISTER,
-	SKILL_VINES,
-	SKILL_HUNGER,
-	SKILL_SHOCKWAVE,
-	SKILL_VOLCANO,
-	SKILL_TORNADO,
-	SKILL_SPIRITOFBARBS,
-	SKILL_SUMMONGRIZZLY,
-	SKILL_FURY,
-	SKILL_ARMAGEDDON,
-	SKILL_HURRICANE,
-	SKILL_FIRETRAUMA,
-	SKILL_CLAWMASTERY,
-	SKILL_PSYCHICHAMMER,
-	SKILL_TIGERSTRIKE,
-	SKILL_DRAGONTALON,
-	SKILL_SHOCKFIELD,
-	SKILL_BLADESENTINEL,
-	SKILL_QUICKNESS,
-	SKILL_FISTSOFFIRE,
-	SKILL_DRAGONCLAW,
-	SKILL_CHARGEDBOLTSENTRY,
-	SKILL_WAKEOFFIRESENTRY,
-	SKILL_WEAPONBLOCK,
-	SKILL_CLOAKOFSHADOWS,
-	SKILL_COBRASTRIKE,
-	SKILL_BLADEFURY,
-	SKILL_FADE,
-	SKILL_SHADOWWARRIOR,
-	SKILL_CLAWSOFTHUNDER,
-	SKILL_DRAGONTAIL,
-	SKILL_LIGHTNINGSENTRY,
-	SKILL_INFERNOSENTRY,
-	SKILL_MINDBLAST,
-	SKILL_BLADESOFICE,
-	SKILL_DRAGONFLIGHT,
-	SKILL_DEATHSENTRY,
-	SKILL_BLADESHIELD,
-	SKILL_VENOM,
-	SKILL_SHADOWMASTER,
-	SKILL_ROYALSTRIKE,
-	SKILL_WAKEOFDESTRUCTIONSENTRY,
-	SKILL_IMPINFERNO,
-	SKILL_IMPFIREBALL,
-	SKILL_BAALTAUNT,
-	SKILL_BAALCORPSEEXPLODE,
-	SKILL_BAALMONSTERSPAWN,
-	SKILL_CATAPULTCHARGEDBALL,
-	SKILL_CATAPULTSPIKEBALL,
-	SKILL_SUCKBLOOD,
-	SKILL_CRYHELP,
-	SKILL_HEALINGVORTEX,
-	SKILL_TELEPORT2,
-	SKILL_SELF_RESURRECT,
-	SKILL_VINEATTACK,
-	SKILL_OVERSEERWHIP,
-	SKILL_BARBSAURA,
-	SKILL_WOLVERINEAURA,
-	SKILL_OAKSAGEAURA,
-	SKILL_IMPFIREMISSILE,
-	SKILL_IMPREGNATE,
-	SKILL_SIEGEBEASTSTOMP,
-	SKILL_MINIONSPAWNER,
-	SKILL_CATAPULTBLIZZARD,
-	SKILL_CATAPULTPLAGUE,
-	SKILL_CATAPULTMETEOR,
-	SKILL_BOLTSENTRY,
-	SKILL_CORPSECYCLER,
-	SKILL_DEATHMAUL,
-	SKILL_DEFENSECURSE,
-	SKILL_BLOODMANA,
-	SKILL_MONINFERNOSENTRY,
-	SKILL_MONDEATHSENTRY,
-	SKILL_SENTRYLIGHTNING,
-	SKILL_FENRISRAGE,
-	SKILL_BAALTENTACLE,
-	SKILL_BAALNOVA,
-	SKILL_BAALINFERNO,
-	SKILL_BAALCOLDMISSILES,
-	SKILL_MEGADEMONINFERNO,
-	SKILL_EVILHUTSPAWNER,
-	SKILL_COUNTESSFIREWALL,
-	SKILL_IMPBOLT,
-	SKILL_HORRORARCTICBLAST,
-	SKILL_DEATHSENTRYLTNG,
-	SKILL_VINECYCLER,
-	SKILL_BEARSMITE,
-	SKILL_RESURRECT2,
-	SKILL_BLOODLORDFRENZY,
-	SKILL_BAALTELEPORT,
-	SKILL_IMPTELEPORT,
-	SKILL_BAALCLONETELEPORT,
-	SKILL_ZAKARUMLIGHTNING,
-	SKILL_VAMPIREMISSILE,
-	SKILL_MEPHISTOMISSILE,
-	SKILL_DOOMKNIGHTMISSILE,
-	SKILL_ROGUEMISSILE,
-	SKILL_HYDRAMISSILE,
-	SKILL_NECROMAGEMISSILE,
-	SKILL_MONBOW,
-	SKILL_MONFIREARROW,
-	SKILL_MONCOLDARROW,
-	SKILL_MONEXPLODINGARROW,
-	SKILL_MONFREEZINGARROW,
-	SKILL_MONPOWERSTRIKE,
-	SKILL_SUCCUBUSBOLT,
-	SKILL_MEPHFROSTNOVA,
-	SKILL_MONICESPEAR,
-	SKILL_SHAMANICE,
-	SKILL_DIABLOGEDDON,
-	SKILL_DELERIUMCHANGE,
-	SKILL_NIHLATHAKCORPSEEXPLOSION,
-	SKILL_SERPENTCHARGE,
-	SKILL_TRAPNOVA,
-	SKILL_UNHOLYBOLTEX,
-	SKILL_SHAMANFIREEX,
-	SKILL_IMPFIREMISSILEEX,
-	SKILL_FIXEDSIEGEBEASTSTOMP,
-};
 
 enum D2C_SkillCastType
 {
@@ -4111,728 +1204,7 @@ enum D2SaveCharType
 };
 #endif // CONSTANTS_UNITS
 
-#ifndef CONSTANTS_MISSILES //Pseudo-Macro to group all constants
-enum D2C_MissilesHcIdx
-{
-	MISSILE_ARROW = 0,
-	MISSILE_JAVELIN,
-	MISSILE_BIGHEAD1,
-	MISSILE_BIGHEAD2,
-	MISSILE_BIGHEAD3,
-	MISSILE_BIGHEAD4,
-	MISSILE_BIGHEAD5,
-	MISSILE_SPIKE1,
-	MISSILE_SPIKE2,
-	MISSILE_SPIKE3,
-	MISSILE_SPIKE4,
-	MISSILE_SPIKE5,
-	MISSILE_FIREARROW,
-	MISSILE_CR_ARROW1,
-	MISSILE_CR_ARROW2,
-	MISSILE_CR_ARROW3,
-	MISSILE_CR_ARROW4,
-	MISSILE_CR_ARROW5,
-	MISSILE_BLOOD1,
-	MISSILE_BLOOD2,
-	MISSILE_BIGBLOOD1,
-	MISSILE_BIGBLOOD2,
-	MISSILE_SHAFIRE1,
-	MISSILE_SHAFIRE2,
-	MISSILE_SHAFIRE3,
-	MISSILE_SHAFIRE4,
-	MISSILE_SHAFIRE5,
-	MISSILE_MAGICARROW,
-	MISSILE_ICEARROW,
-	MISSILE_FIREEXPLODE,
-	MISSILE_ICEEXPLODE,
-	MISSILE_BOLT,
-	MISSILE_ANDARIELSPRAY,
-	MISSILE_BIGHEADEXP,
-	MISSILE_SHAMANEXP,
-	MISSILE_THROWAXE,
-	MISSILE_THROWKNIFE,
-	MISSILE_GLAIVE,
-	MISSILE_POISONJAV,
-	MISSILE_POISONJAVCLOUD,
-	MISSILE_COLDARROW,
-	MISSILE_EXPLODINGARROW,
-	MISSILE_EXPLODINGARROWEXP,
-	MISSILE_PLAGUEJAVELIN,
-	MISSILE_OILPOTION,
-	MISSILE_EXPLOSIVEPOTION,
-	MISSILE_FULMINATINGPOTION,
-	MISSILE_RANCIDGASEPOTION,
-	MISSILE_CHOKINGGASPOITION,
-	MISSILE_STRANGLINGGASPOTION,
-	MISSILE_NOTUSED50,
-	MISSILE_EXPLOSIVEPOTIONEXP,
-	MISSILE_EXPLOSIVEPOTIONDEBRIS1,
-	MISSILE_EXPLOSIVEPOTIONDEBRIS2,
-	MISSILE_EXPLOSIVEPOTIONDEBRIS3,
-	MISSILE_HOLYBOLT,
-	MISSILE_CHARGEDBOLT,
-	MISSILE_SANCTUARYBOLT,
-	MISSILE_FIREBOLT,
-	MISSILE_ICEBOLT,
-	MISSILE_INFERNOFLAME1,
-	MISSILE_INFERNOFLAME2,
-	MISSILE_FIREBALL,
-	MISSILE_MUMMY1,
-	MISSILE_MUMMY2,
-	MISSILE_MUMMY3,
-	MISSILE_MUMMY4,
-	MISSILE_BLAZE,
-	MISSILE_FIREWALLMAKER,
-	MISSILE_FIREWALL,
-	MISSILE_GOOSPIT1,
-	MISSILE_GOOSPIT2,
-	MISSILE_GOOSPIT3,
-	MISSILE_GOOSPIT4,
-	MISSILE_GOOSPIT5,
-	MISSILE_GOOSPLAT,
-	MISSILE_SAND_PILE,
-	MISSILE_UNHOLYBOLT1,
-	MISSILE_UNHOLYBOLT2,
-	MISSILE_UNHOLYBOLT3,
-	MISSILE_UNHOLYBOLT4,
-	MISSILE_SANCTUARYCENTER,
-	MISSILE_FIREEXPLOSION,
-	MISSILE_STUCKARROW,
-	MISSILE_FOOTPRINT,
-	MISSILE_IMMOLATIONARROW,
-	MISSILE_GUIDEDARROW,
-	MISSILE_FREEZINGARROW,
-	MISSILE_FREEZINGARROWEXP1,
-	MISSILE_FREEZINGARROWEXP2,
-	MISSILE_NOVA,
-	MISSILE_ICEBLAST,
-	MISSILE_BLESSEDHAMMER,
-	MISSILE_CHAINLIGHTNING,
-	MISSILE_FISTOFARES,
-	MISSILE_CHILLBLOOD,
-	MISSILE_GLACIALSPIKE,
-	MISSILE_TELEPORT,
-	MISSILE_LIGHTNINGBOLT,
-	MISSILE_LIGHTNINGHIT,
-	MISSILE_METEOR,
-	MISSILE_METEORCENTER,
-	MISSILE_METEORTAIL,
-	MISSILE_METEOREXPLODE,
-	MISSILE_FIRESMALL,
-	MISSILE_FIREMEDIUM,
-	MISSILE_MONBLIZCENTER,
-	MISSILE_MONBLIZ1,
-	MISSILE_MONBLIZ2,
-	MISSILE_MONBLIZ3,
-	MISSILE_MONBLIZ4,
-	MISSILE_MONBLIZEXPLODE1,
-	MISSILE_MONBLIZEXPLODE2,
-	MISSILE_MONBLIZEXPLODE3,
-	MISSILE_TEETH,
-	MISSILE_CORPSEEXPLOSION,
-	MISSILE_POISONCORPSEEXPLOSION,
-	MISSILE_MONSTERCORPSEEXPLODE,
-	MISSILE_POISONNOVA,
-	MISSILE_FROSTNOVA,
-	MISSILE_ROGUE1,
-	MISSILE_ROGUE2,
-	MISSILE_ROGUE3,
-	MISSILE_BAT_LIGHTNING_BOLT,
-	MISSILE_BAT_LIGHTNING_TRAIL,
-	MISSILE_SKMAGE1,
-	MISSILE_SKMAGE2,
-	MISSILE_SKMAGE3,
-	MISSILE_SKMAGE4,
-	MISSILE_VAMPIREFIREBALL,
-	MISSILE_VAMPIREFIREWALLMAKER,
-	MISSILE_VAMPIREFIREWALL,
-	MISSILE_VAMPIREMETEOR,
-	MISSILE_VAMPIREMETEORCENTER,
-	MISSILE_VAMPIREMETEOREXP,
-	MISSILE_RAVEN1,
-	MISSILE_RAVEN2,
-	MISSILE_AMPHIBIANGOO1,
-	MISSILE_AMPHIBIANGOO2,
-	MISSILE_TENTACLEGOO,
-	MISSILE_AMPHIBIANEXPLODE,
-	MISSILE_POISONPUFF,
-	MISSILE_CURSEEFFECTRED,
-	MISSILE_SPIDERGOOLAY,
-	MISSILE_FETISHINFERNO1,
-	MISSILE_FETISHINFERNO2,
-	MISSILE_SPIDERGOO,
-	MISSILE_CURSECAST,
-	MISSILE_HOWL,
-	MISSILE_SHOUT,
-	MISSILE_DUST,
-	MISSILE_REDLIGHTMISSILE,
-	MISSILE_GREENLIGHTMISSILE,
-	MISSILE_BLUELIGHTMISSILE,
-	MISSILE_WHITELIGHTMISSILE,
-	MISSILE_CORPSEPOISONCLOUD,
-	MISSILE_CHILLBLOODCLOUD,
-	MISSILE_CHILLBLOODPUFF,
-	MISSILE_BLIZZARDCENTER,
-	MISSILE_BLIZZARD1,
-	MISSILE_BLIZZARD2,
-	MISSILE_BLIZZARD3,
-	MISSILE_BLIZZARD4,
-	MISSILE_BLIZZARDEXPLODE1,
-	MISSILE_BLIZZARDEXPLODE2,
-	MISSILE_BLIZZARDEXPLODE3,
-	MISSILE_THUNDERSTORM1,
-	MISSILE_THUNDERSTORM2,
-	MISSILE_THUNDERSTORM3,
-	MISSILE_THUNDERSTORM4,
-	MISSILE_MONSTERLIGHT,
-	MISSILE_POISONBALL,
-	MISSILE_DIABLIGHT,
-	MISSILE_REDEMPTION,
-	MISSILE_REDEMPTIONFAIL,
-	MISSILE_HANDOFGOD,
-	MISSILE_DIABFIRE,
-	MISSILE_FINGERMAGESPIDER,
-	MISSILE_ELECTRIC_THROWAXE,
-	MISSILE_DIABWALLMAKER,
-	MISSILE_DIABWALL,
-	MISSILE_CURSEAMPLIFYDAMAGE,
-	MISSILE_CURSEDIMVISION,
-	MISSILE_CURSEWEAKEN,
-	MISSILE_CURSEIRONMAIDEN,
-	MISSILE_CURSETERROR,
-	MISSILE_CURSEATTRACT,
-	MISSILE_CURSEREVERSEVAMPIRE,
-	MISSILE_CURSECONFUSE,
-	MISSILE_CURSEDECREPIFY,
-	MISSILE_CURSELOWERRESIST,
-	MISSILE_CURSECENTER,
-	MISSILE_BONESPEAR,
-	MISSILE_BONESPIRIT,
-	MISSILE_COLDUNIQUE,
-	MISSILE_LIGHTUNIQUE,
-	MISSILE_SKBOWARROW1,
-	MISSILE_SKBOWARROW2,
-	MISSILE_SKBOWARROW3,
-	MISSILE_SKBOWARROW4,
-	MISSILE_SKBOWARROW5,
-	MISSILE_NOVA1,
-	MISSILE_NOVA2,
-	MISSILE_ANDYPOISONBOLT,
-	MISSILE_TEETHEXPLODE,
-	MISSILE_LIGHTNINGJAVELIN,
-	MISSILE_LIGHTNINGFURY,
-	MISSILE_BONEWALLMAKER,
-	MISSILE_NECROMAGE1,
-	MISSILE_NECROMAGE2,
-	MISSILE_NECROMAGE3,
-	MISSILE_NECROMAGE4,
-	MISSILE_SPARKLE,
-	MISSILE_MULTIPLESHOTARROW,
-	MISSILE_MULTIPLESHOTBOLT,
-	MISSILE_CHARGEDSTRIKEBOLT,
-	MISSILE_BONESPEAREXPLODE,
-	MISSILE_POISONEXPLOSIONCLOUD,
-	MISSILE_BONECAST,
-	MISSILE_BATTLECRY,
-	MISSILE_PRIMEPOISONCLOUD,
-	MISSILE_PLAGUEJAVCLOUD,
-	MISSILE_RANCIDGASCLOUD,
-	MISSILE_CHOKINGGASCLOUD,
-	MISSILE_STRANGLINGGASCLOUD,
-	MISSILE_BUGLIGHTNING,
-	MISSILE_PANTHERJAV1,
-	MISSILE_PANTHERJAV2,
-	MISSILE_PANTHERJAV3,
-	MISSILE_PANTHERJAV4,
-	MISSILE_IMMOLATIONFIRE,
-	MISSILE_FURYLIGHTNING,
-	MISSILE_LIGHTNINGSTRIKE,
-	MISSILE_FISTOFTHEHEAVENSDELAY,
-	MISSILE_FISTOFTHEHEAVENSBOLT,
-	MISSILE_WARCRY,
-	MISSILE_BATTLECOMMAND,
-	MISSILE_BATTLEORDERS,
-	MISSILE_PANTHERPOTORANGE,
-	MISSILE_PANTHERPOTGREEN,
-	MISSILE_METEORFIRE,
-	MISSILE_TRAP_SPIKE_RIGHT,
-	MISSILE_TRAP_SPIKE_LEFT,
-	MISSILE_TRAP_CURSED_SKULL_RIGHT,
-	MISSILE_TRAP_CURSED_SKULL_LEFT,
-	MISSILE_TRAP_POISON_BALL_RIGHT,
-	MISSILE_TRAP_POISON_BALL_LEFT,
-	MISSILE_HYDRA,
-	MISSILE_BONESPEARTRAIL,
-	MISSILE_GRIMWARDSMALLSTART,
-	MISSILE_GRIMWARDSMALL,
-	MISSILE_GRIMWARDSMALLSTOP,
-	MISSILE_GRIMWARDMEDIUMSTART,
-	MISSILE_GRIMWARDMEDIUM,
-	MISSILE_GRIMWARDMEDIUMSTOP,
-	MISSILE_GRIMWARDLARGESTART,
-	MISSILE_GRIMWARDLARGE,
-	MISSILE_GRIMWARDLARGESTOP,
-	MISSILE_ZAKARUMLIGHT,
-	MISSILE_GRIMWARDSCARE,
-	MISSILE_FROZENORB,
-	MISSILE_FROZENORBBOLT,
-	MISSILE_FROZENORBNOVA,
-	MISSILE_FROZENORBEXPLODE,
-	MISSILE_CHILLINGARMORBOLT,
-	MISSILE_FIREEXPLOSION2,
-	MISSILE_BLOWGUN,
-	MISSILE_CHAINLIGHTNING2,
-	MISSILE_REVIVESMALL,
-	MISSILE_REVIVEMEDIUM,
-	MISSILE_REVIVELARGE,
-	MISSILE_MONGLACIALSPIKE,
-	MISSILE_ICEBREAKSMALL,
-	MISSILE_ICEBREAKMEDIUM,
-	MISSILE_ICEBREAKLARGE,
-	MISSILE_ICEBREAKSMOKE,
-	MISSILE_MEPHISTO,
-	MISSILE_FIREHEAD,
-	MISSILE_WHILRWIND,
-	MISSILE_ARCANELIGHTNINGBOLT,
-	MISSILE_FROGFIRE,
-	MISSILE_FROGCOLD,
-	MISSILE_FROGPOIS,
-	MISSILE_DESERTFIREBALL,
-	MISSILE_BRDEATHCONTROL,
-	MISSILE_BRDEATHLIGHTNINGBOLT,
-	MISSILE_BRDEATHLIGHTNINGHIT,
-	MISSILE_DENOFEVILLIGHT,
-	MISSILE_CAIRNSTONES,
-	MISSILE_CAIRNSTONESSKY,
-	MISSILE_CAIRNSTONESGROUND,
-	MISSILE_TOWERMIST,
-	MISSILE_TOWERMISTTRAIL,
-	MISSILE_BRDEATHSMOKES1,
-	MISSILE_BRDEATHSMOKENU,
-	MISSILE_BRDEATHSMOKEDT,
-	MISSILE_BRDEATHSPIRITS1,
-	MISSILE_BRDEATHSPIRITNU,
-	MISSILE_BRDEATHSPIRITDT,
-	MISSILE_MEPHISTODEATHCONTROL,
-	MISSILE_MEPHISTOFIREWALLMAKER,
-	MISSILE_MEPHISTOFIREWALL,
-	MISSILE_MEPHISTOFLYINGROCKSBIG,
-	MISSILE_MEPHISTOEXPLOSIONBIG,
-	MISSILE_MEPHISTOFLYINGROCKSSMALL,
-	MISSILE_MEPHISTOEXPLOSIONSMALL,
-	MISSILE_MEPHISTODONOTDRAW,
-	MISSILE_ANDYCONTROL0,
-	MISSILE_ANDYFIREWALLMAKER,
-	MISSILE_ANDYFIREWALL,
-	MISSILE_ANDYCOLUMNFIREBASE,
-	MISSILE_ANDYCOLUMNFIRE,
-	MISSILE_ANDYFALLINGDEBRIS1,
-	MISSILE_ANDYFALLINGDEBRIS2,
-	MISSILE_ANDYFALLINGDEBRIS3,
-	MISSILE_ANDYDEBRISEXPLOSION1,
-	MISSILE_ANDYDEBRISEXPLOSION2,
-	MISSILE_ANDYDEBRISEXPLOSION3,
-	MISSILE_ANDYDEBRISEXPLOSION4,
-	MISSILE_ANDYDEBRISEXPLOSION5,
-	MISSILE_WILLOWISPLIGHTNINGBOLT,
-	MISSILE_QUEENPOISONCLOUD,
-	MISSILE_DIRT_PILE,
-	MISSILE_UNDEADMISSILE1,
-	MISSILE_UNDEADMISSILE2,
-	MISSILE_UNDEADMISSILE3,
-	MISSILE_UNDEADMISSILE4,
-	MISSILE_BONESPIRITEXPLODE,
-	MISSILE_DOPPLEZONEXPLODE,
-	MISSILE_MONBONESPIRIT,
-	MISSILE_TOWERMISTFADE,
-	MISSILE_COUNTESSFIREWALL,
-	MISSILE_TOWERCHESTSPAWNER,
-	MISSILE_HELLMETEORLAUNCH1,
-	MISSILE_HELLMETEORLAUNCH2,
-	MISSILE_HELLMETEORUP,
-	MISSILE_HELLMETEORDOWN,
-	MISSILE_HELLMETEORBALL,
-	MISSILE_HORADRICSTAFF,
-	MISSILE_HORADRICLIGHTNING,
-	MISSILE_HORADRICLIGHT,
-	MISSILE_REGURGITATORCORPSE,
-	MISSILE_REGURGITATORCORPSEEXPL,
-	MISSILE_HIGHPRIESTLIGHTNING,
-	MISSILE_ICEBREAKSMALLMELT,
-	MISSILE_ICEBREAKLARGEMELT,
-	MISSILE_LEAPKNOCKBACK,
-	MISSILE_RADAMENTDEATH,
-	MISSILE_RADAMENTHANDOFGOD,
-	MISSILE_RADAMENTHOLYBOLT,
-	MISSILE_TAINTEDSUNCONTROL,
-	MISSILE_TAINTEDSUNFLASH,
-	MISSILE_TAINTEDSUNBALL,
-	MISSILE_QUEENDEATHCENTER,
-	MISSILE_QUEENDEATHGLOB,
-	MISSILE_QUEENDEATHSPLAT1,
-	MISSILE_QUEENDEATHSPLAT2,
-	MISSILE_HEALINGBOLT,
-	MISSILE_MEPHISTOHOLEDELAY,
-	MISSILE_MEPHISTOHOLEBIRTH,
-	MISSILE_MEPHISTOHOLENEUTRAL,
-	MISSILE_MEPHISTOHOLEDEATH,
-	MISSILE_MEPHISTOHOLEDEAD,
-	MISSILE_DURIELDEATHCONTROL,
-	MISSILE_DURIELDEATHROCK,
-	MISSILE_DURIELDEATHDEBRIS,
-	MISSILE_DURIELDEATHSMOKE,
-	MISSILE_MEPHISTOEXPLOSION,
-	MISSILE_ORBMIST,
-	MISSILE_ORBMISTTRAIL,
-	MISSILE_ORBMISTFADE,
-	MISSILE_PILUM,
-	MISSILE_DIABLO_APPEARS,
-	MISSILE_HFCONTROL,
-	MISSILE_HFFRAGMENT1,
-	MISSILE_HFFRAGMENT2,
-	MISSILE_HFFRAGMENT3,
-	MISSILE_HFSPIRIT1,
-	MISSILE_HFRESERVED3,
-	MISSILE_IZUAL_CONTROL,
-	MISSILE_IZUAL_MIST_LOOP,
-	MISSILE_IZUAL_MIST_FADE,
-	MISSILE_IZUAL_LIGHTNING,
-	MISSILE_IZUAL_LIGHTNING_TRAIL,
-	MISSILE_CAIRN_STONES_BOLT,
-	MISSILE_BOMB_IN_AIR,
-	MISSILE_BOMB_ON_GROUND,
-	MISSILE_BOMB_EXPLOSION,
-	MISSILE_SHOCK_FIELD_IN_AIR,
-	MISSILE_SHOCK_FIELD_ON_GROUND,
-	MISSILE_THROWINGSTAR,
-	MISSILE_ACIDSPRAY,
-	MISSILE_BLADE_CREEPER,
-	MISSILE_DISTRACTION,
-	MISSILE_DISTRACTION_FOG,
-	MISSILE_DISTRACTION_PUFF,
-	MISSILE_DISTRACTION_START,
-	MISSILE_DISTRACTION_END,
-	MISSILE_IMPINFERNOFLAME1,
-	MISSILE_IMPINFERNOFLAME2,
-	MISSILE_BAALLIGHTNINGBOLT,
-	MISSILE_BAALLIGHTNINGTRAIL,
-	MISSILE_BAALLIGHTNINGBOLT2,
-	MISSILE_BAALLIGHTNINGTRAIL2,
-	MISSILE_IMPFIREBALL,
-	MISSILE_IMPFIREBALLEXPLODE,
-	MISSILE_CATAPULTCHARGEDBALL_ON,
-	MISSILE_CATAPULTCHARGEDBALL,
-	MISSILE_CATAPULTCHARGEDBALLBOLT,
-	MISSILE_IMP_SPAWN_MONSTERS,
-	MISSILE_CATAPULT_SPIKE_BALL_ON,
-	MISSILE_CATAPULT_SPIKE_BALL,
-	MISSILE_CATAPULT_SPIKE_IN_AIR,
-	MISSILE_CATAPULT_SPIKE_ON_GROUND,
-	MISSILE_CATAPULT_SPIKE_EXPLOSION,
-	MISSILE_CATAPULT_COLD_BALL_ON,
-	MISSILE_CATAPULT_COLD_BALL,
-	MISSILE_CATAPULT_COLD_EXPLOSION,
-	MISSILE_CATAPULT_PLAGUE_BALL_ON,
-	MISSILE_CATAPULT_PLAGUE_BALL,
-	MISSILE_CATAPULT_PLAGUE_CLOUD,
-	MISSILE_CATAPULT_METEOR_BALL_ON,
-	MISSILE_CATAPULT_METEOR_BALL,
-	MISSILE_CATAPULT_METEOR_FIRE,
-	MISSILE_TOWER_DEATH,
-	MISSILE_HEALING_VORTEX,
-	MISSILE_SUICIDECORPSEEXPLODE,
-	MISSILE_SUICIDEFIREEXPLODE,
-	MISSILE_SUICIDEICEEXPLODE,
-	MISSILE_EXPLODINGJAVALIN,
-	MISSILE_EXPLODINGJAVALINEXP,
-	MISSILE_LIGHTINGTRAILINGJAVALIN,
-	MISSILE_LIGHTJAVALINTRAIL,
-	MISSILE_LIGHTJAVALINEXPLOSION,
-	MISSILE_ICEJAVALIN,
-	MISSILE_ICEJAVALINEXPLODE,
-	MISSILE_PLAGUEJAVELIN2,
-	MISSILE_PLAGUEJAVLINEXPLODE,
-	MISSILE_ADVLIGHTTRAILINGJAV,
-	MISSILE_ADVLIGHTTRAILINGJAV2,
-	MISSILE_ADVLIGHTJAVEXPLODE,
-	MISSILE_SUCFIREBALL,
-	MISSILE_SUCFIREBALLEXPLODE,
-	MISSILE_SUCFIREBALLTRAIL,
-	MISSILE_SUCSHOCKFIELDMISSILE,
-	MISSILE_SUCSHOCKFIELDMISSILEEXP,
-	MISSILE_SUCSHOCKFIELD,
-	MISSILE_HELLFIREMISSILE,
-	MISSILE_HELLFIREEXA,
-	MISSILE_HELLFIREEXB,
-	MISSILE_IMP_CHARGED_BOLT,
-	MISSILE_IMP_TELEPORT,
-	MISSILE_MOLTENBOULDER,
-	MISSILE_MOLTENBOULDEREMERGE,
-	MISSILE_MOLTENBOULDEREXPLODE,
-	MISSILE_MOLTENBOULDERFIREPATH,
-	MISSILE_MOLTENBOULDER_FLYINGROCKS,
-	MISSILE_FIRESTORM,
-	MISSILE_FIRESTORMMAKER,
-	MISSILE_ARCTICBLAST1,
-	MISSILE_ARCTICBLAST2,
-	MISSILE_ERRUPTION_CENTER,
-	MISSILE_ERRUPTION_CRACK_1,
-	MISSILE_ERRUPTION_CRACK_2,
-	MISSILE_ERRUPTION_SMOKE_1,
-	MISSILE_ERRUPTION_SMOKE_2,
-	MISSILE_VINE_BEAST_WALK_1,
-	MISSILE_VINE_BEAST_WALK_2,
-	MISSILE_VINE_BEAST_NEUTRAL,
-	MISSILE_VINE_BEAST_ATTACK,
-	MISSILE_VINE_BEAST_DEATH,
-	MISSILE_VINES,
-	MISSILE_VINES_TRAIL,
-	MISSILE_VINES_WITHER,
-	MISSILE_PLAGUE_VINES,
-	MISSILE_PLAGUE_VINES_TRAIL,
-	MISSILE_PLAGUE_VINES_WITHER,
-	MISSILE_TWISTER,
-	MISSILE_TORNADO,
-	MISSILE_VOLCANO,
-	MISSILE_VOLCANO_OVERLAY_FIRE,
-	MISSILE_VOLCANO_DEBRIS_2,
-	MISSILE_VOLCANO_EXPLOSION,
-	MISSILE_VOLCANO_SMALL_FIRE,
-	MISSILE_DRAGONBREATH_MISSILE,
-	MISSILE_LUREPROJECTILE,
-	MISSILE_LURECENTER,
-	MISSILE_LURECLOUD,
-	MISSILE_IMPMISS1,
-	MISSILE_IMPMISS2,
-	MISSILE_IMPMISS3,
-	MISSILE_IMPMISS4,
-	MISSILE_IMPMISS5,
-	MISSILE_FROZENHORROR_ARCTICBLAST1,
-	MISSILE_FROZENHORROR_ARCTICBLAST2,
-	MISSILE_SENTRYCHARGEDBOLT,
-	MISSILE_SENTRYSPIKEINAIR,
-	MISSILE_SENTRYSPIKEONGROUND,
-	MISSILE_RECYCLER_DELAY,
-	MISSILE_RECYCLER_VINE,
-	MISSILE_RECYCLER_FADE,
-	MISSILE_RECYCLER_EXPLOSION,
-	MISSILE_DEATH_MAULER,
-	MISSILE_DEATH_MAULER_TRAIL,
-	MISSILE_DEATH_MAULER_TRAIL_FADE,
-	MISSILE_BLADEFURY1,
-	MISSILE_BLADEFRAGMENT1,
-	MISSILE_BLADEFURY2,
-	MISSILE_BLADEFRAGMENT2,
-	MISSILE_BLADEFURY3,
-	MISSILE_BLADEFRAGMENT3,
-	MISSILE_SHOCKWAVE,
-	MISSILE_LIGHTNINGTALONS,
-	MISSILE_LIGHTNINGTALONSTRAIL,
-	MISSILE_PHOENIXTRAIL,
-	MISSILE_RABIESPLAGUE,
-	MISSILE_RABIESCONTAGION,
-	MISSILE_WAKE_OF_DESTRUCTION_MAKER,
-	MISSILE_WAKE_OF_DESTRUCTION,
-	MISSILE_DEATHSENTRYEXPLODE,
-	MISSILE_TIGERFURY,
-	MISSILE_TIGERFURYTRAIL,
-	MISSILE_TIGERFURYTRAIL2,
-	MISSILE_INFERNO_SENTRY_1,
-	MISSILE_INFERNO_SENTRY_2,
-	MISSILE_ANCIENT_THROWING_AXE,
-	MISSILE_SENTRYLIGHTNINGBOLT,
-	MISSILE_SENTRYLIGHTNINGHIT,
-	MISSILE_ANYA_CENTER,
-	MISSILE_ANYA_ICICLE,
-	MISSILE_ANYA_ICEIMPACT,
-	MISSILE_ANYA_ICESTEAM,
-	MISSILE_ANYA_ICEMAGIC,
-	MISSILE_DRAGONTAIL_MISSILE,
-	MISSILE_DRAGONFLIGHT,
-	MISSILE_DRAGONFLIGHTMAKER,
-	MISSILE_PROGRESSIVE_RADIUS_DAMAGE,
-	MISSILE_VINE_BEAST_WALK_1_FADE,
-	MISSILE_VINE_BEAST_WALK_2_FADE,
-	MISSILE_VINE_BEAST_NEUTRAL_FADE,
-	MISSILE_VINE_RECYCLER_DELAY,
-	MISSILE_ANCIENT_DEATH_CENTER,
-	MISSILE_ANCIENT_DEATH_CLOUD,
-	MISSILE_LIGHTNING_CHARGE_UP_NOVA,
-	MISSILE_CHAINLIGHTNINGCHARGE_UP,
-	MISSILE_PAIN_WORM_APPEAR,
-	MISSILE_BAAL_TAUNT_CONTROL,
-	MISSILE_BAAL_TAUNT_LIGHTNING,
-	MISSILE_BAAL_TAUNT_LIGHTNING_TRAIL,
-	MISSILE_BAAL_TAUNT_POISON,
-	MISSILE_BAAL_SPAWN_MONSTERS,
-	MISSILE_MINDBLAST_HIT,
-	MISSILE_BLADE_SHIELD_MISSILE,
-	MISSILE_BLADE_SHIELD_ATTACHMENT,
-	MISSILE_BAAL_INFERNO,
-	MISSILE_BAAL_NOVA,
-	MISSILE_FISTSOFFIREEXPLODE,
-	MISSILE_FISTSOFFIREFIREWALL,
-	MISSILE_CLAWSOFTHUNDERBOLT,
-	MISSILE_CLAWSOFTHUNDERNOVA,
-	MISSILE_BLADESOFICEEXPLODE,
-	MISSILE_BLADESOFICECUBES,
-	MISSILE_BLADESOFICECUBESMELT,
-	MISSILE_ROYALSTRIKEMETEOR,
-	MISSILE_ROYALSTRIKEMETEORCENTER,
-	MISSILE_ROYALSTRIKEMETEORTAIL,
-	MISSILE_ROYALSTRIKEMETEOREXPLODE,
-	MISSILE_ROYALSTRIKEMETEORFIRE,
-	MISSILE_ROYALSTRIKECHAINLIGHTNING,
-	MISSILE_ROYALSTRIKECHAOSICE,
-	MISSILE_WORLD_STONE_CHIP_1,
-	MISSILE_WORLD_STONE_CHIP_2,
-	MISSILE_WORLD_STONE_CHIP_3,
-	MISSILE_WORLD_STONE_CHIP_4,
-	MISSILE_HIGHPRIESTLIGHTNING2,
-	MISSILE_INFERNOFLAME3,
-	MISSILE_MINDBLAST_CENTER,
-	MISSILE_ARMAGEDDONCONTROL,
-	MISSILE_ARMAGEDDONROCK,
-	MISSILE_ARMAGEDDONTAIL,
-	MISSILE_ARMAGEDDONEXPLOSION,
-	MISSILE_HURRICANESWOOSH,
-	MISSILE_HURRICANECART,
-	MISSILE_HURRICANEROCK,
-	MISSILE_HURRICANESACK,
-	MISSILE_HURRICANETREE,
-	MISSILE_HURRICANEVASE,
-	MISSILE_BAALCORPSEEXPLODEDELAY,
-	MISSILE_BAALCORPSEEXPLODEEXPL,
-	MISSILE_BAAL_COLD_MAKER,
-	MISSILE_BAAL_COLD_TRAIL,
-	MISSILE_BAAL_SPAWN_MONSTERS_EXP,
-	MISSILE_IMPMISS21,
-	MISSILE_IMPMISS22,
-	MISSILE_IMPMISS23,
-	MISSILE_IMPMISS24,
-	MISSILE_IMPMISS25,
-	MISSILE_ANYASTEAM1,
-	MISSILE_ANYASTEAM2,
-	MISSILE_ANCIENTSGUIDE,
-	MISSILE_ANCIENTSMARKER,
-	MISSILE_ANCIENTSCONTROL,
-	MISSILE_OVERSEERCONTROL,
-	MISSILE_NIHLITHAK1,
-	MISSILE_NIHLITHAK2,
-	MISSILE_NIHLITHAK3,
-	MISSILE_NEHLITHAKCONTROL,
-	MISSILE_NEHLITHAKSWOOSH,
-	MISSILE_NEHLITHAKDEBRIS1,
-	MISSILE_NEHLITHAKDEBRIS2,
-	MISSILE_NEHLITHAKDEBRIS3,
-	MISSILE_NEHLITHAKDEBRIS4,
-	MISSILE_NEHLITHAKGLOW,
-	MISSILE_BAALTELEPORT,
-	MISSILE_BAALCLONEDEATH,
-	MISSILE_ANYASTEAMVENT,
-	MISSILE_ANYASTEAM,
-	MISSILE_NEHLITHAKHOLE,
-	MISSILE_NEHLITHAKHOLELIGHT,
-	MISSILE_VOLCANOFIRETRAIL,
-	MISSILE_NEHLITHAKGLOW2,
-	MISSILE_NEHLITHAKBONECHIPS,
-	MISSILE_BAALCORPSEEXPLODEFADE,
-	MISSILE_ARMAGEDDONFIRE,
-	MISSILE_ICESPARKLE,
-	MISSILE_BAALFX_CONTROL,
-	MISSILE_BAALFX_SPIRIT_1,
-	MISSILE_BAALFX_SPIRIT_2,
-	MISSILE_BAALFX_SPIRIT_3,
-	MISSILE_BAALFX_SPIRIT_4,
-	MISSILE_BAALFX_SPIRIT_5,
-	MISSILE_BAALFX_BAAL_HEAD_APPEAR,
-	MISSILE_BAALFX_BAAL_HEAD_1,
-	MISSILE_BAALFX_BAAL_HEAD_2,
-	MISSILE_BAALFX_BAAL_HEAD_3,
-	MISSILE_BAALFX_TYREAL_DEBRIS_1,
-	MISSILE_BAALFX_TYREAL_DEBRIS_2,
-	MISSILE_BAALFX_TYREAL_DEBRIS_3,
-	MISSILE_BAALFX_TYREAL_DEBRIS_BREAK,
-	MISSILE_WORLDSTONE_SHAKE,
-	MISSILE_BLESSEDHAMMEREX,
-	MISSILE_SENTRYLIGHTNINGBOLT2,
-	MISSILE_SENTRYLIGHTNINGHIT2,
-	MISSILE_LIGHTNINGTOWERNOVA,
-	MISSILE_SKBOWARROW6,
-	MISSILE_SKBOWARROW7,
-	MISSILE_SKBOWARROW8,
-	MISSILE_BIGHEAD6,
-	MISSILE_VIPER_POISJAV,
-	MISSILE_VIPER_POISJAVCLOUD,
-	MISSILE_VIPER_FIRE,
-	MISSILE_VIPER_FIRECLOUD,
-	MISSILE_VIPER_BONESPEAR,
-	MISSILE_COUNTESSFIREWALLMAKER,
-	MISSILE_BAAL_TAUNT_LIGHTNING_CONTROL,
-	MISSILE_BAAL_TAUNT_POISON_CONTROL,
-	MISSILE_EXPLODINGARROWEXP2,
-	MISSILE_FREEZINGARROWEXP3,
-	MISSILE_PANTHERJAV5,
-	MISSILE_SPIKE6,
-	MISSILE_CR_ARROW6,
-	MISSILE_SKMAGEPOIS,
-	MISSILE_SKMAGECOLD,
-	MISSILE_SKMAGEFIRE,
-	MISSILE_SKMAGELTNG,
-	MISSILE_SUCCUBUSMISS,
-	MISSILE_WILLOWISPLIGHTNINGBOLT2,
-	MISSILE_MUMMYEX,
-	MISSILE_GOOSPITEX,
-	MISSILE_IMPMISSEX,
-	MISSILE_DIABLOGEDDONCONTROL,
-	MISSILE_DIABLOGEDDONROCK,
-	MISSILE_DIABLOGEDDONTAIL,
-	MISSILE_DIABLOGEDDONEXPLOSION,
-	MISSILE_DIABLOGEDDONFIRE,
-	MISSILE_MEGADEMONINFERNO,
-	MISSILE_TRAPFIREBOLT,
-	MISSILE_TRAPPOISONJAVCLOUD,
-	MISSILE_TRAPNOVA,
-	MISSILE_MEPHFROSTNOVA,
-	MISSILE_MEPHLIGHT,
-	MISSILE_VAMPIREMETEORFIRE,
-	MISSILE_STRAFEARROW,
-	MISSILE_STRAFEBOLT,
-	MISSILE_SIEGE_BEAST_STOMP_MISSILE
-};
-#endif // CONSTANTS_MISSILES
-
 #ifndef CONSTANTS_DATATABLES //Pseudo-Macro to group all constants
-enum D2C_TxtFieldTypes
-{
-	TXTFIELD_NONE, 
-	TXTFIELD_ASCII, 
-	TXTFIELD_DWORD, 
-	TXTFIELD_WORD, 
-	TXTFIELD_BYTE, 
-	TXTFIELD_UNKNOWN1, 
-	TXTFIELD_UNKNOWN2, 
-	TXTFIELD_BYTE2, 
-	TXTFIELD_DWORD2, 
-	TXTFIELD_RAW, 
-	TXTFIELD_ASCIITOCODE, 
-	TXTFIELD_UNKNOWN3, 
-	TXTFIELD_UNKNOWN4, 
-	TXTFIELD_CODETOBYTE, 
-	TXTFIELD_UNKNOWN5, 
-	TXTFIELD_CODETOWORD, 
-	TXTFIELD_UNKNOWN6, 
-	TXTFIELD_NAMETOINDEX, 
-	TXTFIELD_NAMETOINDEX2, 
-	TXTFIELD_NAMETODWORD, 
-	TXTFIELD_NAMETOWORD, 
-	TXTFIELD_NAMETOWORD2, 
-	TXTFIELD_KEYTOWORD, 
-	TXTFIELD_CUSTOMLINK, 
-	TXTFIELD_UNKNOWN7, 
-	TXTFIELD_CALCTODWORD, 
-	TXTFIELD_BIT
-};
 
 enum D2C_TxtLinkerIndexTypes
 {
