+++ source/Fog/include/Fog.h
@@ -46,7 +46,45 @@ struct D2BinFileStrc
 	int32_t nCellCount;				//0x0C
 };
 
-struct D2BinFieldStrc;
+enum D2C_TxtFieldTypes : uint32_t
+{
+	TXTFIELD_NONE,
+	TXTFIELD_ASCII,
+	TXTFIELD_DWORD,
+	TXTFIELD_WORD,
+	TXTFIELD_BYTE,
+	TXTFIELD_UNKNOWN1,
+	TXTFIELD_UNKNOWN2,
+	TXTFIELD_BYTE2,
+	TXTFIELD_DWORD2,
+	TXTFIELD_RAW,
+	TXTFIELD_ASCIITOCODE,
+	TXTFIELD_UNKNOWN3,
+	TXTFIELD_UNKNOWN4,
+	TXTFIELD_CODETOBYTE,
+	TXTFIELD_UNKNOWN5,
+	TXTFIELD_CODETOWORD,
+	TXTFIELD_UNKNOWN6,
+	TXTFIELD_NAMETOINDEX,
+	TXTFIELD_NAMETOINDEX2,
+	TXTFIELD_NAMETODWORD,
+	TXTFIELD_NAMETOWORD,
+	TXTFIELD_NAMETOWORD2,
+	TXTFIELD_KEYTOWORD,
+	TXTFIELD_CUSTOMLINK,
+	TXTFIELD_UNKNOWN7,
+	TXTFIELD_CALCTODWORD,
+	TXTFIELD_BIT
+};
+
+struct D2BinFieldStrc
+{
+	char* szFieldName;				//0x00
+	D2C_TxtFieldTypes nFieldType;	//0x04
+	int32_t nFieldLength;			//0x08
+	int32_t nFieldOffset;			//0x0C
+	void* pLinkField;				//0x10
+};
 
 struct D2UnkExcelStrc
 {
